### 2023-02-05 c42e9ab
Add documentation (first draft)
