+++ graphblas/core/ss/vector.py
@@ -425,7 +425,7 @@ class ss:
             Whether to sort indices if the format is "sparse"
         give_ownership : bool, default False
             Perform a zero-copy data transfer to Python if possible.  This gives ownership of
-            the underlying memory buffers to Numpy.
+            the underlying memory buffers to NumPy.
             ** If True, this nullifies the current object, which should no longer be used! **
         raw : bool, default False
             If True, always return array the same size as returned by SuiteSparse.
