### 2023-02-07 #81 @NickleDave
Hi @eriknw!

We're very glad to see you all have gone ahead with a full submission, after discussion in https://github.com/pyOpenSci/python-package-guide/issues/21#issuecomment-1368046000 as you linked above.

I just want to welcome you and let you know we are working on this.

I will finish the initial editor checks by the end of this week. @lwasser is traveling and we will need to co-ordinate about editors, but I expect to get back to you about that by the middle of next week at the latest.

Thank you for providing all the detail in the submission. The context you've provided will be helpful for the review (and will definitely help me with editor checks!). It sounds like you've anticipated some points @lwasser brought up when we discussed in Slack. Looking forward to helping you improve the docs and giving you some blog post material! :grin:
