### 2023-02-07 0b2ac41
Add warning that `to_dense` can create very large arrays
