### 2023-02-08 81f9d9e
Remove deprecated `Matrix.new` (#380)
