### 2023-02-08 1bbce69
`.wait` return `self` so it can be used with method-chaining (#379)
* `.chain` return `self` so it can be used with method-chaining
