### 2023-02-09 #81 @NickleDave
# Editor in Chief checks

These are the basic checks that the package needs to pass to begin review.
Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

  - [x] The package imports properly into a standard Python environment `import package-name`.
    - was able to do both `pip install python-graphbals` in a python 3.10 venv on Linux
    - and install with conda on Mac OS, python 3.10
    - but see comments below
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
    - but see comments below
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
    - Apache 2.0
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---

This package passes all checks; we can begin review.

It's obvious the developers and maintainers have done a ton of work.
Our goal here should be to help make sure *everyone* can appreciate how much work they have done, and how much functionality is packed into python-graphblas.

Along those lines, some notes for the review (none of these need to be addressed before we start):
- [ ] the description section of the README needs a high level, non-technical description so that the use of the package would be clear to anyone with a general scientific background
  - Specifically there should be language that defines *sparse data* - just a simple definition like "data with lots of 0's "and then link to a good description of sparse data from scipy or some other resource
  - Language the submitting authors use above is very close to a good high level description: "python-graphblas ... is a Python library for ... high-performance sparse linear algebra for scalable graph analytics. [Think of it like a faster, more capable scipy.sparse that can implement NetworkX.] Python-graphblas mimics the math notation, making it the most natural way to learn, use, and think about GraphBLAS. In contrast to other high level GraphBLAS bindings, Python-graphblas can fully and cleanly support any implementation of the GraphBLAS C API specification thereby allowing us to be vendor-agnostic." This could be the first thing someone reads both in the README description and on the index of the docs
- [ ] the description section of the README should also clarify why would someone use python-graphblas in place of, or in conjunction, with the scipy sparse module and/or networkx, but in very easy to understand language
- [ ] ideally the README itself should have a short, beginning to end tutorial to get someone started, that really demonstrates the power of the package
  - for example this snippet appears to be almost stand-alone, and the set-up lines will be understandable by most scientific Python users: https://python-graphblas.readthedocs.io/en/latest/getting_started/primer.html#sssp-in-python-graphblas
  - as opposed to the examples that are currently in the README that are meant to give an impression of the package but are not stand-alone runnable snippets; these examples may be useful in API documentation. But the README as currently written may be missing the opportunity to recruit users without a quick friendly walkthrough
- [ ] as noted by submitting author, there is an existing tutorial, the [python-graphblas primer](https://python-graphblas.readthedocs.io/en/latest/getting_started/primer.html#sssp-in-python-graphblas), but there could be additional vignettes;
  - what are typical use cases, perhaps analyses from papers or packages that already use python-graphblas could be added in a how-tos section of the docs? Submitting authors state there may be examples in pygraphblas documentation
