### 2023-02-09 #81 @NickleDave
> it's more important to give the right people the right amount of time to do things right

🙌

finding the right people now! 🙂
