+++ graphblas/core/matrix.py
@@ -1329,7 +1329,33 @@ class Matrix(BaseType):
         return cls.from_coo(row_indices, cols, values, dtype, nrows=nrows, ncols=ncols, name=name)
 
     @classmethod
-    def from_iso_value(cls, value, nrows, ncols, dtype=None, *, name=None):
+    def from_iso_value(cls, value, nrows, ncols, dtype=None, *, name=None, **opts):
+        """Create a fully dense Matrix filled with a scalar value.
+
+        Parameters
+        ----------
+        value : scalar
+            Scalar value used to fill the Matrix.
+        nrows : int
+            Number of rows.
+        ncols : int
+            Number of columns.
+        dtype : DataType, optional
+            Data type of the Matrix. If not provided, the scalar value will be
+            inspected to choose an appropriate dtype.
+        name : str, optional
+            Name to give the Matrix.
+
+        See Also
+        --------
+        from_coo
+        from_dense
+        from_edgelist
+
+        Returns
+        -------
+        Matrix
+        """
         value, dtype = values_to_numpy_buffer(value, dtype, subarray_after=0)
         if backend == "suitesparse" and not dtype._is_udt:
             # `Matrix.ss.import_fullr` does not yet handle all cases with UDTs
@@ -1337,18 +1363,21 @@ class Matrix(BaseType):
                 value, dtype=dtype, nrows=nrows, ncols=ncols, is_iso=True, name=name
             )
         rv = cls(dtype, nrows=nrows, ncols=ncols, name=name)
-        rv << value
+        rv(**opts) << value
         return rv
 
     @classmethod
-    def from_dense(cls, values, missing_value=None, *, dtype=None, name=None):
-        """Create a fully dense Matrix from a NumPy array or list of lists.
+    def from_dense(cls, values, missing_value=None, *, dtype=None, name=None, **opts):
+        """Create a Matrix from a NumPy array or list of lists.
 
         Parameters
         ----------
         values : list or np.ndarray
             List of values.
-        dtype :
+        missing_value : scalar, optional
+            A scalar value to consider "missing"; elements of this value will be dropped.
+            If None, then the resulting Matrix will be dense.
+        dtype : DataType, optional
             Data type of the Matrix. If not provided, the values will be inspected
             to choose an appropriate dtype.
         name : str, optional
@@ -1358,6 +1387,7 @@ class Matrix(BaseType):
         --------
         from_coo
         from_edgelist
+        from_iso_value
         to_dense
 
         Returns
@@ -1396,10 +1426,10 @@ class Matrix(BaseType):
                 name=name,
             )
         if missing_value is not None:
-            rv << select.valuene(rv, missing_value)
+            rv(**opts) << select.valuene(rv, missing_value)
         return rv
 
-    def to_dense(self, fill_value=None, dtype=None):
+    def to_dense(self, fill_value=None, dtype=None, **opts):
         """Convert Matrix to NumPy array of the same shape with missing values filled.
 
         .. warning::
@@ -1409,7 +1439,7 @@ class Matrix(BaseType):
         ----------
         fill_value : scalar, optional
             Value used to fill missing values. This is required if there are missing values.
-        dtype :
+        dtype : DataType, optional
             Requested dtype for the output values array.
 
         See Also
@@ -1450,10 +1480,10 @@ class Matrix(BaseType):
                     )
             dtype = unify(fill_value.dtype, self.dtype, is_left_scalar=True)
 
-        rv = self.dup(dtype, clear=True, name="to_dense")
-        rv << fill_value
[…]
