+++ .github/workflows/test_and_build.yml
@@ -81,7 +81,7 @@ jobs:
         shell: bash -l {0}
     strategy:
       # To "stress test" in CI, set `fail-fast` to `false` and perhaps add more items to `matrix.slowtask`
-      fail-fast: true
+      fail-fast: false
       # The build matrix is [os]x[slowtask] and then randomly chooses [pyver] and [sourcetype].
       # This should ensure we'll have full code coverage (i.e., no chance of getting unlucky),
       # since we need to run all slow tests on Windows and non-Windoes OSes.
@@ -118,10 +118,10 @@ jobs:
             source
             upstream
           weights: |
-            2
             1
             1
             1
+            1000000
       - name: Setup conda
         uses: conda-incubator/setup-miniconda@v2
         with:
@@ -204,7 +204,7 @@ jobs:
           elif [[ ${{ steps.sourcetype.outputs.selected }} == "source" ]]; then
               pip install --no-deps --no-binary=all suitesparse-graphblas
           elif [[ ${{ steps.sourcetype.outputs.selected }} == "upstream" ]]; then
-              pip install --no-deps git+https://github.com/GraphBLAS/python-suitesparse-graphblas.git@main#egg=suitesparse-graphblas
+              pip install --no-deps git+https://github.com/GraphBLAS/python-suitesparse-graphblas.git@b65050921ea71606436f37279c225edaf95fc677#egg=suitesparse-graphblas
           fi
           pip install --no-deps -e .
       - name: Unit tests
