### 2023-02-12 48b14d2
See what happens with KMP_DUPLICATE_LIB_OK=True
