### 2023-02-12 6292c98
Better comment and add a test for openmp enabled
