+++ .github/workflows/test_and_build.yml
@@ -17,9 +17,6 @@ on:
     branches:
       - main
 
-env:
-  KMP_DUPLICATE_LIB_OK: True
-
 jobs:
   rngs:
     # To achieve consistent coverage, we need a little bit of correlated collaboration.
@@ -84,13 +81,17 @@ jobs:
         shell: bash -l {0}
     strategy:
       # To "stress test" in CI, set `fail-fast` to `false` and perhaps add more items to `matrix.slowtask`
-      fail-fast: false
+      fail-fast: true
       # The build matrix is [os]x[slowtask] and then randomly chooses [pyver] and [sourcetype].
       # This should ensure we'll have full code coverage (i.e., no chance of getting unlucky),
       # since we need to run all slow tests on Windows and non-Windoes OSes.
       matrix:
         os: ["ubuntu-latest", "macos-latest", "windows-latest"]
         slowtask: ["pytest_normal", "pytest_bizarro", "notebooks"]
+    env:
+      # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
+      # Setting this is a workaround.
+      KMP_DUPLICATE_LIB_OK: True
     steps:
       - name: Checkout
         uses: actions/checkout@v3
