+++ .pre-commit-config.yaml
@@ -71,7 +71,7 @@ repos:
         additional_dependencies: [tomli]
         files: ^(graphblas|docs)/
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.244
+    rev: v0.0.245
     hooks:
       - id: ruff
   - repo: https://github.com/sphinx-contrib/sphinx-lint
