### 2023-02-15 cefc282
Rename `from_iso_value` to `from_scalar`
