+++ graphblas/unary/__init__.py
@@ -17,6 +17,7 @@ def __getattr__(key):
             f"`{key}` is specific to SuiteSparse:GraphBLAS. "
             f"`gb.unary.{key}` will be removed in version 2023.9.0 or later.",
             DeprecationWarning,
+            stacklevel=2,
         )
         rv = _deprecated[key]
         globals()[key] = rv
