+++ scripts/check_versions.sh
@@ -11,5 +11,5 @@ conda search 'awkward[channel=conda-forge]>=2.0.7'
 conda search 'numba[channel=conda-forge]>=0.56.4'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
 conda search 'flake8-comprehensions[channel=conda-forge]>=3.10.1'
-conda search 'flake8-bugbear[channel=conda-forge]>=23.1.20'
+conda search 'flake8-bugbear[channel=conda-forge]>=23.2.13'
 conda search 'flake8-simplify[channel=conda-forge]>=0.19.3'
