### 2023-02-17 b4a80a4
Add comment for how to create iso-valued objects with structure and scalar
A more flexible way with any mask is e.g.:
```python
w = Vector(v.dtype, size=v.size)
w(~v.S) << value
```
