+++ pyproject.toml
@@ -218,7 +218,7 @@ select = [
     "TRY",  # tryceratops
     # "RSE",  # flake8-raise
     # "SLF",  # flake8-self
-    # "NPY",  # NumPy-specific rules
+    "NPY",  # NumPy-specific rules
     "RUF",  # ruff-specific rules
 ]
 external = [
