### 2023-02-17 d226a51
Add `from_dense` and `to_dense` methods (#382)
* Add `from_dense` and `to_dense` methods

* Add warning that `to_dense` can create very large arrays

* Add `from_iso_value`, `missing_value=` to `from_dense`, and deprecate `io.from_numpy`

* Rename `from_iso_value` to `from_scalar`

* Update `flake8-bugbear` and make improvements

* Add comment for how to create iso-valued objects with structure and scalar
