### 2023-02-17 60e6daa
Add `op.is_idempotent` property to Monoids that means `op(x, x) == x`
