+++ graphblas/core/operator.py
@@ -416,6 +416,11 @@ class TypedBuiltinMonoid(TypedOpBase):
     def type2(self):
         return self.type
 
+    @property
+    def is_idempotent(self):
+        """True if ``monoid(x, x) == x`` for any x."""
+        return self.parent.is_idempotent
+
 
 class TypedBuiltinSemiring(TypedOpBase):
     __slots__ = ()
@@ -558,6 +563,7 @@ class TypedUserMonoid(TypedOpBase):
 
     commutes_to = TypedBuiltinMonoid.commutes_to
     type2 = TypedBuiltinMonoid.type2
+    is_idempotent = TypedBuiltinMonoid.is_idempotent
     __call__ = TypedBuiltinMonoid.__call__
 
 
@@ -756,10 +762,10 @@ class ParameterizedBinaryOp(ParameterizedUdf):
 
 
 class ParameterizedMonoid(ParameterizedUdf):
-    __slots__ = "binaryop", "identity", "__signature__"
+    __slots__ = "binaryop", "identity", "_is_idempotent", "__signature__"
     is_commutative = True
 
-    def __init__(self, name, binaryop, identity, *, anonymous=False):
+    def __init__(self, name, binaryop, identity, *, is_idempotent=False, anonymous=False):
         if type(binaryop) is not ParameterizedBinaryOp:
             raise TypeError("binaryop must be parameterized")
         self.binaryop = binaryop
@@ -776,6 +782,7 @@ class ParameterizedMonoid(ParameterizedUdf):
                     f"    identity{sig}"
                 )
         self.identity = identity
+        self._is_idempotent = is_idempotent
         if name is None:
             name = binaryop.name
         super().__init__(name, anonymous)
@@ -788,9 +795,16 @@ class ParameterizedMonoid(ParameterizedUdf):
         identity = self.identity
         if callable(identity):
             identity = identity(*args, **kwargs)
-        return Monoid.register_anonymous(binary, identity, self.name)
+        return Monoid.register_anonymous(
+            binary, identity, self.name, is_idempotent=self._is_idempotent
+        )
 
     commutes_to = TypedBuiltinMonoid.commutes_to
+
+    @property
+    def is_idempotent(self):
+        """True if ``monoid(x, x) == x`` for any x."""
+        return self._is_idempotent
 
     def __reduce__(self):
         name = f"monoid.{self.name}"
@@ -2491,7 +2505,7 @@ class Monoid(OpBase):
     as well as in the ``graphblas.ops`` combined namespace.
     """
 
-    __slots__ = "_binaryop", "_identity"
+    __slots__ = "_binaryop", "_identity", "_is_idempotent"
     is_commutative = True
     is_positional = False
     _custom_dtype = None
@@ -2517,12 +2531,14 @@ class Monoid(OpBase):
     }
 
     @classmethod
-    def _build(cls, name, binaryop, identity, *, anonymous=False):
+    def _build(cls, name, binaryop, identity, *, is_idempotent=False, anonymous=False):
         if type(binaryop) is not BinaryOp:
             raise TypeError(f"binaryop must be a BinaryOp, not {type(binaryop)}")
         if name is None:
             name = binaryop.name
-        new_type_obj = cls(name, binaryop, identity, anonymous=anonymous)
+        new_type_obj = cls(
+            name, binaryop, identity, is_idempotent=is_idempotent, anonymous=anonymous
+        )
         if not binaryop._is_udt:
             if not isinstance(identity, Mapping):
                 identities = dict.fromkeys(binaryop.types, identity)
@@ -2586,7 +2602,7 @@ class Monoid(OpBase):
         return op
 
     @classmethod
-    def register_anonymous(cls, binaryop, identity, name=None):
+    def register_anonymous(cls, binaryop, identity, name=None, *, is_idempotent=False):
[…]
