+++ graphblas/tests/test_op.py
@@ -1317,7 +1317,7 @@ def test_op_ss():
 def test_deprecated():
     with pytest.warns(DeprecationWarning, match="please use"):
         gb.unary.erf
-    with pytest.warns(DeprecationWarning, match="please use"):
+    with pytest.warns(DeprecationWarning, match="please use `gb.indexunary.rowindex`"):
         gb.unary.positioni
     with pytest.warns(DeprecationWarning, match="please use"):
         gb.binary.firsti
