### 2023-02-18 9aba9ea
Add `op.is_idempotent` property to Monoids that means `op(x, x) == x` (#388)
* Add `op.is_idempotent` property to Monoids that means `op(x, x) == x`

* Handle new setuptools DeprecationWarning
