### 2023-02-18 1a836ec
Enable more linting via ruff (#390)
