+++ pyproject.toml
@@ -138,10 +138,11 @@ filterwarnings = [
     # See: https://docs.python.org/3/library/warnings.html#describing-warning-filters
     # and: https://docs.pytest.org/en/7.2.x/how-to/capture-warnings.html#controlling-warnings
     "error",
+    # MAINT: we can drop support for sparse <0.13 at any time
     "ignore:`np.bool` is a deprecated alias:DeprecationWarning:sparse._umath",  # sparse <0.13
     # setuptools v67.3.0 deprecated `pkg_resources.declare_namespace` on 13 Feb 2023. See:
     # https://setuptools.pypa.io/en/latest/history.html#v67-3-0
-    # TODO: check if this is still necessary in 2025
+    # MAINT: check if this is still necessary in 2025
     "ignore:Deprecated call to `pkg_resources.declare_namespace:DeprecationWarning:pkg_resources",
 ]
 
@@ -177,42 +178,44 @@ select = [
     "F",  # pyflakes
     "E",  # pycodestyle Error
     "W",  # pycodestyle Warning
-    # "C90",  # mccabe
-    # "I",  # isort
-    # "N",  # pep8-naming
+    # "C90",  # mccabe (Too strict, but maybe we should make things less complex)
+    # "I",  # isort (Should we replace `isort` with this?)
+    "N",  # pep8-naming
     "D",  # pydocstyle
     "UP",  # pyupgrade
     "YTT",  # flake8-2020
-    # "ANN",  # flake8-annotations
-    "S",  # bandit
-    # "BLE",  # flake8-blind-except
-    # "FBT",  # flake8-boolean-trap
+    # "ANN",  # flake8-annotations (We don't use annotations yet)
+    "S",  # x. bandit
+    # "BLE",  # flake8-blind-except (Maybe consider)
+    # "FBT",  # flake8-boolean-trap (Why?)
     "B",  # flake8-bugbear
-    # "A",  # flake8-builtins
+    "A",  # flake8-builtins
     "COM",  # flake8-commas
     "C4",  # flake8-comprehensions
     "DTZ",  # flake8-datetimez
     "T10",  # flake8-debugger
-    # "DJ",  # flake8-django
-    # "EM",  # flake8-errmsg
+    # "DJ",  # flake8-django (We don't use django)
+    # "EM",  # xx. flake8-errmsg (Perhaps nicer, but too much work)
     "EXE",  # flake8-executable
     "ISC",  # flake8-implicit-str-concat
-    # "ICN",  # flake8-import-conventions
+    # "ICN",  # flake8-import-conventions (Doesn't allow "_" prefix such as `_np`)
     "G",  # flake8-logging-format
-    # "INP",  # flake8-no-pep420
+    "INP",  # flake8-no-pep420
     "PIE",  # flake8-pie
     "T20",   # flake8-print
-    # "PYI",  # flake8-pyi
+    # "PYI",  # flake8-pyi (We don't have stub files yet)
     "PT",  # flake8-pytest-style
-    "Q",  # flake8-quotes
-    # "RET",  # flake8-return
+    "Q",  # xxx. flake8-quotes
+    "RSE", # flake8-raise
+    "RET",  # flake8-return
+    # "SLF",  # flake8-self (We can use our own private variables--sheesh!)
     "SIM",  # flake8-simplify
-    # "TID",  # flake8-tidy-imports
+    # "TID",  # flake8-tidy-imports (Rely on isort and our own judgement)
     "TCH",  # flake8-type-checking
-    # "ARG",  # flake8-unused-arguments
-    # "PTH",  # flake8-use-pathlib
-    # "ERA",  # eradicate
-    # "PD",  # pandas-vet
[…]
