### 2023-02-20 34ced23
Use `np.random.default_rng` instead of legacy functions
