+++ pyproject.toml
@@ -278,14 +278,16 @@ ignore = [
     "TRY003",  # Avoid specifying long messages outside the exception class (Note: why?)
 ]
 [tool.ruff.per-file-ignores]
-"graphblas/core/operator.py" = ["S102"]
-"graphblas/tests/*py" = ["S101", "T201", "D103", "D100"]
-"graphblas/tests/test_dtype.py" = ["UP003"]
-"graphblas/tests/test_formatting.py" = ["E501"]
-"graphblas/**/__init__.py" = ["F401"]
-"scripts/*.py" = ["INP001"]
-"scripts/create_pickle.py" = ["F403", "F405"]
-"docs/*.py" = ["INP001"]
+"graphblas/core/operator.py" = ["S102"]  # exec is used for UDF
+"graphblas/core/ss/matrix.py" = ["NPY002"]  # numba doesn't support rng generator yet
+"graphblas/core/ss/vector.py" = ["NPY002"]  # numba doesn't support rng generator yet
+"graphblas/ss/_core.py" = ["N999"]  # We want _core.py to be underscopre
+"graphblas/tests/*py" = ["S101", "T201", "D103", "D100"]  # Allow assert, print, and no docstring
+"graphblas/tests/test_formatting.py" = ["E501"]  # Allow long lines
+"graphblas/**/__init__.py" = ["F401"]  # Allow unused imports (w/o defining `__all__`)
+"scripts/*.py" = ["INP001"]  # Not a package
+"scripts/create_pickle.py" = ["F403", "F405"]  # Allow `from foo import *`
+"docs/*.py" = ["INP001"]  # Not a package
 
 [tool.ruff.flake8-builtins]
 builtins-ignorelist = ["copyright", "format", "min", "max"]
