### 2023-02-20 23ff0ea
bump ruff (and use pathlib)
