### 2023-02-21 e0dcbf1
Update use of pathlib
