### 2023-02-21 b7e9485
Use `np.random.default_rng` instead of legacy functions (#392)
* Use `np.random.default_rng` instead of legacy functions

* bump ruff (and use pathlib)
