+++ pyproject.toml
@@ -27,6 +27,12 @@ keywords = [
     "matrix",
     "lagraph",
     "suitesparse",
+    "Networks",
+    "Graph Theory",
+    "Mathematics",
+    "network",
+    "discrete mathematics",
+    "math",
 ]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -44,8 +50,9 @@ classifiers = [
     "Intended Audience :: Other Audience",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
-    "Topic :: Scientific/Engineering :: Mathematics",
     "Topic :: Scientific/Engineering :: Information Analysis",
+    "Topic :: Scientific/Engineering :: Mathematics",
+    "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 dependencies = [
     "suitesparse-graphblas >=7.4.0.0, <7.5",
@@ -150,7 +157,6 @@ filterwarnings = [
 branch = true
 source = ["graphblas"]
 omit = [
-    "graphblas/_version.py",
     "graphblas/viz.py",  # TODO: test and get coverage for viz.py
 ]
 
@@ -185,7 +191,7 @@ select = [
     "UP",  # pyupgrade
     "YTT",  # flake8-2020
     # "ANN",  # flake8-annotations (We don't use annotations yet)
-    "S",  # x. bandit
+    "S",  # bandit
     # "BLE",  # flake8-blind-except (Maybe consider)
     # "FBT",  # flake8-boolean-trap (Why?)
     "B",  # flake8-bugbear
@@ -195,7 +201,7 @@ select = [
     "DTZ",  # flake8-datetimez
     "T10",  # flake8-debugger
     # "DJ",  # flake8-django (We don't use django)
-    # "EM",  # xx. flake8-errmsg (Perhaps nicer, but too much work)
+    # "EM",  # flake8-errmsg (Perhaps nicer, but too much work)
     "EXE",  # flake8-executable
     "ISC",  # flake8-implicit-str-concat
     # "ICN",  # flake8-import-conventions (Doesn't allow "_" prefix such as `_np`)
@@ -205,17 +211,17 @@ select = [
     "T20",   # flake8-print
     # "PYI",  # flake8-pyi (We don't have stub files yet)
     "PT",  # flake8-pytest-style
-    "Q",  # xxx. flake8-quotes
+    "Q",  # flake8-quotes
     "RSE", # flake8-raise
     "RET",  # flake8-return
     # "SLF",  # flake8-self (We can use our own private variables--sheesh!)
     "SIM",  # flake8-simplify
     # "TID",  # flake8-tidy-imports (Rely on isort and our own judgement)
-    "TCH",  # flake8-type-checking
+    # "TCH",  # flake8-type-checking (Note: figure out type checking later)
     # "ARG",  # flake8-unused-arguments (Sometimes helpful, but too strict)
-    # "PTH",  # flake8-use-pathlib (Is there a strong argument for this?)
+    "PTH",  # flake8-use-pathlib (Often better, but not always)
     # "ERA",  # eradicate (We like code in comments!)
-    # "PD",  # xl. pandas-vet (Intended for scripts that use pandas, not libraries)
+    # "PD",  # pandas-vet (Intended for scripts that use pandas, not libraries)
     "PGH",  # pygrep-hooks
     "PL",  # pylint
     "PLC",  # pylint Convention
@@ -225,6 +231,7 @@ select = [
     "TRY",  # tryceratops
     "NPY",  # NumPy-specific rules
     "RUF",  # ruff-specific rules
+    "ALL",  # Try new categories by default (making the above list unnecessary)
 ]
 external = [
[…]
