### 2023-02-21 0213c4b
Remove expired deprecation of using scipy matrix
