### 2023-02-21 096d374
Remove expired deprecation of using scipy matrix (#393)
* Remove expired deprecation of using scipy matrix

* Fix complex tests on Windows (asin and asinh on complex wrong on Windows)
