### 2023-02-22 #81 @NickleDave
Hi @eriknw, @jim22k, @SultanOrazbayev, brief update:
very happy to inform you that @tomalrussell will be guest editor for this review! 🎉   NetworkX contributor, developer of spatial & network tools like [snkit](https://github.com/tomalrussell/snkit).
I will let @tomalrussell take it from here!
