### 2023-02-24 #81 @tomalrussell
Hi @eriknw, @jim22k, @SultanOrazbayev, and thanks to @NickleDave for the introduction.

I've reached out to potential reviewers, and incidentally look forward to taking a closer look at `python-graphblas` myself. I'll update here as soon as, definitely within a week.
