### 2023-02-24 faaf1c7
Update usage of `from_dense` and `to_dense` in docs and notebooks
