### 2023-02-24 00c6539
Ignore warning from new version of pydata/sparse (we should investigate later)
