### 2023-02-25 632ebef
`to_pydata_sparse(v)` on Vector should create 1-d array
