+++ graphblas/tests/test_io.py
@@ -433,7 +433,8 @@ def test_vector_to_from_pydata_sparse():
     assert v.isequal(gb.Vector.from_coo(coords, data, dtype=dtypes.INT64), check_dtype=True)
 
     t = gb.io.to_pydata_sparse(v)
-    assert t == s
+    assert t.shape == s.shape
+    assert (t == s).all()
 
 
 @pytest.mark.skipif("not sparse")
@@ -445,7 +446,8 @@ def test_matrix_to_from_pydata_sparse():
     assert v.isequal(gb.Matrix.from_coo(*coords, data, dtype=dtypes.INT64), check_dtype=False)
 
     t = gb.io.to_pydata_sparse(v)
-    assert t == s
+    assert t.shape == s.shape
+    assert (t == s).all()
 
     # test ndim
     e = sparse.random(shape=(5, 5, 5), density=0)
@@ -464,7 +466,8 @@ def test_matrix_to_from_pydata_sparse():
     assert w.isequal(gb.Matrix.from_coo(*coords, data, dtype=dtypes.INT64), check_dtype=False)
 
     r = gb.io.to_pydata_sparse(w, format="gcxs")
-    assert r == g
+    assert r.shape == g.shape
+    assert (r == g).all()
     with pytest.raises(ValueError, match="format"):
         gb.io.to_pydata_sparse(w, format="badformat")
     with pytest.raises(TypeError, match="sparse.pydata"):
