### 2023-02-26 45a6084
Remove use of `from_numpy` and misc updates/cleanup
