+++ scripts/check_versions.sh
@@ -5,9 +5,10 @@
 # Tip: add `--json` for more information.
 conda search 'numpy[channel=conda-forge]>=1.24.2'
 conda search 'pandas[channel=conda-forge]>=1.5.3'
-conda search 'scipy[channel=conda-forge]>=1.10.0'
+conda search 'scipy[channel=conda-forge]>=1.10.1'
 conda search 'networkx[channel=conda-forge]>=3.0'
 conda search 'awkward[channel=conda-forge]>=2.0.8'
+conda search 'sparse[channel=conda-forge]>=0.14.0'
 conda search 'numba[channel=conda-forge]>=0.56.4'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
 conda search 'flake8-comprehensions[channel=conda-forge]>=3.10.1'
