+++ pyproject.toml
@@ -83,6 +83,7 @@ viz = [
 ]
 test = [
     "pytest",
+    "packaging",
     "pandas >=1.2",
     "scipy >=1.8",
 ]
@@ -94,6 +95,7 @@ complete = [
     "sparse >=0.12",
     "matplotlib >=3.5",
     "pytest",
+    "packaging",
 ]
 
 [tool.setuptools]
@@ -147,6 +149,9 @@ filterwarnings = [
     "error",
     # MAINT: we can drop support for sparse <0.13 at any time
     "ignore:`np.bool` is a deprecated alias:DeprecationWarning:sparse._umath",  # sparse <0.13
+    # sparse 0.14.0 (2022-02-24) began raising this warning; it has been reported and fixed upstream.
+    "ignore:coords should be an ndarray. This will raise a ValueError:DeprecationWarning:sparse._coo.core",
+
     # setuptools v67.3.0 deprecated `pkg_resources.declare_namespace` on 13 Feb 2023. See:
     # https://setuptools.pypa.io/en/latest/history.html#v67-3-0
     # MAINT: check if this is still necessary in 2025
@@ -271,6 +276,7 @@ ignore = [
     "PLR0913",  # Too many arguments to function call
     "PLR0915",  # Too many statements
     "PLR2004",  # Magic number used in comparison, consider replacing magic with a constant variable
+    "PLW2901",  # Outer for loop variable ... overwritten by inner assignment target (Note: good advice, but too strict)
     "RET502",  # Do not implicitly `return None` in function able to return non-`None` value
     "RET503",  # Missing explicit `return` at the end of function able to return non-`None` value
     "RET504",  # Unnecessary variable assignment before `return` statement
