### 2023-03-07 #81 @tomalrussell
:wave: Hi @sneakers-the-rat and @szhorvat! Thank you for volunteering to review
for pyOpenSci!

Please get in touch with any questions or concerns! Your review is due in three weeks: 29th March 2023
