### 2023-03-07 python-graphblas/python-graphblas#396 @szhorvat
[pyos] Broken link in docs
On https://python-graphblas.readthedocs.io/en/latest/user_guide/fundamentals.html, the graphblas.org link is broken, probably due to a double-`:` in `https:://graphblas.org`

https://github.com/pyOpenSci/software-submission/issues/81
