### 2023-03-07 5c7f080
Update fundamentals.rst
Resolve https://github.com/python-graphblas/python-graphblas/issues/396
