### 2023-03-07 #81 @szhorvat
@tomalrussell Could you please clarify if the authors are expected to follow all the above checklist items to the letter? (E.g., all mentioned badges mandatory, links to all "vignettes" _from the README_, all functions have examples of use, etc.)

----

@eriknw :

> includes documentation with examples for all functions.
>
>     * We're working on this! The C API and SuiteSparse:GraphBLAS C library are both well documented. We have a very large API surface area to cover, so "documentation with examples for all functions" is a really, really high bar, but one I hope we achieve someday :)

Indeed, the documentation of most Python projects won't give usage examples for every single function, and doing so would definitely be a lot of work. But just to show that it is possible, and often tremendously useful to users, I wanted to point to Mathematica's documentation where each function has not one, but _many_ examples. See e.g. [LinearSolve](https://reference.wolfram.com/language/ref/LinearSolve.html). I tried to follow the same with my IGraph/M package for Mathematica. R packages also often have at least one example for each function.
