### 2023-03-07 2a6f60e
Handle deprecation warning added in setuptools v67.5.0
