### 2023-03-07 python-graphblas/python-graphblas#398 @eriknw
Yeah, this is worth attention, thanks @szhorvat. Right now we defer to `networkx.to_scipy_sparse`, but they "solved" this problem by adding to the documentation and providing a recipe for how to double the diagonals.  Their recipe involves creating a dense array, which is not so great.

So, at the very minimum, we should document how self-edges are handled in different cases. We should also consider adding a keyword argument to control behavior and changing the default behavior.

See NetworkX discussion about this here: https://github.com/networkx/networkx/pull/1078
And their docs here: https://networkx.org/documentation/stable/reference/generated/networkx.convert_matrix.to_scipy_sparse_array.html

As for how we handle self-edges in `graphblas-algorithms`... this has been a big question mark. Our current plan is to match networkx, and to create tests that run and compare against networkx to help us determine what networkx does.
