### 2023-03-07 80a92e7
Update fundamentals.rst (#397)
Resolve https://github.com/python-graphblas/python-graphblas/issues/396
