### 2023-03-07 python-graphblas/python-graphblas#396 @eriknw
Thanks for the report @szhorvat and the fix @SultanOrazbayev! :heart:
