### 2023-03-07 python-graphblas/python-graphblas#396 @szhorvat
I won't open a separate issue for this, as it's not a big deal, but the doc page doesn't look great in dark mode, and the header image just went missing. I think it was there earlier today.
