### 2023-03-07 python-graphblas/python-graphblas#398 @eriknw
I agree with your points @szhorvat, and thanks for the clear explanations. It's helpful to have a "paper trail", and I'm looking forward on your perspective on more things. We'll discuss this issue in our weekly meeting tomorrow, but here are my current thoughts:

- the current behavior for undirected is "wrong" for the reasons you state and the default behavior should be as you describe
- how self-edges are converted to GraphBLAS adjacency matrix should be explicitly documented
- we should consider adding a keyword argument to control behavior of undirected self-edges
  - are there strong reasons for using the current behavior?
  - having a keyword can serve as a "flag" to bring attention to how to handle self-edges
- we should be consistent when converting undirected graphs from other graph packages
  - converting to/from igraph and graph-tool are within scope
- we should consider upstreaming this concern back to NetworkX, b/c they are using scipy.sparse more often
  - since the default behavior is "wrong", it may be easy for them to introduce subtle errors

> Does the GraphBLAS standard say anything about the interpretation of symmetric matrices as undirected graphs? Or does it sidestep the problem by effectively treating everything as directed? I'm new to GraphBLAS, but I guess the latter?

Despite having "Graph" in the name, GraphBLAS says very little about the interpretation of matrices. The same matrix can be interpreted differently whether it's directed or undirected, and the same graph can have different associated matrices: adjacency, incidence, Laplacian, etc. So, I would say the GraphBLAS approach is to allow users to adhere to well-established theory relating matrices, linear algebra, and graphs. If you know you have an adjecency matrix for an undirected graph, you can treat it as such or wrap it in a `Graph` class as is done in `LAGraph` and `graphblas_algorithms`.

So, we should follow the spirit of adhering to well-established theory, which I think you've convinced me that in this case means changing how we convert self-edges for undirected graphs from networkx.
