### 2023-03-08 709750c
Fix path to logo in docs (include `_static` in path)
