### 2023-03-08 python-graphblas/python-graphblas#396 @eriknw
> I won't open a separate issue for this, as it's not a big deal, but the doc page doesn't look great in dark mode, and the header image just went missing. I think it was there earlier today.

Thanks for reporting. I fixed the header logo in #400. I think this _just_ happened.

As for not looking good in dark mode, the "stable" build of the docs looks fine to me:
https://python-graphblas.readthedocs.io/en/stable/index.html

The stable build uses `pydata-sphinx-theme` 0.12.0. The "latest" build uses version 0.13.1. Looks like something changed (CC @jim22k). Perhaps we should pin to a specific version.

Also, should we point to "stable" docs instead of "latest" docs? As a developer, it's nice to have issues reported for latest.

Regarding the original issue of a broken link in the doc, perhaps we should add a check to ensure links work. For example, `make linkcheck` mostly does this, but seems to have issues for relative links for me.
