### 2023-03-08 #81 @tomalrussell
Thanks both, great start 💯   - I'll aim to check in occasionally or as needed.

> Could you please clarify if the authors are expected to follow all the above checklist items to the letter? (E.g., all mentioned badges mandatory, links to all "vignettes" _from the README_, all functions have examples of use, etc.)

I would comment on anything you notice and let the authors respond, we can always exercise judgment if "to the letter" seems unhelpful.
