### 2023-03-08 python-graphblas/python-graphblas#396 @SultanOrazbayev
Just a quick note to confirm the docs render well for me in the dark mode also and I can see the logo.
