### 2023-03-08 python-graphblas/python-graphblas#396 @jim22k
I just updated readthedocs config so that "stable" is the default.

@eriknw I'll update your PR with a fix for the dark mode.
