### 2023-03-08 61e6953
Fix dark mode for docs
- Update and pin pydata-sphinx-theme to 0.13.1
- Remove circle around light/dark theme change icon
- Remove "Show Source" button
