### 2023-03-10 python-graphblas/python-graphblas#402 @szhorvat
[pyos] Review checklist items
This is regarding the checklist items for the pyOpenSci review, see https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1458245662

 - [ ] It would be nice to briefly mention optional dependencies in the [README](https://github.com/python-graphblas/python-graphblas#install), as that's where most people look first (I know they're already in the docs).
 - [ ] Please ensure that the author/maintainer list, as indicated in this repo, matches the author/maintainer list in the submission (see https://github.com/pyOpenSci/software-submission/issues/81#issue-1571161465)

https://github.com/pyOpenSci/software-submission/issues/81
