### 2023-03-10 #81 @szhorvat
Some of the review checklist items feel a bit R-ish and I wonder if these are intentional, or they just haven't been updated yet after borrowing from rOpenSci. Is the term "vignette" used in the context of Python?  Examples for each function are not typical for Python either (though I have to say that in my personal opinion it wouldn't be bad to borrow this habit from R 😉 )
