### 2023-03-10 #81 @NickleDave
> I wonder if these are intentional.

Yes. If you prefer the term "tutorial", then feel free to use that.

> though I have to say that in my personal opinion it wouldn't be bad to borrow this habit from R

Exactly.

For example see discussion of vignettes here on PyGMT: https://github.com/pyOpenSci/software-submission/issues/43#issuecomment-994023562

Or discussion of examples here on the jointly review: https://github.com/pyOpenSci/software-submission/issues/45#issuecomment-1001172229
