### 2023-03-10 python-graphblas/python-graphblas#403 @szhorvat
[pyos] Installation from PyPI only supports some platforms
Looking at the installation instructions ([readme](https://github.com/python-graphblas/python-graphblas#install) or [docs](https://python-graphblas.readthedocs.io/en/stable/getting_started/#installation)), I got the impression that a simple `pip install python-graphblas` would work for most users, but this does not appear to be the case. Originally I installed from Conda-Forge, so I did not notice this.

It seems that the essential dependency [suitesparse-graphblas](https://pypi.org/project/suitesparse-graphblas/) only provides binaries for Linux. Thus a simple `pip install python-graphblas` will fail for macOS and Windows users, and I expect most of them will be confused.

Can you please explain this situation better, so that the installation experience is as smooth as possible? I.e.,

 - Ideally, provide binaries for all platforms.
 - If this is not possible, point out the issues and their solutions.
     - A simple solution is to use Anaconda.
     - A more complicated one is to compile `suitesparse-graphblas` from source. There should be instructions for this.

https://github.com/pyOpenSci/software-submission/issues/81
