### 2023-03-10 #81 @NickleDave
> Happy to also do the inverse (link back to opened issues) as well

@sneakers-the-rat if you can *also* link in comments that just helps make sure we don't miss it

> So - for this review, will only post text of review when checklist finished, is that what you had in mind?

yes please

@szhorvat @sneakers-the-rat the checklists and workflow we have are fairly well established and up to date. They're designed in part so that @lwasser can parse the reviews in an automated way.

We're of course happy to learn from other communities-- @lwasser frequently talks with Arfon from JOSS, for example--but please let's try to follow the workflow we have in place for now.

I realize from how this review has started we probably can do more to make the workflow clearer. We're happy to hear feedback after the review on how we can improve the workflow and make it clearer--we can discuss in Slack or in the forum.

Thank you!
