### 2023-03-10 #81 @szhorvat
Although I did not respond in words, I noted the comment from @lwasser and I am happy to follow (I believe I am following) the process. If I missed anything, please point it out explicitly, and I'll try to correct it.
