### 2023-03-10 python-graphblas/python-graphblas#403 @szhorvat
I tried to let `suitesparse-graphblas` compile from sources on my machine. I am using macOS 10.14, and I have SuiteSparse_GraphBLAS 4.0.3 installed through MacPorts. Since MacPorts places headers and libraries in a non-standard location, I used the following installation command:

```
CPPFLAGS=-I/opt/local/include LDFLAGS=-L/opt/local/lib pip install python-graphblas
```

The following (and many similar) errors occur:

```
      clang -Wno-unused-result -Wsign-compare -Wunreachable-code -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /Users/szhorvat/opt/miniconda3/envs/gb2/include -fPIC -O2 -isystem /Users/szhorvat/opt/miniconda3/envs/gb2/include -I/opt/local/include -I/Users/szhorvat/opt/miniconda3/envs/gb2/include -I/Users/szhorvat/opt/miniconda3/envs/gb2/include/python3.10 -c suitesparse_graphblas/_graphblas.c -o build/temp.macosx-10.9-x86_64-cpython-310/suitesparse_graphblas/_graphblas.o
      suitesparse_graphblas/_graphblas.c:4655:12: error: use of undeclared identifier 'GxB_COMPRESSION'
        int n = (GxB_COMPRESSION) <= 0;
                 ^
      suitesparse_graphblas/_graphblas.c:4656:30: error: use of undeclared identifier 'GxB_COMPRESSION'
        *o = (unsigned long long)((GxB_COMPRESSION) | 0);  /* check that GxB_COMPRESSION is an integer */
                                   ^
      suitesparse_graphblas/_graphblas.c:4662:12: error: use of undeclared identifier 'GxB_IMPORT'
        int n = (GxB_IMPORT) <= 0;
                 ^
```

I'm guessing SuiteSparse_GraphBLAS 4.0.3 is too old. I'll try with a newer one at a later point.
