### 2023-03-10 561f8e2
Update README.md
optional dependencies copied from docs
