### 2023-03-10 673650d
Fix path to logo in docs (include `_static` in path) (#400)
* Fix path to logo in docs (include `_static` in path)

* Fix dark mode for docs

- Update and pin pydata-sphinx-theme to 0.13.1
- Remove circle around light/dark theme change icon
- Remove "Show Source" button

---------

Co-authored-by: Jim Kitchen <jim22k@gmail.com>
