+++ docs/env.yml
@@ -19,5 +19,5 @@ dependencies:
     - commonmark  # For RTD
     - nbsphinx
     - numpydoc
-    - pydata-sphinx-theme
-    - sphinx-panels
+    - pydata-sphinx-theme=0.13.1
+    - sphinx-panels=0.6.0
