### 2023-03-10 python-graphblas/python-graphblas#403 @eriknw
> Ideally, provide binaries for all platforms.

This is _almost_ ready: https://github.com/GraphBLAS/python-suitesparse-graphblas/pull/66
and is high on our roadmap (WIP): https://github.com/python-graphblas/python-graphblas/issues/360#issuecomment-1404610964

I think there is one last issue to figure out: how to create wheels with OpenMP on OS X Arm. Creating wheels with OpenMP is kind of fraught and is something we should probably ask a maintainer community such as Scientific-Python, PyOpenSci, or even StackOverflow. Any pointers?

(Note: I used to manually build wheels for Windows and OS X, but I lost access to the machine I used to do this, and it was pretty painful to do).

> A more complicated one is to compile suitesparse-graphblas from source. There should be instructions for this.

That's a good idea. We actually test compiling `python-suitesparse-graphblas` from source (and from latest main branch):

https://github.com/python-graphblas/python-graphblas/blob/673650dde0d11f169fb56eb516a427239fc45d19/.github/workflows/test_and_build.yml#L204-L207

When we test this in CI, we install SuiteSparse:GraphBLAS from conda-forge, so we don't need to manipulate any paths.

Building `python-suitesparse-graphblas` from source with SuiteSparse:GraphBLAS installed _elsewhere_ is possible. This definitely deserves being documented once we figure out how to set the paths. This is kind of an advanced workflow, so we'll probably only include instructions for Linux at first where it's most likely. I think your attempt should work with a newer version of SuiteSparse:GraphBLAS, but it would be nice to confirm this.

> I'm guessing SuiteSparse_GraphBLAS 4.0.3 is too old. I'll try with a newer one at a later point.

Yeah, that's _way_ too old. `python-suitesparse-graphblas` version matches the version of SuiteSparse:GraphBLAS, and we require at least version 7.4.0.
