### 2023-03-10 python-graphblas/python-graphblas#403 @szhorvat
> I think there is one last issue to figure out: how to create wheels with OpenMP on OS X Arm. Creating wheels with OpenMP is kind of fraught and is something we should probably ask a maintainer community such as Scientific-Python, PyOpenSci, or even StackOverflow. Any pointers?

python-igraph has binaries for most platforms, including macOS on ARM, so perhaps the setup there can be of interest to you? https://github.com/igraph/python-igraph/blob/main/.github/workflows/build.yml#L102  I'm not deeply familiar with the details, but I can find them out for you, if you have questions.

What I recall:

 - On Windows OpenMP shouldn't be an issue if the OpenMP version provided by MSVC is sufficient for you.
 - On macOS, we build libomp from sources, to make sure that we can control the macOS version that is targeted (we currently support 10.9 or later), and so that we can link it statically, making the igraph binary self-contained. Do not try to link to Homebrew's libomp!   That will not be portable, and you won't be able to support older macOS versions.
