### 2023-03-10 python-graphblas/python-graphblas#402 @szhorvat
@eriknw Marginally related: Is the Discord channel, mentioned in the readme, still active / supported? It appears quite dead.
