### 2023-03-10 python-graphblas/python-graphblas#402 @eriknw
> Is the Discord channel, mentioned in the readme, still active / supported?

Yeah, it's still active/supported. Things often become 1-on-1 pings so the channel doesn't get too noisy with mundane things (but maybe these would be okay to show that we're there?).

The Discord is a perfect place for your recent comments/questions. Sorry I haven't replied yet--I just traveled across the world and am still recovering. I don't know if @jim22k regularly watches the Discord.
