### 2023-03-11 1ede759
Update pyproject.toml
Taken the list from https://github.com/python-graphblas/python-graphblas/graphs/contributors

excluded Dependabot, though not against robots in general.
