### 2023-03-11 7d9452b
Update pyproject.toml
Removed `contributors` (not a valid key for project description it seems):
```
contributors = [
    {GitHub_handle = "eriknw"},
    {GitHub_handle = "jim22k"},
    {GitHub_handle = "ParticularMiner"},
    {GitHub_handle = "SultanOrazbayev"},
    {GitHub_handle = "seibert"},
]
```
