### 2023-03-11 a0c152c
Update README.md (#404)
* Update README.md

Adds Python version badges

optional dependencies copied from docs

* Update pyproject.toml

Add Sultan Orazbayev as maintainer
