### 2023-03-11 python-graphblas/python-graphblas#402 @eriknw
Done. Good suggestions, @szhorvat 👍
