### 2023-03-13 2faf499
Maybe trust `ruff` to fix some things
But run `autoflake`, `isort`, `pyupgrade`, and `black` first (for now).
