+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
       - id: validate-pyproject
         name: Validate pyproject.toml
   - repo: https://github.com/myint/autoflake
-    rev: v2.0.1
+    rev: v2.0.2
     hooks:
       - id: autoflake
         args: [--in-place]
@@ -48,6 +48,11 @@ repos:
     hooks:
       - id: black
       - id: black-jupyter
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.254
+    hooks:
+      - id: ruff
+        args: [--fix-only]
   - repo: https://github.com/PyCQA/flake8
     rev: 6.0.0
     hooks:
