### 2023-03-13 b0544d8
Add a few notes to pre-commit hooks
