### 2023-03-13 #81 @lwasser
thank you all! @sneakers-the-rat @szhorvat we are super happy to discuss this more (maybe we could do this in [discourse](https://pyopensci.discourse.group/) as an open spot to discuss? As @NickleDave said, we want to learn from other communities, however it's also important that our review approach is consistent. As such we want to discuss as a community before making changes to our processes for review. I hope that make sense. I am looking forward to chatting more!! And appreciate everyone's time here on this review!
