### 2023-03-13 5969089
Add `fast-matrix-market` to optional dependencies documentation
