+++ .github/workflows/publish_pypi.yml
@@ -35,7 +35,7 @@ jobs:
       - name: Check with twine
         run: python -m twine check --strict dist/*
       - name: Publish to PyPI
-        uses: pypa/gh-action-pypi-publish@v1.6.4
+        uses: pypa/gh-action-pypi-publish@v1.7.1
         with:
           user: __token__
           password: ${{ secrets.PYPI_TOKEN }}
