+++ .pre-commit-config.yaml
@@ -24,12 +24,6 @@ repos:
     hooks:
       - id: validate-pyproject
         name: Validate pyproject.toml
-  # We can probably remove `autoflake` if we come to trust `ruff --fix`
-  - repo: https://github.com/myint/autoflake
-    rev: v2.0.2
-    hooks:
-      - id: autoflake
-        args: [--in-place]
   # We can probably remove `isort` if we come to trust `ruff --fix`,
   # but we'll need to figure out the configuration to do this in `ruff`
   - repo: https://github.com/pycqa/isort
@@ -53,7 +47,7 @@ repos:
       - id: black
       - id: black-jupyter
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.254
+    rev: v0.0.255
     hooks:
       - id: ruff
         args: [--fix-only]
@@ -66,7 +60,6 @@ repos:
         additional_dependencies: &flake8_dependencies
         # These versions need updated manually
         - flake8==6.0.0
-        - flake8-comprehensions==3.10.1  # Consider removing and rely on `ruff` instead
         - flake8-bugbear==23.2.13
         - flake8-simplify==0.19.3
   - repo: https://github.com/asottile/yesqa
@@ -82,7 +75,7 @@ repos:
         additional_dependencies: [tomli]
         files: ^(graphblas|docs)/
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.254
+    rev: v0.0.255
     hooks:
       - id: ruff
   - repo: https://github.com/sphinx-contrib/sphinx-lint
