+++ pyproject.toml
@@ -209,7 +209,7 @@ select = [
     "B",  # flake8-bugbear
     "A",  # flake8-builtins
     "COM",  # flake8-commas
-    "C4",  # flake8-comprehensions
+    "C40",  # flake8-comprehensions
     "DTZ",  # flake8-datetimez
     "T10",  # flake8-debugger
     # "DJ",  # flake8-django (We don't use django)
