### 2023-03-15 #81 @sneakers-the-rat
@eriknw - clarification q on your intended audience statement above for the sake of providing feedback on the docs: are you expecting people to have prior experience with graphBLAS? or is the intention to be able to bring someone with ~reasonable understanding of sparse data or graphs up to speed as well? Would be happy to be the latter kind of reader and let you know what parts are clear/unclear if that's the intention, but if the expectation is for someone to already understand graphBLAS i'll spare u.
