+++ scripts/check_versions.sh
@@ -12,5 +12,5 @@ conda search 'sparse[channel=conda-forge]>=0.14.0'
 conda search 'fast_matrix_market[channel=conda-forge]>=1.4.5'
 conda search 'numba[channel=conda-forge]>=0.56.4'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
-conda search 'flake8-bugbear[channel=conda-forge]>=23.2.13'
+conda search 'flake8-bugbear[channel=conda-forge]>=23.3.12'
 conda search 'flake8-simplify[channel=conda-forge]>=0.19.3'
