+++ .pre-commit-config.yaml
@@ -47,7 +47,7 @@ repos:
       - id: black
       - id: black-jupyter
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.255
+    rev: v0.0.256
     hooks:
       - id: ruff
         args: [--fix-only]
@@ -60,7 +60,7 @@ repos:
         additional_dependencies: &flake8_dependencies
         # These versions need updated manually
         - flake8==6.0.0
-        - flake8-bugbear==23.2.13
+        - flake8-bugbear==23.3.12
         - flake8-simplify==0.19.3
   - repo: https://github.com/asottile/yesqa
     rev: v1.4.0
@@ -75,7 +75,7 @@ repos:
         additional_dependencies: [tomli]
         files: ^(graphblas|docs)/
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.255
+    rev: v0.0.256
     hooks:
       - id: ruff
   - repo: https://github.com/sphinx-contrib/sphinx-lint
