### 2023-03-15 #81 @lwasser
FWIW - i think that it could be nice if the overview provides enough context for a user less experienced with graphBLAS to follow. it might welcome more users. for instance i was unfamiliar with the term "sparse matrix". then i looked i up and quickly realized i've used that kind of data structure before. so i was familiar with the data type but not the word. of course @eriknw knows best what is feasible here in terms of helping inform users of the package's goals.

my two cents :)
