### 2023-03-15 #81 @eriknw
> are you expecting people to have prior experience with graphBLAS? or is the intention to be able to bring someone with ~reasonable understanding of sparse data or graphs up to speed as well?

Yeah, I think we definitely want to support both, and I know there's work to do on the docs, so all feedback is appreciated. It would be great if we had documentation at an even higher level too such as example applications that data scientists can use without even needing to know much about graphs or sparse.

My personal priority right now is to improve _maintenance_ documentation so that maintenance can continue smoothly if I disappeared for a time (for any reason). My second documentation priority is to establish best practices for how to add examples to docstrings. We have other maintainers and contributors who have improved and plan to continue improving documentation.

> of course @eriknw knows best what is feasible here in terms of helping inform users of the package's goals.

I think you're giving me (and engineers in general) too much credit! I/we are sometimes much too close to engineering problems to remember the bigger picture of "who cares and why?". I would like for us to pivot to a "product-led growth" mindset/orientation this year now that much of the hard engineering is done.
