### 2023-03-16 #81 @lwasser
@eriknw thank you for this feedback. so it sounds like it might be useful for @sneakers-the-rat @szhorvat to provide feedback from their perspective about what is confusing in the documentation. (such great questions you both asked erik!!) i wonder if there is some way for us to also help you with revisions if you need that support given your focus on maintenance docs (which i agree is profoundly important for long term maintenance). it's something that i've been thinking about but we don't have a huge team just yet to support that just yet...
