### 2023-03-16 python-graphblas/python-graphblas#398 @eriknw
Perhaps the easiest thing for us to do here (at least for now) is to double the diagonals ourselves when converting from NetworkX, such as:
```python
if not G.is_directed():
    A += gb.select.diag(A)
```
