### 2023-03-17 python-graphblas/python-graphblas#398 @eriknw
If we double the diagonals by default when converting _from_ NetworkX, then we ought to halve the diagonals when converting _to_ a NetworkX undirected Graph by default.

What are good keyword arguments to control this behavior?
