### 2023-03-17 python-graphblas/python-graphblas#398 @szhorvat
> igraph and graph-tool both use 2.

I have to make a correction here. The Python interface of igraph has not fully transitioned yet, and currently uses 1 by default. graph-tool does use 2.

> What are good keyword arguments to control this behavior?

What we are heading towards in igraph is using a `loops` parameter with settings `"ignore"`, `"once"`, `"twice"` for all functions it makes sense to control this:

https://python.igraph.org/en/stable/api/igraph.GraphBase.html#Adjacency
