### 2023-03-17 python-graphblas/python-graphblas#398 @szhorvat
I wanted to give you a heads up that I won't be reachable until the week after next. I will think more carefully about these issues when I'm back and will try to finish the review by April 2nd.
