### 2023-03-17 python-graphblas/python-graphblas#398 @eriknw
Thanks for the suggestions, questions, and heads up @szhorvat. I expect to be especially busy the next two weeks, so this is good timing for a pause. I do appreciate your scrutiny and involvement with your careful review--you're helping to improve things and push things forward :)

We would like to redo `from_networkx`/`to_networkx` and also add igraph conversions to try to be consistent. We've discussed this at our previous two meetings, but don't have specific APIs yet. We want to support:
- directed/undirected
- keyword argument for how to handle self-edges
- adjacency and incidence matrices / graphs and multigraphs (and maybe someday hypergraphs)

> It seems to me that the GraphBLAS approach of using adjacency matrices is not very suitable for multigraphs.

Right, adjacency matrices aren't the way to go for weighted multigraphs. For unweighted, using the multiplicity for the edge weight can sometimes work as you indicated. We'll need to use incidence matrices in general though. Now I wonder: how are self-edges handled by incidence matrices? Instead of a single row (or column) for an edge with two values to indicate the two end points, I suppose you'd need two rows (or columns) for an edge with one value each to indicate the self-loop.
