### 2023-03-17 python-graphblas/python-graphblas#398 @eriknw
@jim22k, regarding the shape of the incidence matrix, since we assume row-oriented-ness, I think it makes the most sense to have `nodes X edges` (node ids as row indices and edge id as column indices) by default. This way, it's easy to ask the question "which edges are incident upon node `i`" by simply taking `B[i, :]`, which ought to be faster than `B[:, i]` if the underlying storage is CSR or DCSR. This also seems to be the more common convention.
