+++ .pre-commit-config.yaml
@@ -47,7 +47,7 @@ repos:
       - id: black
       - id: black-jupyter
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.256
+    rev: v0.0.257
     hooks:
       - id: ruff
         args: [--fix-only]
@@ -75,7 +75,7 @@ repos:
         additional_dependencies: [tomli]
         files: ^(graphblas|docs)/
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.256
+    rev: v0.0.257
     hooks:
       - id: ruff
   - repo: https://github.com/sphinx-contrib/sphinx-lint
