### 2023-03-20 python-graphblas/python-graphblas#398 @SultanOrazbayev
> Right, adjacency matrices aren't the way to go for weighted multigraphs. For unweighted, using the multiplicity for the edge weight can sometimes work as you indicated. We'll need to use incidence matrices in general though. Now I wonder: how are self-edges handled by incidence matrices? Instead of a single row (or column) for an edge with two values to indicate the two end points, I suppose you'd need two rows (or columns) for an edge with one value each to indicate the self-loop.

Wouldn't this be handled by a simple one-valued row? (so the column with the only value is the self-edge)
