### 2023-03-20 python-graphblas/python-graphblas#398 @SultanOrazbayev
This is not directly relevant to the discussion here, but since incidence matrices were brought up, there is a relationship between incidence and adjacency matrix given [here](https://mathworld.wolfram.com/IncidenceMatrix.html):

> The incidence matrix C of a graph and adjacency matrix L of its line graph are related by
> $L=C^{T}C-2I $

I don't have the book referenced there, but at least for the special case of $ L=2I $, the corresponding value for incidence matrix is $ C=2I $.
