### 2023-03-20 python-graphblas/python-graphblas#398 @eriknw
The relationship I'm more familiar with the relationship connecting the incidence matrix `B` to the Laplacian `L`, adjacency matrix `A`, and diagonal degree matrix `D`.

For unweighted, undirected graph:
```
B @ B.T = D + A
```
For unweighted, directed graph:
```
B @ B.T = D - A = L
```
So, for these to hold, I think the single value in a column for a self-edge should be `2` for unweighted/undirected, and `0` for unweighted/directed.
