### 2023-03-20 eb60ec5
Note that `scipy` is needed for all backends for mmread and mmwrite
