### 2023-03-20 python-graphblas/python-graphblas#406 @sneakers-the-rat
[pyos][docs] In-page TOC overlaps with page content in docs
Not sure if this is an upstream issue with `pydata-sphinx-theme` - but when loading the docs on mobile, or just with a narrow screen, they look like this, which makes them unreadable. Tested in firefox and chrome, though sometimes chrome doesn't show the in-page TOC at all, which is interesting.

If this is an upstream issue with the theme, feel free to close, just wanted to put on your radar

Mobile (simulated):

Related to PyOS review: https://github.com/pyOpenSci/software-submission/issues/81
