### 2023-03-20 python-graphblas/python-graphblas#407 @sneakers-the-rat
[pyos][docs] Link to or include example notebooks in docs
Approaching the package as a naïve potential user, I was looking for some examples after the 'getting started' section. Would recommend linking to the example notebooks as an entry in the Getting Started toctree: https://github.com/python-graphblas/python-graphblas/blob/main/docs/getting_started/index.rst

It's possible to put HTTP links directly in the toctree, or else you could render the notebooks themselves using an extension, eg.: https://docs.readthedocs.io/en/stable/guides/jupyter.html

Part of https://github.com/pyOpenSci/software-submission/issues/81
