### 2023-03-20 python-graphblas/python-graphblas#410 @eriknw
I agree it would be great to more thoroughly document the operators such as from `gb.binary`. We can/should probably even group them as "builtin, part of the spec", "builtin, SuiteSparse:GraphBLAS-specific", and UDF. We can include a lot of relavent information in a table view for the different operators, and we can even indicate what the numpy-equivalent is when possible (since we have this info in e.g. `gb.binary.numpy` namespace). Whatever we do though would need to be autogenerated.

> Aside from the wrapped functions/classes, there are a large number of functions/classes that lack a docstring (351 by an inspection that also includes trivially missing docstrings like module-level docstrings. You all will have a better sense of what counts as "user-facing" in this context, but presumably some subset of those are intended for use by the user.

A few of them, yes, but it's not quite as dire as it looks at a glance.  I commented out the following ignore codes and ran `ruff` via `pre-commit run --all-files`:
https://github.com/python-graphblas/python-graphblas/blob/c3baea8d55df1ab015e27c9b9db16e40f64f446b/pyproject.toml#L248-L255
This probably has similar results to your inspection. When looking through the results, I see very little that I regard as user-facing or particularly important. It also doesn't recognize that `@wrapdoc` decorator adds a docstring. To be sure, though, there are a few user-facing methods and properties such as `.new(...)` that could use docstrings, and we can do better in general. If you see any _specific_ methods, functions, properties, or error messages that should be added or improved during the course of your experiments, please comment here or raise new issues :)

> the positively massive core/operator.py file

Heh, every time I've tried to split it up, I've failed. Here's a quick sketch of the important bits: we call e.g. `UnaryOp._initialize`, which then does a regex on `lib` (the objects from cffi that wraps the C implementation) to find unary operators, then we add all the typed unary operators (as `TypedBuiltinUnaryOp`) to `UnaryOp`, which we save to `gb.unary`. We do the same thing for binary operators, monoids, semirings, and indexunaryops.
