### 2023-03-20 python-graphblas/python-graphblas#408 @eriknw
Yeah, I agree these could use attention and would be nice to get some of them into the documentation.

Your assessment is correct: these are a hodgepodge of notebooks going back _years_ that have not been diligently kept up to date (except to run) or organized. I think it would be nice to update them to use current best practices (such as `gb.semiring.min_plus(A @ B)` instead of `A.mxm(B, gb.semiring.min_plus)`), and then figure out how to best integrate them into docs (and update the discussion in the notebook as appropriate).

Relatedly, we could even get more examples from here (most of these need cleaned up and updated even more):
https://github.com/metagraph-dev/grblas-recipes
