### 2023-03-20 python-graphblas/python-graphblas#409 @eriknw
Just dropping in to say thank you for creating issues such as this one with specific suggestions for improving the documentation and overall user-experience--they're very helpful :)
