### 2023-03-22 4920dc8
Read Matrix Market with `fast_matrix_market` (#391)
* Read Matrix Market with `fast_matrix_market`

* Update usage of `from_dense` and `to_dense` in docs and notebooks

* Ignore warning from new version of pydata/sparse (we should investigate later)

* `to_pydata_sparse(v)` on Vector should create 1-d array

* bump awkward to 2.1.1

* Maybe trust `ruff` to fix some things

But run `autoflake`, `isort`, `pyupgrade`, and `black` first (for now).

* Add a few notes to pre-commit hooks

* Add `fast-matrix-market` to optional dependencies documentation

* Drop autoflake in pre-commit (use ruff inst
[…]
