+++ graphblas/io.py
@@ -26,6 +26,7 @@ def draw(m):  # pragma: no cover
     _warn(
         "`graphblas.io.draw` is deprecated; it has been moved to `graphblas.viz.draw`",
         DeprecationWarning,
+        stacklevel=2,
     )
     viz.draw(m)
 
@@ -93,6 +94,7 @@ def from_numpy(m):  # pragma: no cover (deprecated)
         "`graphblas.io.from_numpy` is deprecated; "
         "use `Matrix.from_dense` and `Vector.from_dense` instead.",
         DeprecationWarning,
+        stacklevel=2,
     )
     if m.ndim > 2:
         raise _GraphblasException("m.ndim must be <= 2")
@@ -336,6 +338,7 @@ def to_numpy(m):  # pragma: no cover (deprecated)
         "`graphblas.io.to_numpy` is deprecated; "
         "use `Matrix.to_dense` and `Vector.to_dense` instead.",
         DeprecationWarning,
+        stacklevel=2,
     )
     try:
         import scipy  # noqa: F401
@@ -566,16 +569,27 @@ def to_pydata_sparse(A, format="coo"):
     return s.asformat(format)
 
 
-def mmread(source, *, dup_op=None, name=None):
+def mmread(source, engine="auto", *, dup_op=None, name=None, **kwargs):
     """Create a GraphBLAS Matrix from the contents of a Matrix Market file.
 
     This uses `scipy.io.mmread
-    <https://docs.scipy.org/doc/scipy/reference/generated/scipy.io.mmread.html>`_.
+    <https://docs.scipy.org/doc/scipy/reference/generated/scipy.io.mmread.html>`_
+    or `fast_matrix_market.mmread
+    <https://github.com/alugowski/fast_matrix_market/tree/main/python>`_.
+
+    By default, ``fast_matrix_market`` will be used if available, because it
+    is faster. Additional keyword arguments in ``**kwargs`` will be passed
+    to the engine's ``mmread``. For example, ``parallelism=8`` will set the
+    number of threads to use to 8 when using ``fast_matrix_market``.
 
     Parameters
     ----------
-    filename : str or file
+    source : str or file
         Filename (.mtx or .mtz.gz) or file-like object
+    engine : {"auto", "scipy", "fmm", "fast_matrix_market"}, default "auto"
+        How to read the matrix market file. "scipy" uses ``scipy.io.mmread``,
+        "fmm" and "fast_matrix_market" uses ``fast_matrix_market.mmread``,
+        and "auto" will use "fast_matrix_market" if available.
     dup_op : BinaryOp, optional
         Aggregation function for duplicate coordinates (if found)
     name : str, optional
@@ -586,11 +600,26 @@ def mmread(source, *, dup_op=None, name=None):
     :class:`~graphblas.Matrix`
     """
     try:
+        # scipy is currently needed for *all* engines
         from scipy.io import mmread
         from scipy.sparse import isspmatrix_coo
     except ImportError:  # pragma: no cover (import)
         raise ImportError("scipy is required to read Matrix Market files") from None
-    array = mmread(source)
+    engine = engine.lower()
+    if engine in {"auto", "fmm", "fast_matrix_market"}:
+        try:
+            from fast_matrix_market import mmread  # noqa: F811
+        except ImportError:  # pragma: no cover (import)
+            if engine != "auto":
+                raise ImportError(
+                    "fast_matrix_market is required to read Matrix Market files "
+                    f'using the "{engine}" engine'
+                ) from None
+    elif engine != "scipy":
+        raise ValueError(
+            f'Bad engine value: {engine!r}. Must be "auto", "scipy", "fmm", or "fast_matrix_market"'
+        )
+    array = mmread(source, **kwargs)
     if isspmatrix_coo(array):
[…]
