+++ environment.yml
@@ -23,7 +23,7 @@ dependencies:
     - pandas
     # For I/O
     - awkward
-    # - fast_matrix_market  # Coming soon...
+    - fast_matrix_market
     - networkx
     - scipy
     - sparse

[…]
