### 2023-03-22 7a5f724
Fix right panel visibility overlap issue
Also split up API reference into separate pages for better navigation experience
