### 2023-03-22 c6850c0
Fix right panel visibility overlap issue (#413)
Also split up API reference into separate pages for better navigation experience
