+++ docs/api_reference/collections.rst
@@ -1,0 +1,23 @@
+Collections
+-----------
+
+Matrix
+~~~~~~
+
+.. autoclass:: graphblas.Matrix
+    :members:
+    :special-members: __getitem__, __setitem__, __delitem__, __contains__, __iter__
+
+Vector
+~~~~~~
+
+.. autoclass:: graphblas.Vector
+    :members:
+    :special-members: __getitem__, __setitem__, __delitem__, __contains__, __iter__
+
+Scalar
+~~~~~~
+
+.. autoclass:: graphblas.Scalar
+    :members:
+    :special-members: __eq__, __bool__
