### 2023-03-24 5f79544
fast_matrix_market not available in PyPy
