### 2023-03-24 cfd5a1c
give pypy more flexibility
