### 2023-03-24 87948d2
Make sure pypy is and remains installed
