+++ scripts/check_versions.sh
@@ -14,3 +14,4 @@ conda search 'numba[channel=conda-forge]>=0.56.4'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
 conda search 'flake8-bugbear[channel=conda-forge]>=23.3.12'
 conda search 'flake8-simplify[channel=conda-forge]>=0.19.3'
+conda search 'python[channel=conda-forge]>=3.8 *pypy*'
