### 2023-03-25 551963c
Try to pip install fast-matrix-market on PyPy
