### 2023-03-25 2e9d794
Use mamba or conda, b/c it's no fun when one fails to install
