+++ .github/workflows/test_and_build.yml
@@ -91,7 +91,7 @@ jobs:
     env:
       # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
       # Setting this is a workaround.
-      KMP_DUPLICATE_LIB_OK: ${{ contains(matrix.os, 'macXXXos') && 'True' || '' }}
+      KMP_DUPLICATE_LIB_OK: ${{ contains(matrix.os, 'macos') && 'TRUE' || 'FALSE' }}
     steps:
       - name: Checkout
         uses: actions/checkout@v3
