### 2023-03-27 #81 @tomalrussell
Hi all - just checking in with @sneakers-the-rat and @szhorvat - thanks for the engagement, discussion and issues opened so far. Are you both on track to submit your initial reviews by the 29th?

(For context - the pyopensci process aims for 3 weeks for review, 2 weeks for subsequent changes, and 1 week for reviewer approval of changes.)
