### 2023-03-27 python-graphblas/python-graphblas#410 @eriknw
@sneakers-the-rat, please let us know if there is anything we can do (chat, meet, reply to issues, etc.) to help you understand our project and perform the review more effectively.
