### 2023-03-28 #81 @szhorvat
I arrived back on Sunday and immediately came down with a bad flu, so I won't be able to finish by tomorrow. If I recover in time, I still hope to finish by Sunday.
