+++ graphblas/tests/test_dtype.py
@@ -249,11 +249,10 @@ def test_has_complex():
         assert dtypes._supports_complex
         return
 
-    # import suitesparse_graphblas as ssgb
-    # from packaging.version import parse
+    import suitesparse_graphblas as ssgb
+    from packaging.version import parse
 
-    # if parse(ssgb.__version__) < parse("7.4.3.1"):
-    #     assert not dtypes._supports_complex
-    # else:
-    #     assert dtypes._supports_complex
-    assert dtypes._supports_complex  # XXX
+    if parse(ssgb.__version__) < parse("7.4.3.1"):
+        assert not dtypes._supports_complex
+    else:
+        assert dtypes._supports_complex
