### 2023-03-28 fea1d4c
equal weights again
