### 2023-03-28 ff1fcb5
Fix infix expression `_value` and `_expr` usage
Previously, `infixexpr._value` was sometimes `MatrixExpression` and sometimes `Matrix`. Scary!
