### 2023-03-28 41fb71a
Fix infix expression `_value` and `_expr` usage (#418)
Previously, `infixexpr._value` was sometimes `MatrixExpression` and sometimes `Matrix`. Scary!
