+++ pyproject.toml
@@ -318,7 +318,8 @@ ignore = [
 "graphblas/core/ss/matrix.py" = ["NPY002"]  # numba doesn't support rng generator yet
 "graphblas/core/ss/vector.py" = ["NPY002"]  # numba doesn't support rng generator yet
 "graphblas/ss/_core.py" = ["N999"]  # We want _core.py to be underscopre
-"graphblas/tests/*py" = ["S101", "T201", "D103", "D100", "SIM300"]  # Allow assert, print, no docstring, and yoda
+# Allow assert, pickle, RNG, print, no docstring, and yoda in tests
+"graphblas/tests/*py" = ["S101", "S301", "S311", "T201", "D103", "D100", "SIM300"]
 "graphblas/tests/test_formatting.py" = ["E501"]  # Allow long lines
 "graphblas/**/__init__.py" = ["F401"]  # Allow unused imports (w/o defining `__all__`)
 "scripts/*.py" = ["INP001"]  # Not a package
