### 2023-03-29 3128130
Split operator.py into multiple files
