+++ graphblas/core/operator.py
@@ -1,3585 +1,0 @@
-import inspect
-import itertools
-import re
-from collections.abc import Mapping
-from functools import lru_cache, reduce
-from operator import getitem, mul
-from types import BuiltinFunctionType, FunctionType, ModuleType
-
-import numba
-import numpy as np
-
-from .. import (
-    _STANDARD_OPERATOR_NAMES,
-    backend,
-    binary,
-    config,
-    indexunary,
-    monoid,
-    op,
-    select,
-    semiring,
-    unary,
-)
-from ..dtypes import (
-    BOOL,
-    FP32,
-    FP64,
-    INT8,
-    INT16,
-    INT32,
-    INT64,
-    UINT8,
-    UINT16,
-    UINT32,
-    UINT64,
-    _sample_values,
-    _supports_complex,
-    lookup_dtype,
-    unify,
-)
-from ..exceptions import UdfParseError, check_status_carg
-from . import ffi, lib
-from .expr import InfixExprBase
-from .utils import libget, output_type
-
-if _supports_complex:
-    from ..dtypes import FC32, FC64
-
-ffi_new = ffi.new
-UNKNOWN_OPCLASS = "UnknownOpClass"
-
-# These now live as e.g. `gb.unary.ss.positioni`
-# Deprecations such as `gb.unary.positioni` will be removed in 2023.9.0 or later.
-_SS_OPERATORS = {
-    # unary
-    "erf",  # scipy.special.erf
-    "erfc",  # scipy.special.erfc
-    "frexpe",  # np.frexp[1]
-    "frexpx",  # np.frexp[0]
-    "lgamma",  # scipy.special.loggamma
-    "tgamma",  # scipy.special.gamma
-    # Positional
-    # unary
-    "positioni",
-    "positioni1",
-    "positionj",
-    "positionj1",
-    # binary
-    "firsti",
-    "firsti1",
-    "firstj",
-    "firstj1",
-    "secondi",
-    "secondi1",
-    "secondj",
-    "secondj1",
-    # semiring
-    "any_firsti",
-    "any_firsti1",
-    "any_firstj",
-    "any_firstj1",
-    "any_secondi",
-    "any_secondi1",
-    "any_secondj",
-    "any_secondj1",
-    "max_firsti",
-    "max_firsti1",
-    "max_firstj",
-    "max_firstj1",
-    "max_secondi",
-    "max_secondi1",
-    "max_secondj",
-    "max_secondj1",
-    "min_firsti",
-    "min_firsti1",
-    "min_firstj",
-    "min_firstj1",
-    "min_secondi",
-    "min_secondi1",
-    "min_secondj",
-    "min_secondj1",
-    "plus_firsti",
-    "plus_firsti1",
-    "plus_firstj",
-    "plus_firstj1",
-    "plus_secondi",
-    "plus_secondi1",
-    "plus_secondj",
-    "plus_secondj1",
-    "times_firsti",
-    "times_firsti1",
-    "times_firstj",
-    "times_firstj1",
-    "times_secondi",
-    "times_secondi1",
-    "times_secondj",
-    "times_secondj1",
-}
-
-
-def _hasop(module, name):
-    return (
-        name in module.__dict__
-        or name in module._delayed
-        or name in getattr(module, "_deprecated", ())
-    )
-
-
-class OpPath:
-    def __init__(self, parent, name):
-        self._parent = parent
-        self._name = name
-        self._delayed = {}
-        self._delayed_commutes_to = {}
-
-    def __getattr__(self, key):
-        if key in self._delayed:
-            func, kwargs = self._delayed.pop(key)
-            return func(**kwargs)
-        self.__getattribute__(key)  # raises
-
-
[…]
