+++ graphblas/core/agg.py
@@ -1,680 +1,17 @@
-from functools import partial
-from operator import getitem
+"""graphblas.core.agg namespace is deprecated; please use graphblas.core.operator.agg instead.
 
-import numpy as np
+.. deprecated:: 2023.3.0
+`graphblas.core.agg` will be removed in a future release.
+Use `graphblas.core.operator.agg` instead.
+Will be removed in version 2023.11.0 or later.
 
-from .. import agg, backend, binary, monoid, semiring, unary
-from ..dtypes import INT64, lookup_dtype
-from .utils import output_type
+"""
+import warnings
 
+from .operator.agg import *
 
-def _get_types(ops, initdtype):
-    """Determine the input and output types of an aggregator based on a list of ops."""
-    if initdtype is None:
-        prev = dict(ops[0].types)
-    else:
-        op = ops[0]
-        prev = {key: get_typed_op(op, key, initdtype).return_type for key in op.types}
-    for op in ops[1:]:
-        cur = {}
-        types = op.types
-        for in_type, out_type in prev.items():
-            if out_type not in types:  # pragma: no cover (safety)
-                continue
-            cur[in_type] = types[out_type]
-        prev = cur
-    return prev
-
-
-class Aggregator:
-    opclass = "Aggregator"
-
-    def __init__(
-        self,
-        name,
-        *,
-        initval=None,
-        monoid=None,
-        semiring=None,
-        switch=False,
-        semiring2=None,
-        finalize=None,
-        composite=None,
-        custom=None,
-        types=None,
-        any_dtype=None,
-    ):
-        self.name = name
-        self._initval_orig = initval
-        self._initval = False if initval is None else initval
-        self._initdtype = lookup_dtype(type(self._initval), self._initval)
-        self._monoid = monoid
-        self._semiring = semiring
-        self._semiring2 = semiring2
-        self._switch = switch
-        self._finalize = finalize
-        self._composite = composite
-        self._custom = custom
-        if types is None:
-            if monoid is not None:
-                types = [monoid]
-            elif semiring is not None:
-                types = [semiring, semiring2]
-                if finalize is not None:
-                    types.append(finalize)
-                initval = self._initval
-            else:  # pragma: no cover (sanity)
-                raise TypeError("types must be provided for composite and custom aggregators")
-        self._types_orig = types
-        self._types = None
-        self._typed_ops = {}
-        self._any_dtype = any_dtype
-
-    @property
-    def types(self):
-        if self._types is None:
-            if type(self._semiring) is str:
-                self._semiring = semiring.from_string(self._semiring)
-                if type(self._types_orig[0]) is str:  # pragma: no branch
-                    self._types_orig[0] = semiring.from_string(self._types_orig[0])
-            self._types = _get_types(
-                self._types_orig, None if self._initval_orig is None else self._initdtype
-            )
-        return self._types
-
-    def __getitem__(self, dtype):
-        dtype = lookup_dtype(dtype)
-        if not self._any_dtype and dtype not in self.types:
-            raise KeyError(f"{self.name} does not work with {dtype}")
-        if dtype not in self._typed_ops:
-            self._typed_ops[dtype] = TypedAggregator(self, dtype)
-        return self._typed_ops[dtype]
-
[…]
