### 2023-03-29 d5a4456
Try to fix test_packages on Windows
