### 2023-03-30 2b566f3
Use `--only-binary ":all:"` when installing wheels
