+++ .github/workflows/test_and_build.yml
@@ -203,7 +203,8 @@ jobs:
           elif [[ ${{ steps.sourcetype.outputs.selected}} == "wheel" ]] ; then
             psgver=$(python -c 'import random ; print(random.choice(["==7.4.3.2", ""]))')
           elif [[ ${{ steps.sourcetype.outputs.selected}} == "source" ]] ; then
-            psgver=$(python -c 'import random ; print(random.choice(["==7.4.0", "==7.4.1", "==7.4.2", "==7.4.3.0", "==7.4.3.1", "==7.4.3.2", ""]))')
+            psgver=$(python -c 'import random ; print(random.choice(["==7.4.3.2", ""]))')
+            # psgver=$(python -c 'import random ; print(random.choice(["==7.4.0", "==7.4.1", "==7.4.2", "==7.4.3.0", "==7.4.3.1", "==7.4.3.2", ""]))')
           else
             psgver=""
           fi
@@ -222,20 +223,24 @@ jobs:
             ${{ steps.sourcetype.outputs.selected == 'upstream' && 'cython' || '' }} \
             ${{ steps.sourcetype.outputs.selected != 'wheel' && '"graphblas>=7.4.0"' || '' }} \
             ${{ contains(steps.pyver.outputs.selected, 'pypy') && 'pypy' || '' }}
-            # ${{ steps.sourcetype.outputs.selected == 'source' && 'setuptools setuptools-git-versioning wheel cython' || '' }} \
       - name: Build extension module
         run: |
           if [[ ${{ steps.sourcetype.outputs.selected }} == "wheel" ]]; then
-              # pip install --no-deps --only-binary ":all:" suitesparse-graphblas${psgver}
-              pip install --no-deps --pre --only-binary ":all:" suitesparse-graphblas==7.4.3.2a1
+              # Add --pre if installing a pre-release
+              pip install --no-deps --only-binary ":all:" suitesparse-graphblas${psgver}
+
+              # Add the below line to the conda install command above if installing from test.pypi.org
+              # ${{ steps.sourcetype.outputs.selected == 'wheel' && 'setuptools setuptools-git-versioning wheel cython' || '' }} \
               # pip install --no-deps --only-binary ":all:" --index-url https://test.pypi.org/simple/ "suitesparse-graphblas>=7.4.3"
           elif [[ ${{ steps.sourcetype.outputs.selected }} == "source" ]]; then
-              # pip install --no-deps --no-binary=all suitesparse-graphblas${psgver}
-              pip install --no-deps --pre --no-binary=all suitesparse-graphblas==7.4.3.2a1
+              # Add --pre if installing a pre-release
+              pip install --no-deps --no-binary=all suitesparse-graphblas${psgver}
+
+              # Add the below line to the conda install command above if installing from test.pypi.org
+              # ${{ steps.sourcetype.outputs.selected == 'source' && 'setuptools setuptools-git-versioning wheel cython' || '' }} \
               # pip install --no-deps --no-build-isolation --no-binary=all --index-url https://test.pypi.org/simple/ suitesparse-graphblas==7.4.3.3
           elif [[ ${{ steps.sourcetype.outputs.selected }} == "upstream" ]]; then
               pip install --no-deps git+https://github.com/GraphBLAS/python-suitesparse-graphblas.git@main#egg=suitesparse-graphblas
-              # pip install --no-deps git+https://github.com/jim22k/python-suitesparse-graphblas.git@dd89007ca750a4d2da352ddff3c12437b86d5946#egg=suitesparse-graphblas
           fi
           pip install --no-deps -e .
       - name: Unit tests
