### 2023-03-31 7030218
Make source weights for psg equal
