### 2023-03-31 40e548d
Split operator.py into multiple files (#420)
* Split operator.py into multiple files

* Ensure tomli is installed when testing

* ruff 260
