+++ graphblas/core/operator/binary.py
@@ -1,0 +1,864 @@
+import inspect
+import re
+from functools import lru_cache
+from types import FunctionType
+
+import numba
+import numpy as np
+
+from ... import _STANDARD_OPERATOR_NAMES, backend, binary, monoid, op
+from ...dtypes import (
+    BOOL,
+    FP32,
+    FP64,
+    INT8,
+    INT16,
+    INT32,
+    INT64,
+    UINT8,
+    UINT16,
+    UINT32,
+    UINT64,
+    _sample_values,
+    _supports_complex,
+    lookup_dtype,
+)
+from ...exceptions import UdfParseError, check_status_carg
+from .. import ffi, lib
+from ..expr import InfixExprBase
+from .base import (
+    _SS_OPERATORS,
+    OpBase,
+    ParameterizedUdf,
+    TypedOpBase,
+    _call_op,
+    _deserialize_parameterized,
+    _get_udt_wrapper,
+    _hasop,
+    _udt_mask,
+)
+
+if _supports_complex:
+    from ...dtypes import FC32, FC64
+
+ffi_new = ffi.new
+
+
+class TypedBuiltinBinaryOp(TypedOpBase):
+    __slots__ = ()
+    opclass = "BinaryOp"
+
+    def __call__(self, left, right=None, *, left_default=None, right_default=None):
+        if left_default is not None or right_default is not None:
+            if (
+                left_default is None
+                or right_default is None
+                or right is not None
+                or not isinstance(left, InfixExprBase)
+                or left.method_name != "ewise_add"
+            ):
+                raise TypeError(
+                    "Specifying `left_default` or `right_default` keyword arguments implies "
+                    "performing `ewise_union` operation with infix notation.\n"
+                    "There is only one valid way to do this:\n\n"
+                    f">>> {self}(x | y, left_default=0, right_default=0)\n\nwhere x and y "
+                    "are Vectors or Matrices, and left_default and right_default are scalars."
+                )
+            return left.left.ewise_union(left.right, self, left_default, right_default)
+        return _call_op(self, left, right)
+
+    @property
+    def monoid(self):
+        rv = getattr(monoid, self.name, None)
+        if rv is not None and self.type in rv._typed_ops:
+            return rv[self.type]
+
+    @property
+    def commutes_to(self):
+        commutes_to = self.parent.commutes_to
+        if commutes_to is not None and (self.type in commutes_to._typed_ops or self.type._is_udt):
+            return commutes_to[self.type]
+
+    @property
+    def _semiring_commutes_to(self):
+        commutes_to = self.parent._semiring_commutes_to
+        if commutes_to is not None and (self.type in commutes_to._typed_ops or self.type._is_udt):
+            return commutes_to[self.type]
+
+    @property
+    def is_commutative(self):
+        return self.commutes_to is self
+
+    @property
+    def type2(self):
+        return self.type if self._type2 is None else self._type2
+
+
+class TypedUserBinaryOp(TypedOpBase):
+    __slots__ = "_monoid"
+    opclass = "BinaryOp"
+
+    def __init__(self, parent, name, type_, return_type, gb_obj, dtype2=None):
+        super().__init__(parent, name, type_, return_type, gb_obj, f"{name}_{type_}", dtype2=dtype2)
+        self._monoid = None
+
+    @property
+    def monoid(self):
+        if self._monoid is None:
+            monoid = self.parent.monoid
+            if monoid is not None and self.type in monoid:
+                self._monoid = monoid[self.type]
[…]
