+++ scripts/check_versions.sh
@@ -13,5 +13,5 @@ conda search 'fast_matrix_market[channel=conda-forge]>=1.4.5'
 conda search 'numba[channel=conda-forge]>=0.56.4'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
 conda search 'flake8-bugbear[channel=conda-forge]>=23.3.23'
-conda search 'flake8-simplify[channel=conda-forge]>=0.19.3'
+conda search 'flake8-simplify[channel=conda-forge]>=0.20.0'
 # conda search 'python[channel=conda-forge]>=3.8 *pypy*'
