### 2023-03-31 python-graphblas/python-graphblas#403 @eriknw
Closing as completed.

`python-suitesparse-graphblas` now has wheels for linux, windows, and macos :tada: (there are a few more wheels we'd like to add)

Issue for instructions of how to install with pre-installed version of SuiteSparse:GraphBLAS is here: https://github.com/GraphBLAS/python-suitesparse-graphblas/issues/79
