### 2023-03-31 abb1c78
Test new wheels for `python-suitesparse-graphblas` on all OSes (#385)
* Test new wheels for `python-suitesparse-graphblas` on all OSes

* Use mamba or conda, b/c it's no fun when one fails to install

* Try coveralls upload twice (it sometimes fails)
