+++ .github/workflows/test_and_build.yml
@@ -88,6 +88,10 @@ jobs:
       matrix:
         os: ["ubuntu-latest", "macos-latest", "windows-latest"]
         slowtask: ["pytest_normal", "pytest_bizarro", "notebooks"]
+    env:
+      # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
+      # Setting this is a workaround.
+      KMP_DUPLICATE_LIB_OK: ${{ contains(matrix.os, 'macos') && 'TRUE' || 'FALSE' }}
     steps:
       - name: Checkout
         uses: actions/checkout@v3
@@ -98,6 +102,9 @@ jobs:
         id: pyver
         with:
           # We should support major Python versions for at least 36-42 months
+          # We could probably support pypy if numba were optional
+          # 3.8.16 0_73_pypy
+          # 3.9.16 0_73_pypy
           contents: |
             3.8
             3.9
@@ -110,24 +117,38 @@ jobs:
         uses: ddradar/choose-random-action@v2.0.2
         id: sourcetype
         with:
-          # Set weight to 0 to skip (such as if 'upstream' is known to not work).
-          # Have slightly higher weight for `conda-forge` for faster CI.
+          # Weights must be natural numbers, so set weights to very large to skip one
+          # (such as if 'upstream' is known to not work).
           contents: |
             conda-forge
             wheel
             source
             upstream
           weights: |
-            2
             1
             1
             1
-      - name: Setup conda
+            1
+      - name: Setup mamba
         uses: conda-incubator/setup-miniconda@v2
+        id: setup_mamba
+        continue-on-error: true
         with:
           miniforge-variant: Mambaforge
           miniforge-version: latest
           use-mamba: true
+          python-version: ${{ steps.pyver.outputs.selected }}
+          channels: conda-forge,nodefaults
+          channel-priority: strict
+          activate-environment: graphblas
+          auto-activate-base: false
+      - name: Setup conda
+        uses: conda-incubator/setup-miniconda@v2
+        id: setup_conda
+        if: steps.setup_mamba.outcome == 'failure'
+        continue-on-error: false
+        with:
+          auto-update-conda: true
           python-version: ${{ steps.pyver.outputs.selected }}
           channels: conda-forge,nodefaults
           channel-priority: strict
@@ -144,17 +165,17 @@ jobs:
           yamlver=$(python -c 'import random ; print(random.choice(["=5.4", "=6.0", ""]))')
           sparsever=$(python -c 'import random ; print(random.choice(["=0.12", "=0.13", "=0.14", ""]))')
           fmmver=$(python -c 'import random ; print(random.choice(["=1.4", ""]))')
-          if [[ ${{ steps.pyver.outputs.selected }} == "3.8" ]]; then
+          if [[ ${{ startsWith(steps.pyver.outputs.selected, '3.8') }} == true ]]; then
             npver=$(python -c 'import random ; print(random.choice(["=1.21", "=1.22", "=1.23", ""]))')
             spver=$(python -c 'import random ; print(random.choice(["=1.8", "=1.9", "=1.10", ""]))')
             pdver=$(python -c 'import random ; print(random.choice(["=1.2", "=1.3", "=1.4", "=1.5", ""]))')
             akver=$(python -c 'import random ; print(random.choice(["=1.9", "=1.10", "=2.0", "=2.1", ""]))')
-          elif [[ ${{ steps.pyver.outputs.selected }} == "3.9" ]]; then
[…]
