### 2023-04-03 8243013
WIP: making numba optional
