### 2023-04-04 a987553
No numba on Python 3.11; use conda
