+++ .github/workflows/test_and_build.yml
@@ -133,23 +133,23 @@ jobs:
             1
             1
             1
-      - name: Setup mamba
-        uses: conda-incubator/setup-miniconda@v2
-        id: setup_mamba
-        continue-on-error: true
-        with:
-          miniforge-variant: Mambaforge
-          miniforge-version: latest
-          use-mamba: true
-          python-version: ${{ steps.pyver.outputs.selected }}
-          channels: conda-forge,${{ contains(steps.pyver.outputs.selected, 'pypy') && 'defaults' || 'nodefaults' }}
-          channel-priority: ${{ contains(steps.pyver.outputs.selected, 'pypy') && 'flexible' || 'strict' }}
-          activate-environment: graphblas
-          auto-activate-base: false
+#      - name: Setup mamba
+#        uses: conda-incubator/setup-miniconda@v2
+#        id: setup_mamba
+#        continue-on-error: true
+#        with:
+#          miniforge-variant: Mambaforge
+#          miniforge-version: latest
+#          use-mamba: true
+#          python-version: ${{ steps.pyver.outputs.selected }}
+#          channels: conda-forge,${{ contains(steps.pyver.outputs.selected, 'pypy') && 'defaults' || 'nodefaults' }}
+#          channel-priority: ${{ contains(steps.pyver.outputs.selected, 'pypy') && 'flexible' || 'strict' }}
+#          activate-environment: graphblas
+#          auto-activate-base: false
       - name: Setup conda
         uses: conda-incubator/setup-miniconda@v2
         id: setup_conda
-        if: steps.setup_mamba.outcome == 'failure'
+#        if: steps.setup_mamba.outcome == 'failure'
         continue-on-error: false
         with:
           auto-update-conda: true
@@ -217,8 +217,8 @@ jobs:
           else
             numbaver=$(python -c 'import random ; print(random.choice(["=0.56", ""]))')
           fi
-          if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') }} == true || $(python -c 'import random ; print(random.random() < .1)') == True ]]; then
-            # Some packages aren't available for pypy (and 10% chance to not install these otherwise)
+          if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') }} == true || ${{ startsWith(steps.pyver.outputs.selected, '3.11') }} == true || $(python -c 'import random ; print(random.random() < .1)') == True ]]; then
+            # Some packages aren't available for pypy or Python 3.11 (and 10% chance to not install these otherwise)
             numba=""
             sparse=""
             awkward=""
