### 2023-04-04 d31b3de
skip notebooks and try mamba
