+++ scripts/check_versions.sh
@@ -4,7 +4,7 @@
 # This may be helpful when updating dependency versions in CI.
 # Tip: add `--json` for more information.
 conda search 'numpy[channel=conda-forge]>=1.24.2'
-conda search 'pandas[channel=conda-forge]>=1.5.3'
+conda search 'pandas[channel=conda-forge]>=2.0.0'
 conda search 'scipy[channel=conda-forge]>=1.10.1'
 conda search 'networkx[channel=conda-forge]>=3.0'
 conda search 'awkward[channel=conda-forge]>=2.1.1'
