### 2023-04-04 61f5660
Don't test against pypy in CI
