### 2023-04-04 71af5e7
Update classifiers (we don't officially support PyPy)
