+++ pyproject.toml
@@ -90,51 +90,37 @@ scipy = [
     "scipy >=1.8",
 ]
 suitesparse-udf = [
-    "python-graphblas[suitesparse]",
-    "python-graphblas[numba]",
+    "python-graphblas[suitesparse,numba]",
 ]
 repr = [
     "python-graphblas[pandas]",
 ]
 io = [
-    "python-graphblas[networkx]",
-    "python-graphblas[numba]",
-    "python-graphblas[scipy]",
+    "python-graphblas[networkx,numba,scipy]",
     "awkward >=1.9",
     "fast-matrix-market >=1.4.5",
     "sparse >=0.12",
 ]
 viz = [
-    "python-graphblas[networkx]",
-    "python-graphblas[scipy]",
+    "python-graphblas[networkx,scipy]",
     "matplotlib >=3.5",
 ]
 datashade = [
-    "python-graphblas[numba]",
-    "python-graphblas[pandas]",
-    "python-graphblas[scipy]",
+    "python-graphblas[numba,pandas,scipy]",
     "datashader >=0.12",
     "hvplot >=0.7",
 ]
 test = [
-    "python-graphblas[suitesparse]",
-    "python-graphblas[pandas]",
-    "python-graphblas[scipy]",
+    "python-graphblas[suitesparse,pandas,scipy]",
     "packaging >=21",
     "pytest >=6.2",
     "tomli >=1",
 ]
 default = [
-    "python-graphblas[suitesparse-udf]",
-    "python-graphblas[pandas]",
-    "python-graphblas[scipy]",
+    "python-graphblas[suitesparse-udf,pandas,scipy]",
 ]
 complete = [
-    "python-graphblas[default]",
-    "python-graphblas[io]",
-    "python-graphblas[viz]",
-    "python-graphblas[test]",
-    "python-graphblas[datashade]",
+    "python-graphblas[default,io,viz,test,datashade]",
 ]
 
 [tool.setuptools]
