### 2023-04-07 3af869c
Make suitesparse_graphblas optional (in the future); don't install numba on 3.11 yet
