+++ .github/workflows/imports.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     # strategy:
     #   matrix:
-    #     python-version: ["3.8", "3.9", "3.10"]
+    #     python-version: ["3.8", "3.9", "3.10", "3.11"]
     steps:
       - uses: actions/checkout@v3
       - name: RNG for Python version
@@ -22,14 +22,16 @@ jobs:
             3.8
             3.9
             3.10
+            3.11
           weights: |
             1
             1
             1
+            1000
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ steps.pyver.outputs.selected }}
           # python-version: ${{ matrix.python-version }}
       - run: python -m pip install --upgrade pip
-      - run: pip install -e .
+      - run: pip install -e .[default]
       - run: ./scripts/test_imports.sh
