### 2023-04-07 ce0a667
Ignore deprecation warning raised by awkward in Python 3.11
