+++ pyproject.toml
@@ -189,6 +189,9 @@ filterwarnings = [
     # https://setuptools.pypa.io/en/latest/history.html#v67-5-0
     "ignore:pkg_resources is deprecated as an API:DeprecationWarning:pkg_resources",
 
+    # sre_parse deprecated in 3.11; this is triggered by awkward 0.10
+    "ignore:module 'sre_parse' is deprecated:DeprecationWarning:",
+
     # pypy gives this warning
     "ignore:can't resolve package from __spec__ or __package__:ImportWarning:",
 ]
