### 2023-04-07 0f66e65
sre_constants also deprecated
