+++ graphblas/tests/test_op.py
@@ -35,6 +35,8 @@ from graphblas.dtypes import (
 )
 from graphblas.exceptions import DomainMismatch, UdfParseError
 
+from .conftest import shouldhave
+
 if dtypes._supports_complex:
     from graphblas.dtypes import FC32, FC64
 
@@ -630,7 +632,7 @@ def test_binary_updates():
     vec4 = Vector.from_coo([0], [-3], dtype=dtypes.INT64)
     result2 = vec4.ewise_mult(vec2, binary.cdiv).new()
     assert result2.isequal(Vector.from_coo([0], [-1], dtype=dtypes.INT64), check_dtype=True)
-    if supports_udfs or hasattr(binary, "floordiv"):
+    if shouldhave(binary, "floordiv"):
         result3 = vec4.ewise_mult(vec2, binary.floordiv).new()
         assert result3.isequal(Vector.from_coo([0], [-2], dtype=dtypes.INT64), check_dtype=True)
 
@@ -695,13 +697,13 @@ def test_op_namespace():
     assert op.plus is binary.plus
     assert op.plus_times is semiring.plus_times
 
-    if supports_udfs or hasattr(unary.numpy, "fabs"):
+    if shouldhave(unary.numpy, "fabs"):
         assert op.numpy.fabs is unary.numpy.fabs
-    if supports_udfs or hasattr(binary.numpy, "subtract"):
+    if shouldhave(binary.numpy, "subtract"):
         assert op.numpy.subtract is binary.numpy.subtract
-    if supports_udfs or hasattr(binary.numpy, "add"):
+    if shouldhave(binary.numpy, "add"):
         assert op.numpy.add is binary.numpy.add
-    if supports_udfs or hasattr(semiring.numpy, "add_add"):
+    if shouldhave(semiring.numpy, "add_add"):
         assert op.numpy.add_add is semiring.numpy.add_add
     assert len(dir(op)) > 300
     if supports_udfs:
@@ -769,10 +771,10 @@ def test_op_namespace():
 @pytest.mark.slow
 def test_binaryop_attributes_numpy():
     # Some coverage from this test depends on order of tests
-    if supports_udfs or hasattr(monoid.numpy, "add"):
+    if shouldhave(monoid.numpy, "add"):
         assert binary.numpy.add[int].monoid is monoid.numpy.add[int]
         assert binary.numpy.add.monoid is monoid.numpy.add
-    if supports_udfs or hasattr(binary.numpy, "subtract"):
+    if shouldhave(binary.numpy, "subtract"):
         assert binary.numpy.subtract[int].monoid is None
         assert binary.numpy.subtract.monoid is None
 
@@ -794,12 +796,12 @@ def test_binaryop_attributes():
         assert op[int].parent is op
 
     assert binary.plus[int].parent is binary.plus
-    if supports_udfs or hasattr(binary.numpy, "add"):
+    if shouldhave(binary.numpy, "add"):
         assert binary.numpy.add[int].parent is binary.numpy.add
 
     # bad type
     assert binary.plus[bool].monoid is None
-    if supports_udfs or hasattr(binary.numpy, "equal"):
+    if shouldhave(binary.numpy, "equal"):
         assert binary.numpy.equal[int].monoid is None
         assert binary.numpy.equal[bool].monoid is monoid.numpy.equal[bool]  # sanity
 
@@ -824,7 +826,7 @@ def test_monoid_attributes():
     assert monoid.plus.binaryop is binary.plus
     assert monoid.plus.identities == {typ: 0 for typ in monoid.plus.types}
 
-    if supports_udfs or hasattr(monoid.numpy, "add"):
+    if shouldhave(monoid.numpy, "add"):
         assert monoid.numpy.add[int].binaryop is binary.numpy.add[int]
[…]
