+++ .github/workflows/imports.yml
@@ -7,11 +7,30 @@ on:
       - main
 
 jobs:
-  test_imports:
+  rngs:
     runs-on: ubuntu-latest
+    outputs:
+      os: ${{ steps.os.outputs.selected }}
+    steps:
+      - name: RNG for os
+        uses: ddradar/choose-random-action@v2.0.2
+        id: os
+        with:
+          contents: |
+            ubuntu-latest
+            macos-latest
+            windows-latest
+          weights: |
+            1
+            1
+            1
+  test_imports:
+    needs: rngs
+    runs-on: ${{ needs.rngs.outputs.os }}
     # strategy:
     #   matrix:
     #     python-version: ["3.8", "3.9", "3.10", "3.11"]
+    #     os: ["ubuntu-latest", "macos-latest", "windows-latest"]
     steps:
       - uses: actions/checkout@v3
       - name: RNG for Python version

[…]
