### 2023-04-09 e1ed1f0
Recipes for L1norm and Linfnorm aggregators
