### 2023-04-09 a074aa0
Better error message for default operators that are UDFs
