+++ .github/workflows/imports.yml
@@ -11,6 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     outputs:
       os: ${{ steps.os.outputs.selected }}
+      pyver: ${{ steps.pyver.outputs.selected }}
     steps:
       - name: RNG for os
         uses: ddradar/choose-random-action@v2.0.2
@@ -24,15 +25,6 @@ jobs:
             1
             1
             1
-  test_imports:
-    needs: rngs
-    runs-on: ${{ needs.rngs.outputs.os }}
-    # strategy:
-    #   matrix:
-    #     python-version: ["3.8", "3.9", "3.10", "3.11"]
-    #     os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-    steps:
-      - uses: actions/checkout@v3
       - name: RNG for Python version
         uses: ddradar/choose-random-action@v2.0.2
         id: pyver
@@ -47,9 +39,19 @@ jobs:
             1
             1
             1
+  test_imports:
+    needs: rngs
+    runs-on: ${{ needs.rngs.outputs.os }}
+    # runs-on: ${{ matrix.os }}
+    # strategy:
+    #   matrix:
+    #     python-version: ["3.8", "3.9", "3.10", "3.11"]
+    #     os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+    steps:
+      - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
-          python-version: ${{ steps.pyver.outputs.selected }}
+          python-version: ${{ needs.rngs.outputs.pyver }}
           # python-version: ${{ matrix.python-version }}
       - run: python -m pip install --upgrade pip
       - run: pip install -e .[default]
