+++ graphblas/tests/test_vector.py
@@ -1,7 +1,6 @@
 import inspect
 import itertools
 import pickle
-import platform
 import sys
 import types
 import weakref
@@ -23,13 +22,12 @@ from graphblas.exceptions import (
     OutputNotEmpty,
 )
 
-from .conftest import autocompute, compute
+from .conftest import autocompute, compute, pypy
 
 from graphblas import Matrix, Scalar, Vector  # isort:skip (for dask-graphblas)
 
 
 suitesparse = backend == "suitesparse"
-pypy = platform.python_implementation() == "PyPy"
 
 
 @pytest.fixture
