### 2023-04-10 python-graphblas/python-graphblas#428 @sneakers-the-rat
[pyos][docs] - Intra-docs references
Making a single issue to collect any times the docs refer to some other part of the docs but that reference is unclear. will edit this top issue with any additional probs

- [ ] https://python-graphblas.readthedocs.io/en/stable/user_guide/collections.html#opaque-objects says to look to the API docs for documentation of the `ss` namespace but it's unclear to me where that documentation is

Part of: https://github.com/pyOpenSci/software-submission/issues/81
