### 2023-04-10 python-graphblas/python-graphblas#427 @eriknw
Interesting perspective. I'm curious to hear what other people think. I think I'll need more time to digest this.

The modules in `graphblas.core` are useful for implementing `python-graphblas` itself. Although not technically hidden/private (so we still need to have deprecations cycles such as [here](https://github.com/python-graphblas/python-graphblas/blob/f023d9ca8b382e1900ccbdb9ff2f35b61809a441/graphblas/core/agg.py)), they are tucked away under `.core`, which is common practice for scientific Python packages. See [`numpy.core`](https://github.com/numpy/numpy/tree/main/numpy/core), [`pandas.core`](https://github.com/pandas-dev/pandas/tree/main/pandas/core), [`xarray.core`](https://github.com/pydata/xarray/tree/main/xarray/core), [`numba.core`](https://github.com/numba/numba/tree/main/numba/core), [`sympy.core`](https://github.com/sympy/sympy/tree/master/sympy/core), and [`bokeh.core`](https://github.com/bokeh/bokeh/tree/branch-3.2/src/bokeh/core) just to name a few.

In other words, users shouldn't need to use or understand `graphblas.core` under normal circumstances. We've thought carefully about how to structure the package and what to expose to the user. We want the non-core modules (and objects) to be easily explorable via tab-completion such that `gb.<tab>`, `gb.unary.<tab>`, `A.<tab>`, etc. show only what is relevant and are not messy (btw, don't use ipython 8.12.0, which has a bug that ruins tab-completion). After all, I believe the best way to use (and learn) most packages is to play with them in a REPL.

As you've noted, there is a gap between between what is user-facing and understanding _how_ things are implemented. I expect experimenting with objects would typically be more edifying than diving into the code for most things for most people, but we should still strive for making the code more comprehendible.
