+++ .github/workflows/test_and_build.yml
@@ -210,16 +210,21 @@ jobs:
           else
             psgver=""
           fi
+          # TODO: drop 0.57.0rc1 and use 0.57 once numba 0.57 is properly released
           if [[ $npver == "=1.21" ]] ; then
-            numbaver=$(python -c 'import random ; print(random.choice(["=0.55", "=0.56", "=0.57", ""]))')
+            numbaver=$(python -c 'import random ; print(random.choice(["=0.55", "=0.56", "=0.57.0rc1", ""]))')
+          elif [[ $npver == "=1.24" ]] ; then
+            numbaver=$(python -c 'import random ; print(random.choice(["=0.57.0rc1", ""]))')
           else
-            numbaver=$(python -c 'import random ; print(random.choice(["=0.56", "=0.57", ""]))')
+            numbaver=$(python -c 'import random ; print(random.choice(["=0.56", "=0.57.0rc1", ""]))')
           fi
-          numbaver="=0.57.0rc1"  # XXX: testing numba
           fmm=fast_matrix_market${fmmver}
           awkward=awkward${akver}
-          if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') || startsWith(steps.pyver.outputs.selected, '3.12') }} == true || $(python -c 'import random ; print(random.random() < .40)') == True ]]; then
-            # Some packages aren't available for pypy or Python 3.12; 40% otherwise
+          if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') || \
+                startsWith(steps.pyver.outputs.selected, '3.12') }} == true || \
+                ( $(python -c 'import random ; print(random.random() < .50)') == True && ${{ matrix.slowtask != 'notebooks'}} == true ) ]]
+          then
+            # Some packages aren't available for pypy or Python 3.12; 50% otherwise (if not running notebooks)
             numba=""
             numbaver=NA
             sparse=""
@@ -241,7 +246,7 @@ jobs:
           fi
           echo "versions: np${npver} sp${spver} pd${pdver} ak${akver} nx${nxver} numba${numbaver} yaml${yamlver} sparse${sparsever} psgver${psgver}"
 
-          # XXX: added `-c numba`
+          # TODO: remove `-c numba` when numba 0.57 is properly released
           $(command -v mamba || command -v conda) install -c numba packaging pytest coverage coveralls=3.3.1 pytest-randomly cffi donfig tomli \
             pyyaml${yamlver} ${sparse} pandas${pdver} scipy${spver} numpy${npver} ${awkward} \
             networkx${nxver} ${numba} ${fmm} ${psg} \
