+++ .github/workflows/test_and_build.yml
@@ -86,7 +86,8 @@ jobs:
       # This should ensure we'll have full code coverage (i.e., no chance of getting unlucky),
       # since we need to run all slow tests on Windows and non-Windoes OSes.
       matrix:
-        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        # os: ["ubuntu-latest", "macos-latest", "windows-latest"]
+        os: ["windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest"]  # XXX
         slowtask: ["pytest_normal", "pytest_bizarro", "notebooks"]
     env:
       # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
@@ -218,11 +219,13 @@ jobs:
           else
             numbaver=$(python -c 'import random ; print(random.choice(["=0.56", "=0.57.0rc1", ""]))')
           fi
+          numbaver="=0.57.0rc1"  # XXX
           fmm=fast_matrix_market${fmmver}
           awkward=awkward${akver}
           if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') ||
                 startsWith(steps.pyver.outputs.selected, '3.12') }} == true ||
-                ( $(python -c 'import random ; print(random.random() < .50)') == True && ${{ matrix.slowtask != 'notebooks'}} == true ) ]] ; then
+                ( $(python -c 'import random ; print(random.random() < .01)') == True && ${{ matrix.slowtask != 'notebooks'}} == true ) ]] ; then
+            # XXX: change percentage above from .01 to .50
             # Some packages aren't available for pypy or Python 3.12; 50% otherwise (if not running notebooks)
             echo "skipping numba"
             numba=""
@@ -338,7 +341,7 @@ jobs:
             if [[ $H && $normal  ]] ; then if [[ $macos  ]] ; then echo " $suitesparse" ; elif [[ $windows ]] ; then echo " $vanilla"     ; fi ; fi)$( \
             if [[ $H && $bizarro ]] ; then if [[ $macos  ]] ; then echo " $vanilla"     ; elif [[ $windows ]] ; then echo " $suitesparse" ; fi ; fi)
           echo $args
-          coverage run -a -m pytest --color=yes --randomly -v $args \
+          coverage run -a -m pytest -s --color=yes --randomly -v $args \
             ${{ matrix.slowtask == 'pytest_bizarro' && '--runslow' || '' }}
           git checkout .  # Undo changes to scalar default
       - name: Miscellaneous tests
@@ -359,7 +362,7 @@ jobs:
           rm script.py
           # Tests whose coverage depend on order of tests :/
           # TODO: understand why these are order-dependent and try to fix
-          coverage run -a -m pytest --color=yes -x --no-mapnumpy --runslow -k test_binaryop_attributes_numpy graphblas/tests/test_op.py
+          coverage run -a -m pytest -s --color=yes -x --no-mapnumpy --runslow -k test_binaryop_attributes_numpy graphblas/tests/test_op.py
           # coverage run -a -m pytest --color=yes -x --no-mapnumpy -k test_npmonoid graphblas/tests/test_numpyops.py --runslow
       - name: Auto-generated code check
         if: matrix.slowtask == 'pytest_bizarro'
