### 2023-04-13 3f4a3ee
Always test with numpy 1.24
