+++ graphblas/core/operator/unary.py
@@ -235,8 +235,12 @@ class UnaryOp(OpBase):
                         z[0] = unary_udf(x[0])  # pragma: no cover (numba)
 
                 print("wrapping", name, type_, input_type, return_type, file=sys.stderr)
-                unary_wrapper = numba.cfunc(wrapper_sig, nopython=True)(unary_wrapper)
+                cfunc = numba.cfunc(wrapper_sig, nopython=True)
+                print("AAA", file=sys.stderr)
+                unary_wrapper = cfunc(unary_wrapper)
+                print("BBB", file=sys.stderr)
                 new_unary = ffi_new("GrB_UnaryOp*")
+                print("CCC", file=sys.stderr)
                 check_status_carg(
                     lib.GrB_UnaryOp_new(
                         new_unary, unary_wrapper.cffi, ret_type.gb_obj, type_.gb_obj
@@ -244,7 +248,9 @@ class UnaryOp(OpBase):
                     "UnaryOp",
                     new_unary,
                 )
+                print("DDD", file=sys.stderr)
                 op = TypedUserUnaryOp(new_type_obj, name, type_, ret_type, new_unary[0])
+                print("EEE", file=sys.stderr)
                 new_type_obj._add(op)
                 success = True
                 return_types[type_] = ret_type
