+++ .github/workflows/test_and_build.yml
@@ -87,7 +87,7 @@ jobs:
       # since we need to run all slow tests on Windows and non-Windoes OSes.
       matrix:
         # os: ["ubuntu-latest", "macos-latest", "windows-latest"]
-        os: ["windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest"]  # XXX
+        os: ["windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest", "windows-latest"]  # XXX
         slowtask: ["pytest_normal", "pytest_bizarro", "notebooks"]
     env:
       # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
@@ -307,7 +307,7 @@ jobs:
             if [[ $H && $bizarro ]] ; then if [[ $macos  ]] ; then echo " $suitesparse" ; elif [[ $windows ]] ; then echo " $vanilla"     ; fi ; fi)
           echo $args
           pytest -v --pyargs suitesparse_graphblas
-          python -c 'import graphblas as gb ; gb.config.set(mapnumpy=False) ; gb.unary.numpy.exp2'
+          # python -c 'import graphblas as gb ; gb.config.set(mapnumpy=False) ; gb.unary.numpy.exp2'
           coverage run -m pytest -s --color=yes --randomly -v $args \
             ${{ matrix.slowtask == 'pytest_normal' && '--runslow' || '' }}
       - name: Unit tests (bizarro scalars)
@@ -343,7 +343,7 @@ jobs:
             if [[ $H && $normal  ]] ; then if [[ $macos  ]] ; then echo " $suitesparse" ; elif [[ $windows ]] ; then echo " $vanilla"     ; fi ; fi)$( \
             if [[ $H && $bizarro ]] ; then if [[ $macos  ]] ; then echo " $vanilla"     ; elif [[ $windows ]] ; then echo " $suitesparse" ; fi ; fi)
           echo $args
-          python -c 'import graphblas as gb ; gb.config.set(mapnumpy=False) ; gb.unary.numpy.exp2'
+          # python -c 'import graphblas as gb ; gb.config.set(mapnumpy=False) ; gb.unary.numpy.exp2'
           coverage run -a -m pytest -s --color=yes --randomly -v $args \
             ${{ matrix.slowtask == 'pytest_bizarro' && '--runslow' || '' }}
           git checkout .  # Undo changes to scalar default
