+++ .github/workflows/test_and_build.yml
@@ -128,10 +128,10 @@ jobs:
             source
             upstream
           weights: |
+            1000
             1
-            10000
-            1
-            1
+            1000
+            1000
       - name: Setup mamba
         uses: conda-incubator/setup-miniconda@v2
         id: setup_mamba
