+++ .github/workflows/test_and_build.yml
@@ -211,10 +211,10 @@ jobs:
             psgver=""
           fi
           # TODO: drop 0.57.0rc1 and use 0.57 once numba 0.57 is properly released
-          if [[ $npver == "=1.21" ]] ; then
-            numbaver=$(python -c 'import random ; print(random.choice(["=0.55", "=0.56", "=0.57.0rc1", ""]))')
-          elif [[ $npver == "=1.24" ]] ; then
+          if [[ $npver == "=1.24" || ${{ startsWith(steps.pyver.outputs.selected, '3.11') }} == true ]] ; then
             numbaver=$(python -c 'import random ; print(random.choice(["=0.57.0rc1", ""]))')
+          elif [[ $npver == "=1.21" ]] ; then
+            numbaver=$(python -c 'import random ; print(random.choice(["=0.55", "=0.56", "=0.57.0rc1", ""]))')
           else
             numbaver=$(python -c 'import random ; print(random.choice(["=0.56", "=0.57.0rc1", ""]))')
           fi
