### 2023-04-15 e6f94ec
lazy indexunary handling and a bit more coverage
