+++ graphblas/core/formatting.py
@@ -880,6 +880,8 @@ def format_index_expression_html(expr):
         computed = get_expr_result(expr, html=True)
         if "__EXPR__" in computed:
             return computed.replace("<tt>__EXPR__</tt>", topline)
+        # BRANCH NOT COVERED
+        raise Exception("XXX")
 
     keys = []
     values = []
