+++ graphblas/core/operator/select.py
@@ -38,6 +38,8 @@ class ParameterizedSelectOp(ParameterizedUdf):
     __slots__ = "func", "__signature__", "_is_udt"
 
     def __init__(self, name, func, *, anonymous=False, is_udt=False):
+        # NOT COVERED
+        raise Exception("XXX")
         self.func = func
         self.__signature__ = inspect.signature(func)
         self._is_udt = is_udt
@@ -46,11 +48,15 @@ class ParameterizedSelectOp(ParameterizedUdf):
         super().__init__(name, anonymous)
 
     def _call(self, *args, **kwargs):
+        # NOT COVERED
+        raise Exception("XXX")
         sel = self.func(*args, **kwargs)
         sel._parameterized_info = (self, args, kwargs)
         return SelectOp.register_anonymous(sel, self.name, is_udt=self._is_udt)
 
     def __reduce__(self):
+        # NOT COVERED
+        raise Exception("XXX")
         name = f"select.{self.name}"
         if not self._anonymous and name in _STANDARD_OPERATOR_NAMES:
             return name
@@ -58,6 +64,8 @@ class ParameterizedSelectOp(ParameterizedUdf):
 
     @staticmethod
     def _deserialize(name, func, anonymous):
+        # NOT COVERED
+        raise Exception("XXX")
         if anonymous:
             return SelectOp.register_anonymous(func, name, parameterized=True)
         if (rv := SelectOp._find(name)) is not None:
@@ -126,6 +134,8 @@ class SelectOp(OpBase):
         """
         cls._check_supports_udf("register_anonymous")
         if parameterized:
+            # NOT COVERED
+            raise Exception("XXX")
             return ParameterizedSelectOp(name, func, anonymous=True, is_udt=is_udt)
         iop = IndexUnaryOp._build(name, func, anonymous=True, is_udt=is_udt)
         return SelectOp._from_indexunary(iop)
@@ -192,16 +202,22 @@ class SelectOp(OpBase):
         self.is_positional = is_positional
         self._is_udt = is_udt
         if is_udt:
+            # NOT COVERED
+            raise Exception("XXX")
             self._udt_types = {}  # {dtype: DataType}
             self._udt_ops = {}  # {dtype: TypedUserIndexUnaryOp}
 
     def __reduce__(self):
         if self._anonymous:
             if hasattr(self.orig_func, "_parameterized_info"):
+                # NOT COVERED
+                raise Exception("XXX")
                 return (_deserialize_parameterized, self.orig_func._parameterized_info)
             return (self.register_anonymous, (self.orig_func, self.name))
         if (name := f"select.{self.name}") in _STANDARD_OPERATOR_NAMES:
             return name
+        # NOT COVERED
+        raise Exception("XXX")
         return (self._deserialize, (self.name, self.orig_func))
 
     __call__ = TypedBuiltinSelectOp.__call__
