+++ graphblas/core/vector.py
@@ -1989,7 +1989,7 @@ class Vector(BaseType):
             # If we know the dtype, then using `np.fromiter` is much faster
             dtype = lookup_dtype(dtype)
             if dtype.np_type.subdtype is not None and np.__version__[:5] in {"1.21.", "1.22."}:
-                values, dtype = values_to_numpy_buffer(list(d.values()), dtype)
+                values, dtype = values_to_numpy_buffer(list(d.values()), dtype)  # flaky coverage
             else:
                 values = np.fromiter(d.values(), dtype.np_type)
         if size is None and indices.size == 0:
