+++ graphblas/core/utils.py
@@ -132,7 +132,6 @@ def get_shape(nrows, ncols, dtype=None, **arrays):
                 if arr.ndim >= 3:
                     break
                 # BRANCH NOT COVERED
-                raise Exception("XXX")
             elif arr.ndim == 2:
                 break
         else:
@@ -361,7 +360,7 @@ def _autogenerate_code(
     end="# End auto-generated code",
 ):
     """Super low-tech auto-code generation used by automethods.py and infixmethods.py."""
-    with filepath.open() as f:  # flaky coverage
+    with filepath.open() as f:  # pragma: no branch (flaky)
         orig_text = f.read()
     if specializer:
         begin = f"{begin}: {specializer}"
@@ -381,7 +380,7 @@ def _autogenerate_code(
     new_text = orig_text
     for start, stop in reversed(boundaries):
         new_text = f"{new_text[:start]}{begin}{text}{new_text[stop:]}"
-    with filepath.open("w") as f:  # flaky coverage
+    with filepath.open("w") as f:  # pragma: no branch (flaky)
         f.write(new_text)
     import subprocess
 
