### 2023-04-16 7c3c227
Add logo and images to README (#432)
* Add logo and images to README

* link to pypi
