### 2023-04-16 f6c6d27
Make PNGs smaller via https://tinypng.com
