### 2023-04-16 10706aa
mark a couple more lines as flaky coverage
