+++ graphblas/semiring/numpy.py
@@ -134,7 +134,7 @@ __all__ = list(_semiring_names)
 
 def __dir__():
     if not _supports_udfs and not _config.get("mapnumpy"):
-        return globals().keys()
+        return globals().keys()  # FLAKY COVERAGE
     attrs = _delayed.keys() | _semiring_names
     if not _supports_udfs:
         attrs &= {
