+++ graphblas/monoid/numpy.py
@@ -160,7 +160,7 @@ _idempotent = {
 
 def __dir__():
     if not _supports_udfs and not _config.get("mapnumpy"):
-        return globals().keys()
+        return globals().keys()  # FLAKY COVERAGE
     attrs = _delayed.keys() | _monoid_identities.keys()
     if not _supports_udfs:
         attrs &= _numpy_to_graphblas.keys()
