### 2023-04-16 0eb490f
making numba optional (#423)
* make numba optional

* bump pandas 2.0.0

* Don't test against pypy in CI

* expand optional dependencies

* networkx 3.1

* simplify/combine deps

* Make suitesparse_graphblas optional (in the future); don't install numba on 3.11 yet

* Ignore deprecation warning raised by awkward in Python 3.11

* try to ignore warning from sparse

* drop sparse 0.12 (which is really old and annoying to support)

* sre_constants also deprecated

* `shouldhave` make tests a little more clear (I think)

Also, bump version of `fast_matrix_market`, and randomly generate OS for impo
[…]
