### 2023-04-17 python-graphblas/python-graphblas#429 @sneakers-the-rat
OK I see the `_initialize` method in the individual operators now. following that. Yes I think this would all be useful to have in the docs. A simple block diagram would go a long way here I think, to be able to trace "these are the functions/classes/methods brought up from cffi, they are initialized here, and then they end up in this namespace" - nothing too fancy, just something to guide people who are curious about the implementation and might want to become contributors.

Yes upstream metadata would be nice. I pieced through this myself, wondering how much could be inferred from the cffi imports themselves, and it doesn't look like much - I see why a lot is hardcoded here, it's otherwise impossible to tell which method/function goes with what. getting started reading again and hoping to finish today
