### 2023-04-17 python-graphblas/python-graphblas#433 @sneakers-the-rat
[pyos] Pop out constants from methods
I am seeing that what look like constants that describe the various operations/etc. things imported from graphblas are defined within functions in some places, eg. here:
https://github.com/python-graphblas/python-graphblas/blob/0eb490fe6dcb96264fed6637517de01c26f036cb/graphblas/core/operator/unary.py#L342

and then it looks like those have to be redeclared with some additional metadata in the tests, eg here:
https://github.com/python-graphblas/python-graphblas/blob/0eb490fe6dcb96264fed6637517de01c26f036cb/graphblas/tests/test_operator_types.py#L44

I'm not quite sure what that additional metadata is doing, it looks like the tests there are testing whether the list defined within the test matches the modules imported within the different operators, so maybe defining twice is desired behavior, but it seems like regardless you might do yourself a favor and move all the places where some constant-like thing is defined ephemerally as in the example above to a place where you might reuse it or allow it to be inspected. It looks like other constant-like things like regexes to match different types are stored as class attributes, so that might be how you want to do it?
