### 2023-04-17 python-graphblas/python-graphblas#434 @sneakers-the-rat
[pyos] 'C' is not defined in Operations docs
The [Operations docs](https://python-graphblas.readthedocs.io/en/stable/user_guide/operations.html)

have examples like

```python
A = gb.Matrix.from_coo([0, 0, 1, 1, 2], [1, 2, 0, 2, 1],
                       [2.0, 5.0, 1.5, 4.0, 0.5])
B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2], [1, 2, 0, 1, 1, 2],
                       [3., -2., 0., 6., 3., 1.])

# These are equivalent
C << A.ewise_mult(B, op='min')  # method style
C << gb.binary.min(A & B)  # functional style
```

but when I try and do either of these I get `'C' is not defined` which is what I expected from prior documentation. Would recommend making these [doctests](https://www.sphinx-doc.org/en/master/usage/extensions/doctest.html) so that the docs don't include erroneous examples

part of: https://github.com/pyOpenSci/software-submission/issues/81
