### 2023-04-17 python-graphblas/python-graphblas#436 @sneakers-the-rat
[pyos][docs] `to_awkward` and `from_awkward` undocumented
The functions `to_awkward` and `from_awkward` are undocumented:
- https://python-graphblas.readthedocs.io/en/stable/api_reference/io.html
- https://python-graphblas.readthedocs.io/en/stable/user_guide/io.html

+1 for using automodule and splitting i/o module up in package and submodules with module-level docstrings so you don't have to worry about this in the future and just have docs in one place.

Part of: https://github.com/pyOpenSci/software-submission/issues/81
