### 2023-04-17 python-graphblas/python-graphblas#437 @sneakers-the-rat
[pyos][docs] `parameterized`, `is_udt`, and `lazy` undocumented for user functions
The `register_new` and `register_anonymous` class methods for different operators have parameters `parameterized`, `is_udt`, and `lazy` that are tested and seem to work but I don't see them documented.

From what I can see in their implementation, this would include
- what each parameter does
- that `parameterized`, `lazy`, and both being undefined are mutually exclusive (`if: elif: else:`)
- return type (in the case that `lazy=False`)
- any constraints that the function needs to have (these are present implicitly in the descriptions of the operator types but it would be nice to have specific "this function needs to return one of x type" and etc)

Additionally, `register_new` and `register_anonymous` are present in all `OpBase` subclasses, but not in the base class as an abstract method.

part of: https://github.com/pyOpenSci/software-submission/issues/81
