### 2023-04-17 python-graphblas/python-graphblas#438 @sneakers-the-rat
[pyos][checklist] Add link to comparison with `pygraphblas` in README
The review checklist asks me to affirm there is a comparison to similar packages in the ecosystem in the README. I think that could be as simple as linking to https://python-graphblas.readthedocs.io/en/stable/getting_started/faq.html#what-is-the-relationship-between-python-graphblas-and-pygraphblas

Close this if you have already done that, I just looked and didn't see it, but i just scanned really briefly.

Part of: https://github.com/pyOpenSci/software-submission/issues/81
