### 2023-04-17 #81 @sneakers-the-rat
Alright I made it through the checklist, there are some straggling thoughts in the remaining open issues. We can hash those out there and then I'll write my final review :)
