### 2023-04-18 python-graphblas/python-graphblas#433 @eriknw
The current code and approach has served us well up to this point. We have had a multi-year effort to _discover_ how we want operators to behave. For example, we want `gb.unary.exp(1)` to automatically coerce an integer input to use the float operator instead of raise an exception. This code has been updated incrementally and is complicated. By effectively "doing it twice" by doing it differently in the tests serves as a good sanity/consistency check, which has proven helpful. The tests will also be super-helpful if we ever refactor how we do operators, which seems likely.

I do appreciate the scrutiny you've given the operators. They're important, and they're worth making more understandable/maintainable. I agree with much of your input, and you've given a solid "shape" of how things should be after a refactor. We want things to be much more declarative and easier to modify info or behavior of a single operator at a time (such as improve a docstring). When we do this refactor, it will probably address handing multiple backends (cffi-based and not cffi-based) that may have different operators. I expect such a declarative approach will be similar to, but different than the data in the tests. So, I expect the tests to remain largely the same post-refactor, and we can always discard the tests if they are no longer useful and/or become too burdensome.

We could perhaps do a light refactor before this larger refactor where we _name_ and make introspectable all of the hard-coded sets, lists, and dicts that appear within the `_initialize` methods, which I think would be sufficient to close this issue.
