### 2023-04-18 python-graphblas/python-graphblas#434 @eriknw
Thanks for the report! I consider this a bug. Code examples should be copy/paste-able.

I'm also a big fan of using doctests for ensuring correctness, so I'll add this as an item on our roadmap.
