### 2023-04-18 python-graphblas/python-graphblas#435 @eriknw
We've been (slowly) exploring type annotations for a while now in #179.

The argument and return types are already in the numpydoc-style docstrings, so useful is it to have inline type annotations as well?

I think the approach we are leaning toward is to add stub files for type annotations. If we do this, does it make sense to add inline annotations to _some_ functions?

> splitting up the i/o module into a package with submodules

This actually sounds reasonable to me.

Another approach for doing what you want may be to add e.g.:
```python
import typing

if typing.TYPE_CHECKING:
    try:
        from networkx import DiGraph, Graph
    except importError:
        ...
```
