### 2023-04-18 python-graphblas/python-graphblas#438 @eriknw
I added a link to the FAQ to the README, which I think should be sufficient and I would argue is more useful (especially as we get more questions 😉 ) than using valuable space (and attention) to discuss a defunct package.
