### 2023-04-18 python-graphblas/python-graphblas#438 @eriknw
Regarding "comparing to similar packages in the ecosystem", I think it would actually be informative to compare to e.g. `scipy.sparse`, `networkx`, and `igraph`. I think such a comparison could go in the README and/or documentation.
