### 2023-04-19 a3e60c6
Update README.md
Follows up on #438:

>Regarding "comparing to similar packages in the ecosystem", I think it would actually be informative to compare to e.g. scipy.sparse, networkx, and igraph. I think such a comparison could go in the README and/or documentation.
