### 2023-04-19 528f13f
Update io.rst
As raised in #436
