### 2023-04-19 17d5c79
Create output objects in operations.rst examples (#440)
