+++ docs/user_guide/operations.rst
@@ -30,6 +30,7 @@ a Vector is treated as an nx1 column matrix.
                            [2., 5., 1.5, 4.25, 0.5], nrows=4, ncols=4)
     B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2, 3, 3], [1, 2, 0, 1, 1, 2, 0, 1],
                            [3., 2., 9., 6., 3., 1., 0., 5.])
+    C = gb.Matrix(float, A.nrows, B.ncols)
 
     # These are equivalent
     C << A.mxm(B, op='min_plus')  # method style
@@ -69,6 +70,7 @@ a Vector is treated as an nx1 column matrix.
     A = gb.Matrix.from_coo([0, 0, 1, 1, 2], [1, 2, 2, 3, 3],
                            [2., 5., 1.5, 4.25, 0.5], nrows=4, ncols=4)
     v = gb.Vector.from_coo([0, 1, 3], [10., 20., 40.])
+    w = gb.Vector(float, A.nrows)
 
     # These are equivalent
     w << A.mxv(v, op='plus_times')  # method style
@@ -102,6 +104,7 @@ a Vector is treated as an nx1 column matrix.
     v = gb.Vector.from_coo([0, 1, 3], [10., 20., 40.])
     B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2, 3, 3], [1, 2, 0, 1, 1, 2, 0, 1],
                            [3., 2., 9., 6., 3., 1., 0., 5.])
+    u = gb.Vector(float, B.ncols)
 
     # These are equivalent
     u << v.vxm(B, op='plus_plus')  # method style
@@ -149,6 +152,7 @@ Example usage:
                            [2.0, 5.0, 1.5, 4.0, 0.5])
     B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2], [1, 2, 0, 1, 1, 2],
                            [3., -2., 0., 6., 3., 1.])
+    C = gb.Matrix(float, A.nrows, A.ncols)
 
     # These are equivalent
     C << A.ewise_mult(B, op='min')  # method style
@@ -221,10 +225,11 @@ should be used with the functional syntax, ``left_default`` and ``right_default`
 
 .. code-block:: python
 
-    A = gb.Matrix.from_coo([0, 0, 0, 1, 1, 1], [0, 1, 2, 0, 2],
+    A = gb.Matrix.from_coo([0, 0, 0, 1, 1], [0, 1, 2, 0, 2],
                            [9.0, 2.0, 5.0, 1.5, 4.0], nrows=3)
     B = gb.Matrix.from_coo([0, 0, 0, 2, 2, 2], [0, 1, 2, 0, 1, 2],
                            [4., 0., -2., 6., 3., 1.])
+    C = gb.Matrix(float, A.nrows, A.ncols)
 
     # These are equivalent
     C << A.ewise_add(B, op='minus')  # method style
@@ -258,10 +263,11 @@ should be used with the functional syntax, ``left_default`` and ``right_default`
 
 .. code-block:: python
 
-    A = gb.Matrix.from_coo([0, 0, 0, 1, 1, 1], [0, 1, 2, 0, 2],
+    A = gb.Matrix.from_coo([0, 0, 0, 1, 1], [0, 1, 2, 0, 2],
                            [9.0, 2.0, 5.0, 1.5, 4.0], nrows=3)
     B = gb.Matrix.from_coo([0, 0, 0, 2, 2, 2], [0, 1, 2, 0, 1, 2],
                            [4., 0., -2., 6., 3., 1.])
+    C = gb.Matrix(float, A.nrows, A.ncols)
 
     # These are equivalent
[…]
