### 2023-04-19 8617f24
Add "2.0" as the spec version here in readme too
