### 2023-04-20 22a1971
Don't install numba rc on Windows (fix)
