### 2023-04-21 d1beef6
Document parameters in operator registrations methods
