### 2023-04-25 python-graphblas/python-graphblas#433 @sneakers-the-rat
Agreed :). And i am only not adding "of course y'all know the needs and structure of the package better than I do, and I am assuming that there are reasons for everything being the way it is" to all my issues here for brevity lol.

Feel free to close this or keep it open as a reminder, whatever is more useful!
