### 2023-04-25 2bd028c
Add more details to docstrings about UDF func
