+++ graphblas/core/operator/binary.py
@@ -636,6 +636,7 @@ class BinaryOp(OpBase):
         func : FunctionType
             The function to compile. For all current backends, this must be able
             to be compiled with ``numba.njit``.
+            ``func`` takes two input parameters of any dtype and returns any dtype.
         name : str, optional
             The name of the operator. This *does not* show up as ``gb.binary.{name}``.
         parameterized : bool, default False
@@ -676,6 +677,7 @@ class BinaryOp(OpBase):
         func : FunctionType
             The function to compile. For all current backends, this must be able
             to be compiled with ``numba.njit``.
+            ``func`` takes two input parameters of any dtype and returns any dtype.
         parameterized : bool, default False
             When True, create a parameterized user-defined operator, which means
             additional parameters can be "baked into" the operator when used.
