+++ graphblas/core/operator/indexunary.py
@@ -251,7 +251,7 @@ class IndexUnaryOp(OpBase):
             The function to compile. For all current backends, this must be able
             to be compiled with ``numba.njit``.
             ``func`` takes four input parameters--any dtype, int64, int64,
-             any dtype and returns any dtype. The first argument (any dtype) is
+            any dtype and returns any dtype. The first argument (any dtype) is
             the value of the input Matrix or Vector, the second argument (int64)
             is the row index of the Matrix or the index of the Vector, the third
             argument (int64) is the column index of the Matrix or 0 for a Vector,
@@ -300,7 +300,7 @@ class IndexUnaryOp(OpBase):
             The function to compile. For all current backends, this must be able
             to be compiled with ``numba.njit``.
             ``func`` takes four input parameters--any dtype, int64, int64,
-             any dtype and returns any dtype. The first argument (any dtype) is
+            any dtype and returns any dtype. The first argument (any dtype) is
             the value of the input Matrix or Vector, the second argument (int64)
             is the row index of the Matrix or the index of the Vector, the third
             argument (int64) is the column index of the Matrix or 0 for a Vector,
