### 2023-04-26 python-graphblas/python-graphblas#407 @eriknw
Thanks for the suggestions and pointers. Example notebooks in the docs sound great to me. We should probably give the existing notebooks some attention and revise them as appropriate.
