### 2023-04-27 python-graphblas/python-graphblas#438 @sneakers-the-rat
Good with both of those :)
