### 2023-05-03 096c5ed
Document parameters in operator registrations methods (#446)
* Document parameters in operator registrations methods

* Change to `pip install .[all]` instead of `pip install .[complete]`

* numba 0.57.0 released
