### 2023-05-03 0cf2bbd
Use external link for images in README (to show up on PyPI)
