### 2023-05-03 0eeebfc
Use external link for images in README (to show up on PyPI) (#450)
