### 2023-05-06 5a2ad3c
bump ruff (also, test pre-commit.ci)
