+++ graphblas/tests/test_vector.py
@@ -858,7 +858,7 @@ def test_indexunary_udf(v):
     delattr(indexunary, "iin")
     delattr(select, "iin")
     with pytest.raises(UdfParseError, match="Unable to parse function using Numba"):
-        indexunary.register_new("bad", lambda x, row, col, thunk: result)
+        indexunary.register_new("bad", lambda x, row, col, thunk: result) # pragma: no branch
 
 
 def test_reduce(v):
@@ -2425,7 +2425,7 @@ def test_lambda_udfs(v):
     # with pytest.raises(TypeError):
     v.ewise_add(v, lambda x, y: x + y)  # pragma: no branch (numba)
     with pytest.raises(TypeError):
-        v.inner(v, lambda x, y: x + y)
+        v.inner(v, lambda x, y: x + y)  # pragma: no branch (numba)
 
 
 def test_get(v):
