### 2023-05-06 756dc68
pre-commit.ci works fine, let's only use it for linting
