### 2023-05-06 f140f9c
include conftest.py in MANIFEST.in (#451)
* include conftest.py in MANIFEST.in

* Run slow tests when running via `pytest --pyargs graphblas`

* bump ruff (also, test pre-commit.ci)

* pre-commit.ci works fine, let's only use it for linting
