+++ MANIFEST.in
@@ -2,6 +2,7 @@ recursive-include graphblas *.py
 prune docs
 prune scripts
 include setup.py
+include conftest.py
 include README.md
 include LICENSE
 include MANIFEST.in
