### 2023-05-08 #81 @eriknw
I'd like to share most of the PRs that the pyOpenSci reviews have strongly influenced:
- https://github.com/python-graphblas/python-graphblas/pull/385
- https://github.com/python-graphblas/python-graphblas/pull/397
- https://github.com/python-graphblas/python-graphblas/pull/400
- https://github.com/python-graphblas/python-graphblas/pull/404
- https://github.com/python-graphblas/python-graphblas/pull/413
- https://github.com/python-graphblas/python-graphblas/pull/420
- https://github.com/python-graphblas/python-graphblas/pull/432
- https://github.com/python-graphblas/python-graphblas/pull/439
- https://github.com/python-graphblas/python-graphblas/pull/440
- https://github.com/python-graphblas/python-graphblas/pull/441
- https://github.com/python-graphblas/python-graphblas/pull/442
- https://github.com/python-graphblas/python-graphblas/pull/443
- https://github.com/python-graphblas/python-graphblas/pull/446
- https://github.com/python-graphblas/python-graphblas/pull/447
- https://github.com/python-graphblas/python-graphblas/pull/448
- https://github.com/GraphBLAS/python-suitesparse-graphblas/pull/85
- https://github.com/python-graphblas/graphblas-algorithms/pull/54
- https://github.com/python-graphblas/graphblas-algorithms/pull/60

Since the review began, we have also:
- automated wheel building of the upstream dependency that wraps the C library.
  - It supports all architectures/platforms/os/python versions that numpy supports.
  - See: https://pypi.org/project/suitesparse-graphblas/7.4.4.1a1/#files
  - This was 90% done by an outside contributor
- added support for Python 3.11 (numba was finally released for 3.11, but we also just made numba optional)
- presented a poster at PyConUS

The reviewers have also indicated where we can/should refactor code (operators!) to make it more understandable and maintainable, and specific suggestions for how to further improve documentation. I expect implementing these changes will keep us busy for most of the year.

@sneakers-the-rat @szhorvat if there are any specific issues you would like us to address before you finish your reviews (or as conditional for acceptance), please let us know and please be specific. There are many issues, checklists, and discussions for us to read that it's hard to know exactly where things stand (but the reviews have been great--thank you!). For example, I think a specific enhancement that y'all probably really, really want is for the operator objects to be listed in the API reference of the documentation (definitely a good idea). Should we add this "quick and ugly" right now, then redo it later in the year after we refactor operators to be more declarative, or just do it later after the refactor?
