+++ graphblas/core/ss/matrix.py
@@ -63,7 +63,7 @@ def head(matrix, n=10, dtype=None, *, sort=False):
 
 
 def _concat_mn(tiles, *, is_matrix=None):
-    """Argument checking for `Matrix.ss.concat` and returns number of tiles in each dimension."""
+    """Argument checking for ``Matrix.ss.concat`` and returns number of tiles in each dimension."""
     from ..matrix import Matrix, TransposedMatrix
     from ..vector import Vector
 
@@ -261,8 +261,8 @@ class ss:
         vector : Vector
             Create a diagonal from this Vector.
         k : int, default 0
-            Diagonal in question.  Use `k>0` for diagonals above the main diagonal,
-            and `k<0` for diagonals below the main diagonal.
+            Diagonal in question.  Use ``k>0`` for diagonals above the main diagonal,
+            and ``k<0`` for diagonals below the main diagonal.
 
         See Also
         --------
@@ -282,12 +282,12 @@ class ss:
         """
         GxB_Matrix_split.
 
-        Split a Matrix into a 2D array of sub-matrices according to `chunks`.
+        Split a Matrix into a 2D array of sub-matrices according to ``chunks``.
 
         This performs the opposite operation as ``concat``.
 
-        `chunks` is short for "chunksizes" and indicates the chunk sizes for each dimension.
-        `chunks` may be a single integer, or a length 2 tuple or list.  Example chunks:
+        ``chunks`` is short for "chunksizes" and indicates the chunk sizes for each dimension.
+        ``chunks`` may be a single integer, or a length 2 tuple or list.  Example chunks:
 
         - ``chunks=10``
             - Split each dimension into chunks of size 10 (the last chunk may be smaller).
@@ -295,7 +295,7 @@ class ss:
             - Split rows into chunks of size 10 and columns into chunks of size 20.
         - ``chunks=(None, [5, 10])``
             - Don't split rows into chunks, and split columns into two chunks of size 5 and 10.
-        ` ``chunks=(10, [20, None])``
+        - ``chunks=(10, [20, None])``
             - Split columns into two chunks of size 20 and ``ncols - 20``
 
         See Also
@@ -366,9 +366,9 @@ class ss:
 
         Concatenate a 2D list of Matrix objects into the current Matrix.
         Any existing values in the current Matrix will be discarded.
-        To concatenate into a new Matrix, use `graphblas.ss.concat`.
+        To concatenate into a new Matrix, use ``graphblas.ss.concat``.
 
-        Vectors may be used as `Nx1` Matrix objects.
+        Vectors may be used as ``Nx1`` Matrix objects.
 
         This performs the opposite operation as ``split``.
 
@@ -542,8 +542,8 @@ class ss:
         Parameters
         ----------
         format : str, optional
-            If `format` is not specified, this method exports in the currently stored format.
-            To control the export format, set `format` to one of:
+            If ``format`` is not specified, this method exports in the currently stored format.
+            To control the export format, set ``format`` to one of:
                 - "csr"
                 - "csc"
                 - "hypercsr"
@@ -578,7 +578,7 @@ class ss:
 
         Returns
         -------
-        dict; keys depend on `format` and `raw` arguments (see below).
+        dict; keys depend on ``format`` and ``raw`` arguments (see below).
 
         See Also
         --------
@@ -732,10 +732,10 @@ class ss:
         """
         GxB_Matrix_unpack_xxx.
 
-        `unpack` is like `export`, except that the Matrix remains valid but empty.
-        `pack_*` methods are the opposite of `unpack`.
+        ``unpack`` is like ``export``, except that the Matrix remains valid but empty.
+        ``pack_*`` methods are the opposite of ``unpack``.
 
[…]
