### 2023-05-14 4c16653
Use double (not single) backtick in docstrings (#454)
* Use double (not single) backtick in docstrings

* That's annoying; why are scipy tests missing?!
