+++ graphblas/core/agg.py
@@ -1,8 +1,8 @@
 """graphblas.core.agg namespace is deprecated; please use graphblas.core.operator.agg instead.
 
 .. deprecated:: 2023.3.0
-`graphblas.core.agg` will be removed in a future release.
-Use `graphblas.core.operator.agg` instead.
+``graphblas.core.agg`` will be removed in a future release.
+Use ``graphblas.core.operator.agg`` instead.
 Will be removed in version 2023.11.0 or later.
 
 """
