+++ graphblas/viz.py
@@ -67,7 +67,7 @@ def draw(m):  # pragma: no cover
 
 
 def spy(M, *, centered=False, show=True, figure=None, axes=None, figsize=None, **kwargs):
-    """Plot the sparsity pattern of a Matrix using `matplotlib.spy`.
+    """Plot the sparsity pattern of a Matrix using ``matplotlib.spy``.
 
     See:
     - https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.spy.html
@@ -106,8 +106,8 @@ def spy(M, *, centered=False, show=True, figure=None, axes=None, figsize=None, *
 def datashade(M, agg="count", *, width=None, height=None, opts_kwargs=None, **kwargs):
     """Interactive plot of the sparsity pattern of a Matrix using hvplot and datashader.
 
-    The `datashader` library rasterizes large data into a 2d grid of pixels.  Each pixel
-    may contain multiple data points, which are combined by an aggregator (`agg="count"`).
+    The ``datashader`` library rasterizes large data into a 2d grid of pixels.  Each pixel
+    may contain multiple data points, which are combined by an aggregator (``agg="count"``).
     Common aggregators are "count", "sum", "mean", "min", and "max".  See full list here:
     - https://datashader.org/api.html#reductions
 
