### 2023-05-17 f3d7e7d
blacken docs (and other misc)
This doesn't actually add `blacken-docs` as pre-commit hook.

Also, do a few misc changes pointed out by https://scikit-hep.org/developer/reporeview
