+++ docs/user_guide/operations.rst
@@ -26,14 +26,22 @@ a Vector is treated as an nx1 column matrix.
 
 .. code-block:: python
 
-    A = gb.Matrix.from_coo([0, 0, 1, 1, 2], [1, 2, 2, 3, 3],
-                           [2., 5., 1.5, 4.25, 0.5], nrows=4, ncols=4)
-    B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2, 3, 3], [1, 2, 0, 1, 1, 2, 0, 1],
-                           [3., 2., 9., 6., 3., 1., 0., 5.])
+    A = gb.Matrix.from_coo(
+        [0, 0, 1, 1, 2],
+        [1, 2, 2, 3, 3],
+        [2., 5., 1.5, 4.25, 0.5],
+        nrows=4,
+        ncols=4
+    )
+    B = gb.Matrix.from_coo(
+        [0, 0, 1, 1, 2, 2, 3, 3],
+        [1, 2, 0, 1, 1, 2, 0, 1],
+        [3., 2., 9., 6., 3., 1., 0., 5.]
+    )
     C = gb.Matrix(float, A.nrows, B.ncols)
 
     # These are equivalent
-    C << A.mxm(B, op='min_plus')  # method style
+    C << A.mxm(B, op="min_plus")  # method style
     C << gb.semiring.min_plus(A @ B)  # functional style
 
 .. csv-table:: A
@@ -67,13 +75,18 @@ a Vector is treated as an nx1 column matrix.
 
 .. code-block:: python
 
-    A = gb.Matrix.from_coo([0, 0, 1, 1, 2], [1, 2, 2, 3, 3],
-                           [2., 5., 1.5, 4.25, 0.5], nrows=4, ncols=4)
+    A = gb.Matrix.from_coo(
+        [0, 0, 1, 1, 2],
+        [1, 2, 2, 3, 3],
+        [2., 5., 1.5, 4.25, 0.5],
+        nrows=4,
+        ncols=4
+    )
     v = gb.Vector.from_coo([0, 1, 3], [10., 20., 40.])
     w = gb.Vector(float, A.nrows)
 
     # These are equivalent
-    w << A.mxv(v, op='plus_times')  # method style
+    w << A.mxv(v, op="plus_times")  # method style
     w << gb.semiring.plus_times(A @ v)  # functional style
 
 .. csv-table:: A
@@ -102,12 +115,15 @@ a Vector is treated as an nx1 column matrix.
 .. code-block:: python
 
     v = gb.Vector.from_coo([0, 1, 3], [10., 20., 40.])
-    B = gb.Matrix.from_coo([0, 0, 1, 1, 2, 2, 3, 3], [1, 2, 0, 1, 1, 2, 0, 1],
-                           [3., 2., 9., 6., 3., 1., 0., 5.])
+    B = gb.Matrix.from_coo(
+        [0, 0, 1, 1, 2, 2, 3, 3],
+        [1, 2, 0, 1, 1, 2, 0, 1],
+        [3., 2., 9., 6., 3., 1., 0., 5.]
+    )
     u = gb.Vector(float, B.ncols)
 
     # These are equivalent
-    u << v.vxm(B, op='plus_plus')  # method style
+    u << v.vxm(B, op="plus_plus")  # method style
     u << gb.semiring.plus_plus(v @ B)  # functional style
 
 .. csv-table:: v
@@ -148,14 +164,20 @@ Example usage:
 
 .. code-block:: python
 
-    A = gb.Matrix.from_coo([0, 0, 1, 1, 2], [1, 2, 0, 2, 1],
-                           [2.0, 5.0, 1.5, 4.0, 0.5])
[…]
