+++ .github/workflows/test_and_build.yml
@@ -17,6 +17,10 @@ on:
     branches:
       - main
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
 jobs:
   rngs:
     # To achieve consistent coverage, we need a little bit of correlated collaboration.
