+++ docs/user_guide/udf.rst
@@ -21,7 +21,7 @@ Example user-defined UnaryOp:
             return x + 1
         return x
 
-    unary.register_new('force_odd', force_odd_func)
+    unary.register_new("force_odd", force_odd_func)
 
     v = Vector.from_coo([0, 1, 3, 4, 5], [1, 2, 3, 8, 14])
     w = v.apply(unary.force_odd).new()
