+++ graphblas/ss/__init__.py
@@ -1,1 +1,2 @@
-from ._core import about, concat, config, diag
+from ..core.ss.context import Context, global_context
+from ._core import about, concat, config, diag, jit
