+++ graphblas/core/base.py
@@ -571,7 +571,7 @@ class BaseExpression:
         ):
             if opts:
                 # Ignore opts for now
-                desc = descriptor_lookup(**opts)  # noqa: F841
+                desc = descriptor_lookup(**opts)  # noqa: F841 (keep desc in scope for context)
             if self._is_scalar and self._value._is_cscalar != is_cscalar:
                 return self._value.dup(is_cscalar=is_cscalar, name=name)
             rv = self._value
