+++ graphblas/core/ss/context.py
@@ -1,19 +1,10 @@
 import threading
 
-from ...exceptions import check_status, check_status_carg
+from ...exceptions import InvalidValue, check_status, check_status_carg
 from .. import ffi, lib
 from .config import BaseConfig
 
 ffi_new = ffi.new
-
-
-class ThreadLocal(threading.local):
-    """Hold the active context for the current thread."""
-
-    context = None
-
-
-threadlocal = ThreadLocal()
 
 
 class Context(BaseConfig):
@@ -26,62 +17,88 @@ class Context(BaseConfig):
     _defaults = {
         "nthreads": 0,
         "chunk": 0,
-        "gpu_id": -1,  # -1 means no GPU (I think)
+        "gpu_id": -1,  # -1 means no GPU
     }
 
-    def __init__(self, engage=True, *, nthreads=None, chunk=None, gpu_id=None):
+    def __init__(self, engage=True, *, stack=True, nthreads=None, chunk=None, gpu_id=None):
         super().__init__()
-        if nthreads is not None:
-            self["nthreads"] = nthreads
-        if chunk is not None:
-            self["chunk"] = chunk
-        if gpu_id is not None:
-            self["gpu_id"] = gpu_id
-        if engage:
-            self.engage()
-
-    def __new__(cls, engage=True, **opts):
-        self = object.__new__(cls)
         self.gb_obj = ffi_new("GxB_Context*")
         check_status_carg(lib.GxB_Context_new(self.gb_obj), "Context", self.gb_obj[0])
-        return self
+        if stack:
+            context = threadlocal.context
+            self["nthreads"] = context["nthreads"] if nthreads is None else nthreads
+            self["chunk"] = context["chunk"] if chunk is None else chunk
+            self["gpu_id"] = context["gpu_id"] if gpu_id is None else gpu_id
+        else:
+            if nthreads is not None:
+                self["nthreads"] = nthreads
+            if chunk is not None:
+                self["chunk"] = chunk
+            if gpu_id is not None:
+                self["gpu_id"] = gpu_id
+        self._prev_context = None
+        if engage:
+            self.engage()
 
     @classmethod
     def _from_obj(cls, gb_obj=None):
         self = object.__new__(cls)
         self.gb_obj = gb_obj
-        self.__init__(engage=False)
-        return self
-
-    @classmethod
-    def _maybe_new(cls):
-        if threadlocal.context is not None:
-            return threadlocal.context
-        self = cls(engage=False)
-        check_status(lib.GxB_Context_engage(self._carg), self)
-        # Don't assign to threadlocal.context; instead, let it disengage upon going out of scope
+        self._prev_context = None
+        super().__init__(self)
         return self
 
     @property
     def _carg(self):
         return self.gb_obj[0]
 
+    def dup(self, engage=True, *, nthreads=None, chunk=None, gpu_id=None):
+        if nthreads is None:
+            nthreads = self["nthreads"]
+        if chunk is None:
+            chunk = self["chunk"]
+        if gpu_id is None:
+            gpu_id = self["gpu_id"]
+        return type(self)(engage, stack=False, nthreads=nthreads, chunk=chunk, gpu_id=gpu_id)
+
     def __del__(self):
         gb_obj = getattr(self, "gb_obj", None)
         if gb_obj is not None and lib is not None:  # pragma: no branch (safety)
             try:
-                lib.GxB_Context_disengage(gb_obj[0])
-            finally:
-                check_status(lib.GxB_Context_free(gb_obj), self)
+                self.disengage()
+            except InvalidValue:
+                pass
+            lib.GxB_Context_free(gb_obj)
 
     def engage(self):
+        if self._prev_context is None and (context := threadlocal.context) is not self:
+            self._prev_context = context
         check_status(lib.GxB_Context_engage(self._carg), self)
[…]
