+++ graphblas/core/expr.py
@@ -421,7 +421,7 @@ class Updater:
             # Fast path using assignElement
             if self.opts:
                 # Ignore opts for now
-                desc = descriptor_lookup(**self.opts)  # noqa: F841
+                desc = descriptor_lookup(**self.opts)  # noqa: F841 (keep desc in scope for context)
             self.parent._assign_element(resolved_indexes, obj)
         else:
             mask = self.kwargs.get("mask")
