### 2023-05-25 #81 @sneakers-the-rat
> For example, I think a specific enhancement that y'all probably really, really want is for the operator objects to be listed in the API reference of the documentation (definitely a good idea). Should we add this "quick and ugly" right now, then redo it later in the year after we refactor operators to be more declarative, or just do it later after the refactor?

specifically on this - I would not want to encourage you to do anything that is quick and dirty or otherwise not part of your normal planned development process. I trust y'all at this point to make recommended changes as you are able to, and given that there are several ways to list those functions/methods/classes as is, and the docs are to some degree intertwined with the graphviz docs, I don't think their immediate absence is disqualifying to the "all user-facing functions documented" check.
