+++ docs/api_reference/io.rst
@@ -52,12 +52,19 @@ These methods require `scipy <https://scipy.org/>`_ to be installed.
 Awkward Array
 ~~~~~~~~~~~~~
 
-`Awkward Array <https://awkward-array.org/doc/main/>`_ is a library for nested, variable-sized data, including arbitrary-length lists,
-records, mixed types, and missing data, using NumPy-like idioms.
-
-.. autofunction:: graphblas.io.from_awkward
+`Awkward Array <https://awkward-array.org/doc/main/>`_ is a library for nested,
+variable-sized data, including arbitrary-length lists, records, mixed types,
+and missing data, using NumPy-like idioms. Note that the intended use of the
+`awkward-array`-related `io` functions is to convert `graphblas` objects to awkward,
+perform necessary computations/transformations and, if required, convert the
+awkward array back to `graphblas` format. To facilitate this conversion process, 
+`graphblas.io.to_awkward` adds top-level attribute `format`, describing the
+format of the `graphblas` object (this attributed is used by the 
+`graphblas.io.from_awkward` function to reconstruct the `graphblas` object).
 
 .. autofunction:: graphblas.io.to_awkward
+
+.. autofunction:: graphblas.io.from_awkward
 
 Visualization
 ~~~~~~~~~~~~~
