### 2023-06-02 55ab07d
Update the doc string and change the order of the functions to reflect their intended use case.
