### 2023-06-02 052ff08
Show bash command
