### 2023-06-02 c238f5e
use `set -x` instead
