### 2023-06-02 79dc531
Don't install python-suitesparse-graphblas from upstream (which needs ss:gb 8.0)
