+++ .github/workflows/test_and_build.yml
@@ -131,9 +131,9 @@ jobs:
             source
             upstream
           weights: |
-            1
-            1
-            1
+            1000000
+            1000000
+            1000000
             1
       - name: Setup mamba
         uses: conda-incubator/setup-miniconda@v2
