### 2023-06-02 3b6e654
Use `set -x` more strategically
