+++ .github/workflows/test_and_build.yml
@@ -259,7 +259,7 @@ jobs:
             ${{ matrix.slowtask == 'pytest_bizarro' && 'black' || '' }} \
             ${{ matrix.slowtask == 'notebooks' && 'matplotlib nbconvert jupyter "ipython>=7"' || '' }} \
             ${{ steps.sourcetype.outputs.selected == 'upstream' && 'cython' || '' }} \
-            "${{ steps.sourcetype.outputs.selected != 'wheel' && 'graphblas>=7.4.0,<8.0' || 'python' }}" \
+            '"'${{ steps.sourcetype.outputs.selected != 'wheel' && 'graphblas>=7.4.0,<8.0' || 'python' }}'"' \
             ${{ contains(steps.pyver.outputs.selected, 'pypy') && 'pypy' || '' }}
       - name: Build extension module
         run: |
