+++ graphblas/ss/__init__.py
@@ -1,2 +1,2 @@
 from ..core.ss.context import Context, global_context
-from ._core import about, concat, config, diag, jit
+from ._core import about, concat, config, diag
