+++ docs/env.yml
@@ -8,7 +8,7 @@ dependencies:
     # python-graphblas dependencies
     - donfig
     - numba
-    - python-suitesparse-graphblas>=8.0.0.0
+    - python-suitesparse-graphblas>=8.0.1.0
     - pyyaml
     # extra dependencies
     - matplotlib
