### 2023-06-03 a1b757e
Make `dtypes` a directory so we can add `gb.dtypes.ss`
