+++ graphblas/dtypes/_core.py
@@ -1,20 +1,16 @@
-import warnings as _warnings
+import warnings
 
-import numpy as _np
-from numpy import find_common_type as _find_common_type
-from numpy import promote_types as _promote_types
+import numpy as np
+from numpy import find_common_type, promote_types
 
-from . import backend
-from .core import NULL as _NULL
-from .core import _has_numba
-from .core import ffi as _ffi
-from .core import lib as _lib
+from .. import backend, dtypes
+from ..core import NULL, _has_numba, ffi, lib
 
 if _has_numba:
-    import numba as _numba
+    import numba
 
 # Default assumption unless FC32/FC64 are found in lib
-_supports_complex = hasattr(_lib, "GrB_FC64") or hasattr(_lib, "GxB_FC64")
+_supports_complex = hasattr(lib, "GrB_FC64") or hasattr(lib, "GxB_FC64")
 
 
 class DataType:
@@ -26,7 +22,7 @@ class DataType:
         self.gb_name = gb_name
         self.c_type = c_type
         self.numba_type = numba_type
-        self.np_type = _np.dtype(np_type)
+        self.np_type = np.dtype(np_type)
 
     def __repr__(self):
         return self.name
@@ -62,7 +58,7 @@ class DataType:
 
     @property
     def _is_anonymous(self):
-        return globals().get(self.name) is not self
+        return getattr(dtypes, self.name, None) is not self
 
     @property
     def _is_udt(self):
@@ -80,27 +76,27 @@ class DataType:
 def register_new(name, dtype):
     if not name.isidentifier():
         raise ValueError(f"`name` argument must be a valid Python identifier; got: {name!r}")
-    if name in _registry or name in globals():
+    if name in _registry or hasattr(dtypes, name):
         raise ValueError(f"{name!r} name for dtype is unavailable")
     rv = register_anonymous(dtype, name)
     _registry[name] = rv
-    globals()[name] = rv
+    setattr(dtypes, name, rv)
     return rv
 
 
 def register_anonymous(dtype, name=None):
     try:
-        dtype = _np.dtype(dtype)
+        dtype = np.dtype(dtype)
     except TypeError:
         if isinstance(dtype, dict):
             # Allow dtypes such as `{'x': int, 'y': float}` for convenience
-            dtype = _np.dtype([(key, lookup_dtype(val).np_type) for key, val in dtype.items()])
+            dtype = np.dtype([(key, lookup_dtype(val).np_type) for key, val in dtype.items()])
         elif isinstance(dtype, str) and "[" in dtype and dtype.endswith("]"):
             # Allow dtypes such as `"INT64[3, 4]"` for convenience
             base_dtype, shape = dtype.split("[", 1)
             base_dtype = lookup_dtype(base_dtype)
-            shape = _np.lib.format.safe_eval(f"[{shape}")
-            dtype = _np.dtype((base_dtype.np_type, shape))
+            shape = np.lib.format.safe_eval(f"[{shape}")
+            dtype = np.dtype((base_dtype.np_type, shape))
         else:
             raise
     if dtype in _registry:
@@ -114,36 +110,36 @@ def register_anonymous(dtype, name=None):
     if dtype.hasobject:
         raise ValueError("dtype must not allow Python objects")
 
-    from .exceptions import check_status_carg
[…]
