+++ pyproject.toml
@@ -138,6 +138,7 @@ packages = [
     "graphblas.core",
     "graphblas.core.operator",
     "graphblas.core.ss",
+    "graphblas.dtypes",
     "graphblas.indexunary",
     "graphblas.io",
     "graphblas.monoid",
