### 2023-06-07 8ec612f
bump ruff and mark more slow tests
