### 2023-06-07 fbdeffb
Update io.rst
remove trailing blanks
