### 2023-06-08 #81 @eriknw
Thanks for checking in @sneakers-the-rat, and I hope you (and everyone) are feeling better.

I realize the timing on this review has been a bit... flexible, that March 29 (https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1458089302) has come and gone, and we're pretty far off the six week schedule that pyopensci aims for (https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1484897108).

If possible, can we refocus and try to get this review finished-finished before the SciPy conference, July 10th?

We've gotten a lot of value from this review already, and I appreciate all the time and effort the volunteers have spent. Thanks.
