+++ scripts/check_versions.sh
@@ -12,6 +12,6 @@ conda search 'sparse[channel=conda-forge]>=0.14.0'
 conda search 'fast_matrix_market[channel=conda-forge]>=1.6.0'
 conda search 'numba[channel=conda-forge]>=0.57.0'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
-conda search 'flake8-bugbear[channel=conda-forge]>=23.5.9'
+conda search 'flake8-bugbear[channel=conda-forge]>=23.6.5'
 conda search 'flake8-simplify[channel=conda-forge]>=0.20.0'
 # conda search 'python[channel=conda-forge]>=3.8 *pypy*'
