+++ graphblas/dtypes/ss.py
@@ -1,0 +1,1 @@
+from ..core.ss.dtypes import register_new  # noqa: F401
