+++ graphblas/unary/ss.py
@@ -1,3 +1,4 @@
 from ..core import operator
+from ..core.ss.unary import register_new
 
 del operator
