+++ pyproject.toml
@@ -351,7 +351,7 @@ ignore = [
     "TCH",  # flake8-type-checking (Note: figure out type checking later)
     "ARG",  # flake8-unused-arguments (Sometimes helpful, but too strict)
     "TD",  # flake8-todos (Maybe okay to add some of these)
-    "T", "T001", "T002", "T003", "T004",  # flake8-fixme (like flake8-todos)
+    "FIX",  # flake8-fixme (like flake8-todos)
     "ERA",  # eradicate (We like code in comments!)
     "PD",  # pandas-vet (Intended for scripts that use pandas, not libraries)
 ]
