### 2023-06-16 #81 @sneakers-the-rat
SORRY YES i am still here i am just trying to catch up at work. i can do the 21st for sure. i will shoot for tomorrow.
