### 2023-06-18 e44c08f
JIT for binary and indexunary operators (untested, undocumented)
