+++ graphblas/indexunary/ss.py
@@ -1,0 +1,4 @@
+from ..core import operator
+from ..core.ss.indexunary import register_new  # noqa: F401
+
+del operator
