+++ graphblas/binary/ss.py
@@ -1,3 +1,4 @@
 from ..core import operator
+from ..core.ss.binary import register_new  # noqa: F401
 
 del operator
