+++ scripts/check_versions.sh
@@ -7,7 +7,7 @@ conda search 'numpy[channel=conda-forge]>=1.24.3'
 conda search 'pandas[channel=conda-forge]>=2.0.2'
 conda search 'scipy[channel=conda-forge]>=1.10.1'
 conda search 'networkx[channel=conda-forge]>=3.1'
-conda search 'awkward[channel=conda-forge]>=2.2.1'
+conda search 'awkward[channel=conda-forge]>=2.2.2'
 conda search 'sparse[channel=conda-forge]>=0.14.0'
 conda search 'fast_matrix_market[channel=conda-forge]>=1.6.0'
 conda search 'numba[channel=conda-forge]>=0.57.0'
