+++ graphblas/unary/ss.py
@@ -1,4 +1,4 @@
 from ..core import operator
-from ..core.ss.unary import register_new
+from ..core.ss.unary import register_new  # noqa: F401
 
 del operator
