+++ graphblas/indexunary/__init__.py
@@ -4,7 +4,7 @@ _delayed = {}
 
 
 def __dir__():
-    return globals().keys() | _delayed.keys()
+    return globals().keys() | _delayed.keys() | {"ss"}
 
 
 def __getattr__(key):
@@ -13,6 +13,18 @@ def __getattr__(key):
         rv = func(**kwargs)
         globals()[key] = rv
         return rv
+    if key == "ss":
+        from .. import backend
+
+        if backend != "suitesparse":
+            raise AttributeError(
+                f'module {__name__!r} only has attribute "ss" when backend is "suitesparse"'
+            )
+        from importlib import import_module
+
+        ss = import_module(".ss", __name__)
+        globals()["ss"] = ss
+        return ss
     raise AttributeError(f"module {__name__!r} has no attribute {key!r}")
 
 
