### 2023-06-21 620735c
Add test for suitesparse jit
