+++ graphblas/select/ss.py
@@ -1,5 +1,5 @@
 from ..core import operator
-from ..core.ss.binary import register_new  # noqa: F401
+from ..core.ss.select import register_new  # noqa: F401
 
 _delayed = {}
 
