+++ graphblas/indexunary/ss.py
@@ -1,4 +1,6 @@
 from ..core import operator
 from ..core.ss.indexunary import register_new  # noqa: F401
 
+_delayed = {}
+
 del operator
