### 2023-06-21 2e53a9c
Support numpy 1.25 (`np.find_common_type` is deprecated)
