### 2023-06-21 ba8719d
Ignore `np.find_common_type` DeprecationWarning for now, b/c other dependencies use it
