### 2023-06-21 d17e376
Try to make valid numpy 1.25 environment
