### 2023-06-21 2b3b250
Don't make xfail strict while awkward is failing
