### 2023-06-21 581dd46
Don't install scipy.sparse 1.8 with numpy 1.25
