+++ .github/workflows/test_and_build.yml
@@ -216,6 +216,9 @@ jobs:
           fi
           if [[ ${npver} == "=1.25" ]] ; then
             numbaver=""
+            if [[ ${spver} == "=1.8" ]] ; then
+              spver=$(python -c 'import random ; print(random.choice(["=1.9", "=1.10", ""]))')
+            fi
           elif [[ ${npver} == "=1.24" || ${{ startsWith(steps.pyver.outputs.selected, '3.11') }} == true ]] ; then
             numbaver=$(python -c 'import random ; print(random.choice(["=0.57", ""]))')
           elif [[ ${npver} == "=1.21" ]] ; then
