### 2023-06-22 #81 @sneakers-the-rat
Here are my final review comments, I'll also add them into my review above. Thanks again to the authors for their work here, it was a pleasure reading the package <3

# Review

## General Comments

`python-graphblas` is an exemplary package among scientific python packages, with excellent docs, tests, code quality, examples, and contributor engagement. The package is an opinionated wrapper and interface to GraphBLAS, which is well justified and differentiated from prior wrappers. Throughout the review process, the authors have been receptive to feedback and I have all faith that they will address any continuing suggestions I have in future development work. I have no hesitation recommending this package to anyone who wants to use GraphBLAS from Python.

My outstanding recommendations in the remaining open issues are all future suggestions that the authors can take or leave, none are mission-critical. I want to compliment the authors on this excellent work, I'm glad to have had a reason to have read it. I would be happy to respond to any questions the authors have about this review and otherwise continue to engage on open issues.

## Code Quality

Library wrappers have their own sets of challenges and idioms, and `python-graphblas` handles them well with some room for future improvement. Wrapping code needs to make decisions about how to map the underlying API between languages and interact with the C library. The strategy used to abstract and expose the underlying library API can have large impacts on the maintainability and readability of the code, as can strategies for keeping the version of the wrapping package in sync with the changing API of the wrapped library.

`python-graphblas` further complicates these choices by admirably taking on the challenge of having multiple backend implementations (numpy and SuiteSparse:GraphBlas), and multiple export formats (NetworkX, Scipy.Sparse, PyData.Sparse, numpy, SuiteSparse, Matrix Market).

The `python-graphblas` team has chosen a bifurcated abstraction style which fits with the design of their wrapper. Rather than a transparent wrapper, the package introduces its own calling syntax to logically separate the i/o parts of a GraphBLAS call from the computation, which is well described and justified in the documentation.

At the time of reviewing,
- the basic container types are implmented as classes within the `core` module (eg [`core.vector.Vector`](https://github.com/python-graphblas/python-graphblas/blob/f0e03249fab46f36e12b66806a01f6e8e94cfde1/graphblas/core/vector.py#L135))
- operators implemented as separate modules within the main module, each with separate submodules for numpy and suitesparse implementations (eg. [`graphblas.unary`](https://github.com/python-graphblas/python-graphblas/tree/f0e03249fab46f36e12b66806a01f6e8e94cfde1/graphblas/unary))
	- ss implementations are actually implemented in the `core.operator` module, which then makes the symbols available in the top-level module namespace as a side effect of importing, eg. ([`graphblas.unary.ss`](https://github.com/python-graphblas/python-graphblas/blob/f0e03249fab46f36e12b66806a01f6e8e94cfde1/graphblas/unary/ss.py))
- the `core` module also implements the other syntax elements like infix operators that are used throughout the package.

This structure makes for a nice user-facing interface, being able to refer to classes and methods in a natural way (eg. `gb.Vector`, `gb.unary.abs`), obscuring the underlying code structure.

The tradeoff is the significant amount of indirection and implicitness in the code which presents a relatively high barrier to entry for new contributors. The mappings between the suitesparse GraphBLAS implementation and Python are programmed in several places, eg. within a [method](https://github.com/python-graphblas/python-graphblas/blob/0eb490fe6dcb96264fed6637517de01c26f036cb/graphblas/core/operator/unary.py#L342) and then again in the [tests](https://github.com/python-graphblas/python-graphblas/blob/0eb490fe6dcb96264fed6637517de01c26f036cb/graphblas/tests/test_operator_types.py#L44), which the authors describe as being useful as a [consistency check](https://github.com/python-graphblas/python-graphblas/issues/433#issuecomment-1513625131). Having some parts of the library dependent on import time side-effects is less than optimal and makes the code more difficult to reason about, but is certainly not fatal to the usability of the package. Again, the nature of a wrapping package requires making decisions about abstraction, so the only concern I have for the current code structure is the impact on maintainability. The current maintainers seem to have no trouble reasoning about the package, and I believe they are aware of these challenges and are actively working on them (they refactored a formerly massive `operator` file during the course of the review (https://github.com/python-graphblas/python-graphblas/pull/420)). The maintainers can do themselves and future contributors a favor by writing some additional developer docs that explain the structure of the library in greater detail, and I believe they will!

The other question I had was about the relationship between versioning in the API implementation in Python and in the underlying C library. GraphBLAS is a relatively mature API and seems to be only rarely changed, with care for backwards compatibility, so this is less of a concern than for a wrapper around a more actively evolving API. The authors have chosen to formally support a specific version of GraphBLAS ( https://github.com/python-graphblas/python-graphblas/pull/441 ) rather than build version-compatibility infrastructure within the package, which seems like an appropriate decision to me, given their other comments on how refactoring their backend system is on their development roadmap.

I want to also emphasize several of the "nice to have" features in the package, including a very impressive Recorder class that can keep a record of all the GraphBLAS calls made for debugging and reproducibility, automatic compilation of user-supplied functions including lambdas using numba, and the excellent i/o module. These indicate the authors are actively invested in user experience above and beyond the tidy API they expose.

Aside from the above comments, the fundamentals of the package design are strong: modern packaging with setuptools, automated code quality checks, CI, and coverage reports. The decisions made in code structure seem principled, responsive to the constraints of the problem, and result in a very usable user interface - my compliments to the Authors.

## Docs

The package is well documented from an introduction to the problem that GraphBLAS attempts to solve through package design decisions and practical examples. It is suitable for a general audience with some exposure to graph operations and programming in python, which is impressive given the highly specialized nature of the library.

Future suggestions for the authors include embedding their example notebooks in the documentation, and improving the API documentation. Currently, I assume due to some of the abstraction decisions made in the rest of the package, there is not comprehensive documentation of every operation available to the user that might otherwise be accomplished with calling `autodoc` over a class or a module, but some common operators are listed [here](https://python-graphblas.readthedocs.io/en/stable/user_guide/operators.html) and the list of available operators are present in the GraphBLAS documentation as well as through object inspection within an interactive Python session. Given the nature of wrapping code where the underlying operations are well documented elsewhere, this is less of a problem than it would be in other packages.

Altogether the docs are excellent with several clear points of improvement, but far above average in the landscape of scientific python packages.

## Tests

Who among us can claim 100% test coverage? https://coveralls.io/github/python-graphblas/python-graphblas

The tests are well organized and comprehensive, and I was able to find a corresponding test for every package feature I looked at easily. In more security-sensitive contexts one would want to do more adversarial input fuzzing, but I don't think that's all that relevant here since I've never seen graph data analysis libraries used as a malware vector. I have no notes on the tests, this is good work.
