### 2023-06-22 a762cae
Fix testing fix_test_suitesparse_graphblas (#464)
* Fix testing fix_test_suitesparse_graphblas

* Support numpy 1.25 (`np.find_common_type` is deprecated)

* Ignore `np.find_common_type` DeprecationWarning for now, b/c other dependencies use it

* Don't make xfail strict while awkward is failing

* Don't install scipy.sparse 1.8 with numpy 1.25

* Retry coveralls again, and don't fail if retry fails
