+++ graphblas/tests/test_io.py
@@ -369,7 +369,7 @@ def test_scipy_sparse():
 
 
 @pytest.mark.skipif("not ak")
-@pytest.mark.xfail(reason="Need to investigate test failure")
+# @pytest.mark.xfail(reason="Need to investigate test failure")
 def test_awkward_roundtrip():
     # Vector
     v = gb.Vector.from_coo([1, 3, 5], [20, 21, -5], size=22)
@@ -391,7 +391,7 @@ def test_awkward_roundtrip():
 
 
 @pytest.mark.skipif("not ak")
-@pytest.mark.xfail(reason="Need to investigate test failure")
+# @pytest.mark.xfail(reason="Need to investigate test failure")
 def test_awkward_iso_roundtrip():
     # Vector
     v = gb.Vector.from_coo([1, 3, 5], [20, 20, 20], size=22)
