+++ .github/workflows/test_and_build.yml
@@ -92,6 +92,7 @@ jobs:
       matrix:
         os: ["ubuntu-latest", "macos-latest", "windows-latest"]
         slowtask: ["pytest_normal", "pytest_bizarro", "notebooks"]
+        akver: ["2.2.2", "2.2.1", "2.2.0", "2.1.4", "2.1.3", "2.1.2", "2.1.1", "2.1.0", "2.0.10", "2.0.9", "2.0.8", "2.0.7", "2.0.6", "2.0.5", "1.10.3", "1.10.2", "1.10.1", "1.9.0"]
     env:
       # Wheels on OS X come with an OpenMP that conflicts with OpenMP from conda-forge.
       # Setting this is a workaround.
@@ -227,6 +228,10 @@ jobs:
             numbaver=$(python -c 'import random ; print(random.choice(["=0.56", "=0.57", ""]))')
           fi
           fmm=fast_matrix_market${fmmver}
+          akver=${{ matrix.akver }}
+          if [[ ${{ startsWith(steps.pyver.outputs.selected, '3.11') }} == true && ${awkward} == "awkward=1.9.0"]]; then
+            akver=""
+          fi
           awkward=awkward${akver}
           if [[ ${{ contains(steps.pyver.outputs.selected, 'pypy') ||
                 startsWith(steps.pyver.outputs.selected, '3.12') }} == true ||
