### 2023-06-23 d05644e
Clarify in docs that monoids are commutative and associative
