### 2023-06-25 python-graphblas/python-graphblas#471 @szhorvat
[pyos][docs] Versions and documentation
It would be nice to indicate in the documentation, especially the API documentation, what specific version the docs are for.

It would also be useful to be able to switch to the docs of older versions.

Part of: https://github.com/pyOpenSci/software-submission/issues/81
