### 2023-06-25 python-graphblas/python-graphblas#410 @szhorvat
Most of these can be looked up in the GraphBLAS spec/API, as Erik mentioned, but I'd also have liked to see them directly in the python-graphblas docs.

However, there are things which are specific to this package and are not in the docs, such as the `viz` namespace.

`io.draw` is documented, but it shows a warning saying that it's deprecated in favour of `viz.draw`. `viz.draw` does not appear in the API docs or guide.
