### 2023-06-25 python-graphblas/python-graphblas#473 @szhorvat
[pyos] viz.datashade() wrong positioning of matrix elements
I'm not sure if something's wrong with my setup, but with `viz.datashade()` I am not getting the output I expect. Here's an example:

```python
import graphblas as gb
M=gb.Matrix.from_edgelist([(0,1),(0,3),(2,3)])
gb.viz.datashade(M)
```

This is a 3-by-4 matrix, and this is reflected in the axes ranges ([0,3] vertical, [0,4] horizontal). However, the matrix elements only appear if I zoom out, and they're not positioned at the coordinates corresponding to their indices:
