### 2023-06-25 python-graphblas/python-graphblas#474 @szhorvat
[pyos][visualization] Reciprocal directed edges overlap
When plotting a graph/matrix, reciprocal directed edges overlap. The result is that only the weight of one of them is visible, which is confusing. I believe NetworkX does support curved edges, which could be a solution. However, generally, visualization does not seem to be a priority for NetworkX, so perhaps it's not the best idea to rely on it in the long term.

```python
M = gb.Matrix.from_edgelist([(0,1,10),(1,0,20)], nrows=2, ncols=2)
gb.viz.draw(M)
```

Both weights, 10 and 20, should be visible, and it should be clear which belongs to which edge.

I think this is important since `draw()` visualizes matrices specifically as directed graphs, with no undirected option being present.
