### 2023-06-25 python-graphblas/python-graphblas#475 @szhorvat
[pyos][io] Conversion to NetworkX drops isolated vertices
When converting to NetworkX, isolated vertices are dropped. Example:

```python
M=gb.Matrix.from_edgelist([(0,1),(3,0)],nrows=5,ncols=5)
G=gb.io.to_networkx(M)
```

`G.nodes` gives `NodeView((0, 1, 3))`. Nodes `2` and `4` are both missing.

Is this intentional or a bug? It is at least very surprising, and if intentional, it should be clearly documented.

Part of: https://github.com/pyOpenSci/software-submission/issues/81
