### 2023-06-26 #81 @szhorvat
tl;dr This is a very nice package, technically sound, with a well-thought out Pythonic interface. My perception was that in order to bring GraphBLAS to Python users (and thus fully realize its promise of implementing graph algorithm in terms of high-level building blocks), what we need the most is Python-centric documentation and training materials.

## Introduction

First of all, let me say that python-graphblas is a well thought out and technically solid package. It provides a Pythonic interface to the C-based GraphBLAS API, currently supporting SuiteSparse:GraphBLAS (which is the only usable GraphBLAS implementatin available at the moment).

The review should concern python-graphblas, but it's impossible to talk about it without also discussing GraphBLAS itself. And here I must make it clear that I am a newcomer to GraphBLAS. GraphBLAS is a sandardized C API aiming to provide high-level yet general building blocks for graph algorithms, similarly to how BLAS and LAPACK provide building blocks for linear algebra. It does this through some very elegant math, employing the language of linear algebra and abstract algebra, generalizing the usual `(+, *)` operations to a wider class of semirings. For example, with the usual `(+, *)` matrix product the `k`th power of a graph's adjacency matrix gives the number of walks between any two vertices. Replacing `+` by `min` and `*` by `+` will give us shortest path lengths up to `k` instead. What is not yet clear to me where the limits of this approach are: Are there common graph concepts that cannot be expressed in this framework? Are there some which can be expressed, but the most efficient _algorithms_ to compute them can't be expressed with GraphBLAS?

## Documentation and target audience

python-graphblas is realizing one of the major advantages of the GraphBLAS approach: it is now possible to implement graph algorithms in a convenient and high-level langauge like Python without the need for explicit loops, and thus without a significant loss of performance. Indeed, the python-graphblas authors have also created the graphblas-algorithm package, which achieves performance comparable to C/C++ libraries while being written in pure Python. (To be fair, it's not clear to me how much of the performance comes from parallelization, but the benchmarks are impressive.)

This brings me to a point about how python-graphblas is presented. This is the first paragraph in the docs:

> python-graphblas is a Pythonic interface to the highly performant SuiteSparse:GraphBLAS library for performing graph analytics in the language of linear algebra.

This makes the impression that python-graphblas is just an interface to a C library, perhaps even aimed at people who already understand that library. I think this is backwards: if one of the main GraphBLAS advantages is usability from high-level languages, then the target audience should be users of such languages. It seems to me that a project like this should spend at least as much effort on good documentation and training materials as on technical bindings. And this is precisely the weak point of python-graphblas. It is impossible to learn the system without referring to external material that describes the C API. In fact the competing package pygraphblas seems to be doing a little bit better on this front, though still not well enough for users to be able to avoid C-based documentation.

My main recommendation---and I realize that this is a long-term project---is making significant improvements to the documentation with the pure-Python user in mind.

In the intro:

 - Explain how this package is different from graph libraries like igraph or NetworkX? Make it clear that it is not at all an alternative to them.
 - Who is the target audience? Algorithm implementors, not people looking for existig algorithms.

More in depth:

 - Explain the basics of GraphBLAS. There's a very good starting point [here](https://python-graphblas.readthedocs.io/en/stable/getting_started/primer.html#primer).
 - Sufficient documentation for people to learn to work with GraphBLAS without ever leaving Python.
 - This includes better API docs as well, e.g. the discoverability of available operators and semirings.

## Interoperability

In addition to the technically sound and well thought out GraphBLAS interface, the package contains functionality for interoperability with graph theory / network analysis tools, as well as some visualization tools. The quality of these is in need of some improvement:

 - Interoperability is not trivial, and requires agreeing on a specific interpreatation of the data. Some discussion of this in the docs will be useful. See https://github.com/python-graphblas/python-graphblas/issues/398. It will be good to be explicit that GraphBLAS doesn't really operate with graphs, but with matrices, whose interpretation is up to the algorithm implementor.
 - Since a NetworkX converter is included, the interpretation issues should be documented. See also https://github.com/python-graphblas/python-graphblas/issues/475

I opened some visualization issues, https://github.com/python-graphblas/python-graphblas/issues/473 https://github.com/python-graphblas/python-graphblas/issues/474 https://github.com/python-graphblas/python-graphblas/issues/475

All this said, interoperability and visualization are not core functionality. As I see it, they are for convenience, and the package should not be judges based on these.
