### 2023-06-26 python-graphblas/python-graphblas#475 @SultanOrazbayev
Thank you for raising this!

By creating isolates in `networkx` for empty rows/cols,  we'd be making an assumption that such nodes exist. As such, this could be an additional convenience feature (as an additional `kwarg` such as `create_isolates=True`). Definitely something that can be documented.
