### 2023-06-26 8f03411
Allow isolate creation during export to networkx.
Fixes #475.
