### 2023-06-26 423f824
Add test for `create_isolates`.
