### 2023-06-26 python-graphblas/python-graphblas#475 @szhorvat
> for example, in networkx, creating a graph by adding edges [(0,1), (3,0)] would not create nodes 2 and 4, though I understand this is not exactly apples to apples comparision here

Indeed these are not comparable concepts. In your example, 0, 1 and 3 are not vertex indices. They are just "names". They might as well be `"foo"`, `(2,3)` and  `3.14`.

In comparison, in python-graphblas we have indices: 1st row (index 0), second row (index 1), etc.

The current behaviour is not consistent with the customary interpretation of adjacency matrices.
