### 2023-06-26 python-graphblas/python-graphblas#471 @eriknw
Agreed. @jim22k any reason why we don't do this?
