### 2023-06-26 python-graphblas/python-graphblas#474 @eriknw
Yeah, this isn't good. This version of `draw` was part of the initial PoC that @jim22k created years ago. We do plan to give vizualization a revamp and to show examples by including notebooks in the docs (#407), b/c viz is great for user applications and make great examples.

It might be nifty to be able to use `netgraph` or `graph-tool` (or ...) for viz.
- https://netgraph.readthedocs.io/en/latest/sphinx_gallery_output/index.html
- https://graph-tool.skewed.de/

I think @paul-tqh-nguyen may be interested in looking into visualization
