### 2023-06-26 python-graphblas/python-graphblas#474 @szhorvat
In case you're considering multiple viz backends, igraph does have good supports for both self-loops and multi-edges, and how has a matplotlib back-end.
