### 2023-06-27 python-graphblas/python-graphblas#410 @sneakers-the-rat
So I actually just tried to make this work for an hour or so using autodoc.

operator methods are all present in the `dir(operator)` call, but also each of the operator modules overrides `__dir__`, so that might be something.

I tried the following automodule commands:

```rst
.. _operator_methods:

Operator Methods
----------------

Binary
~~~~~~~

.. automodule:: graphblas.binary
    :members:
    :undoc-members:
    :imported-members:
    :private-members:

numpy
~~~~~

.. automodule:: graphblas.binary.numpy
    :members:
    :undoc-members:
    :imported-members:
    :private-members:

ss
~~

.. automodule:: graphblas.binary.ss
    :members:
    :undoc-members:
    :imported-members:
    :private-members:

op
~~

.. automodule:: graphblas.core.operator.binary
    :members:
    :undoc-members:
    :imported-members:
    :private-members:

```

And only the last one gave a sort of partial documentation here:
(zip of the generated html file, it's missing all the stylesheets and whatnot, but the content of the html is what's important)
[operator_methods.html.zip](https://github.com/python-graphblas/python-graphblas/files/11876482/operator_methods.html.zip)

So it seems to get some of them, but some of the methods are also in the `_delayed` module-level dict, and I ran out of time to chase down the other indirection methods.

This is I think related to both
https://github.com/python-graphblas/python-graphblas/issues/433
and
https://github.com/python-graphblas/python-graphblas/issues/429

The thing that is making this difficult is that **I'm still not sure if there is a single source of truth for all the available operator names.** I understand that they are generated at import time, but it should be possible for me to, say, call `binary.list_operators()` or something and get a definitive list of all possible operators, whether that be string representations, the actual objects, or both. As is, calling `dir()` or using tab completions is suboptimal since `dir()` returns all the globals as well, and tab completion requires you to hunt around in dir.

If there is such a thing, then documenting all available directives would just be a matter of making a new sphinx directive that calls `autoclass` for each of the operators. Or, if that's out of reach, simply pprinting the result of that function in the docs would do.

It is a small thing relative to the scale of the whole package, but it actually is a pretty important thing from a user discovery POV, and I feel like it shouldn't be too hard to implement by the devs who understand the abstraction of how the operators are registered better than I do
