### 2023-06-27 b4234e9
Handle scipy 1.11.0
