+++ scripts/check_versions.sh
@@ -5,12 +5,12 @@
 # Tip: add `--json` for more information.
 conda search 'numpy[channel=conda-forge]>=1.25.0'
 conda search 'pandas[channel=conda-forge]>=2.0.2'
-conda search 'scipy[channel=conda-forge]>=1.10.1'
+conda search 'scipy[channel=conda-forge]>=1.11.0'
 conda search 'networkx[channel=conda-forge]>=3.1'
-conda search 'awkward[channel=conda-forge]>=2.2.2'
+conda search 'awkward[channel=conda-forge]>=2.2.4'
 conda search 'sparse[channel=conda-forge]>=0.14.0'
-conda search 'fast_matrix_market[channel=conda-forge]>=1.6.0'
-conda search 'numba[channel=conda-forge]>=0.57.0'
+conda search 'fast_matrix_market[channel=conda-forge]>=1.7.2'
+conda search 'numba[channel=conda-forge]>=0.57.1'
 conda search 'pyyaml[channel=conda-forge]>=6.0'
 conda search 'flake8-bugbear[channel=conda-forge]>=23.6.5'
 conda search 'flake8-simplify[channel=conda-forge]>=0.20.0'
