### 2023-06-27 36be766
Add ways to use easily use burble in tests
