+++ graphblas/tests/conftest.py
@@ -1,4 +1,5 @@
 import atexit
+import contextlib
 import functools
 import itertools
 import platform
@@ -112,6 +113,24 @@ def ic():  # pragma: no cover (debug)
     icecream.install()
     # icecream.ic.disable()  # This disables icecream; do ic.enable() to re-enable
     return icecream.ic
+
+
+@contextlib.contextmanager
+def burble():
+    """Show the burble diagnostics within a context."""
+    # Don't keep track of previous state; always set to False when done
+    gb.ss.config["burble"] = True
+    try:
+        yield
+    finally:
+        gb.ss.config["burble"] = False
+
+
+@pytest.fixture(scope="session", autouse=True)
+def burble_all():  # pragma: no cover (debug)
+    """Show the burble diagnostics for the entire test."""
+    with burble():
+        yield burble
 
 
 def autocompute(func):
