+++ pyproject.toml
@@ -63,7 +63,7 @@ dependencies = [
     "pyyaml >=5.4",
     # These won't be installed by default after 2024.3.0
     # Use e.g. "python-graphblas[suitesparse]" or "python-graphblas[default]" instead
-    "suitesparse-graphblas >=8.0.1.0, <8.1",
+    "suitesparse-graphblas >8.0.0, <8.1",
     "numba >=0.55; python_version<'3.12'",  # make optional where numba is not supported
 ]
 
