### 2023-06-27 4632af5
Backport many of the SS:GB 8 changes to run on 7
