+++ graphblas/core/operator/binary.py
@@ -19,10 +19,10 @@ from ...dtypes import (
     UINT16,
     UINT32,
     UINT64,
-    _sample_values,
     _supports_complex,
     lookup_dtype,
 )
+from ...dtypes._core import _sample_values
 from ...exceptions import UdfParseError, check_status_carg
 from .. import _has_numba, _supports_udfs, ffi, lib
 from ..expr import InfixExprBase
@@ -506,7 +506,7 @@ class BinaryOp(OpBase):
                         type_.gb_obj,
                     ),
                     "BinaryOp",
-                    new_binary,
+                    new_binary[0],
                 )
                 op = TypedUserBinaryOp(new_type_obj, name, type_, ret_type, new_binary[0])
                 new_type_obj._add(op)
@@ -611,7 +611,7 @@ class BinaryOp(OpBase):
                 new_binary, binary_wrapper.cffi, ret_type._carg, dtype._carg, dtype2._carg
             ),
             "BinaryOp",
-            new_binary,
+            new_binary[0],
         )
         op = TypedUserBinaryOp(
             self,
