### 2023-06-27 52f0f36
CI: fail fast
