### 2023-06-27 #81 @tomalrussell
Thanks for the above, @sneakers-the-rat and @szhorvat !

@eriknw @jim22k @SultanOrazbayev - recognising that you've been engaged with the review already, and that some things may be pushed to a longer timeframe, are you happy to respond and make any priority changes by the end of the week (30th June), to try and keep approximately to our updated timeline?
- [editor-in-chief comments](https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1424901406)
- review 1 [checklist](https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1459181121) and [summary](https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1601951475)
- review 2 [checklist](https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1458245662) and [summary](https://github.com/pyOpenSci/software-submission/issues/81#issuecomment-1606360455)
