### 2023-06-27 7495d99
Make fast_matrix_market <1.7 and scipy >=1.11 play nicely together
