### 2023-06-27 1ef78c8
fix mmread to handle sparse and dense arrays
