### 2023-06-27 python-graphblas/python-graphblas#471 @jim22k
No specific reason other than it was easier to focus on writing current docs. No thought was given to how those versions of the docs are maintained over time.

I agree we should do this moving forward.
