+++ graphblas/monoid/ss.py
@@ -1,0 +1,5 @@
+from ..core import operator
+
+_delayed = {}
+
+del operator
