+++ .github/workflows/test_and_build.yml
@@ -374,6 +374,9 @@ jobs:
           # Tests lazy loading of lib, ffi, and NULL in gb.core
           echo "from graphblas.core import base" > script.py
           coverage run -a script.py
+          # Test another code pathway for loading lib
+          echo "from graphblas.core import lib" > script.py
+          coverage run -a script.py
           rm script.py
           # Tests whose coverage depend on order of tests :/
           # TODO: understand why these are order-dependent and try to fix
