### 2023-06-27 python-graphblas/python-graphblas#471 @eriknw
Almost done.

I added an automation rule on readthedocs to automatically activate non-prerelease versions when we do new releases in the future.

I also activated docs versions going back to `2022.11.1`, b/c that's the last version that built today without any issues.

The version and available versions now appear in the [flyout menu](https://docs.readthedocs.io/en/stable/flyout-menu.html) in the lower left in all pages except the initial landing page.

**The only remaining thing to do before closing this issue is to somehow show the version on the landing page (or decide not to).**
