### 2023-06-27 python-graphblas/python-graphblas#473 @eriknw
Oh no! That doesn't look right.

Testing of viz requires that the output is rendered and viewable, not just that the code runs, so we need to enhance our docs.

One option I've been considering is to have visualization in a different package similar to how we have algorithms in `graphblas-algorithms`. I like this for many reasons, including:
- dependencies handled by it
- it can have its own documentation
- this helps keep the scope of `python-graphblas` narrow (but I still like having most conversions in `python-graphblas`)
- other maintainers can "own" a significant package
- it can be more inviting and easier to contribute to for new contributors

I think the last point is really important. I think we can make the "barrier of entry" to contribute to `graphblas-algorithms` and e.g. `graphblaz-viz` _much_ lower compared to `python-graphblas`.
