### 2023-06-28 480e7a7
Refine language about commutative monoids.
