+++ docs/user_guide/operators.rst
@@ -90,8 +90,10 @@ Monoids
 -------
 
 Monoids extend the concept of a binary operator to require a single domain for all inputs and the output.
-Monoids are also associative and commutative, so the order of operations and inputs do not matter
-(for example, ``(a + b) + c == b + (c + a)``).
+Monoids are also associative so the order of operations does not matter
+(for example, ``(a + b) + c == a + (b + c)``).
+GraphBLAS primarily uses *commutative monoids* (for example, ``a + b == b + a``),
+and all standard monoids in python-graphblas commute.
 And finally, monoids have a default identity such that ``A op identity == A``.
 
 Monoids are commonly for reductions, collapsing all elements down to a single value.
