### 2023-06-28 cd45197
Backport many of the SS:GB 8 changes to run on 7 (#478)
* Backport many of the SS:GB 8 changes to run on 7

* Make fast_matrix_market <1.7 and scipy >=1.11 play nicely together

* fix mmread to handle sparse and dense arrays
