+++ README.md
@@ -213,7 +213,7 @@ convenient high-level interface for data pre-processing and transformation
 (e.g. `pandas`, `scipy.sparse`), visualization (e.g. `networkx`, `igraph`),
 interactive exploration and analysis (e.g. `networkx`, `igraph`) or for
 algoritms that are not (yet) implemented in `graphblas-algorithms` (e.g.
-`networkx`, `igraph`, `scipy.sparse`). To facilitate communication with
+`networkx`, `igraph`, `scipy.sparse.csgraph`). To facilitate communication with
 other libraries, `graphblas.io` contains multiple connectors, see the
 following section.
 
