+++ graphblas/binary/ss.py
@@ -3,6 +3,4 @@ from ..core.ss.binary import register_new  # noqa: F401
 
 _delayed = {}
 
-_delayed = {}
-
 del operator
