### 2023-06-28 05d3db8
Try to fix SS JIT config when testing on Linux
