+++ graphblas/tests/test_ssjit.py
@@ -56,7 +56,8 @@ def _setup_jit():
     elif sys.platform == "win32":
         pass
     else:
-        gb.ss.config["jit_c_compiler_name"] = f"{conda_prefix}/x86_64-apple-darwin13.4.0-clang"
+        # gb.ss.config["jit_c_compiler_name"] = f"{conda_prefix}/x86_64-apple-darwin13.4.0-clang"
+        gb.ss.config["jit_c_compiler_name"] = f"{conda_prefix}/bin/cc"
         gb.ss.config["jit_c_compiler_flags"] = (
             "-march=core2 -mtune=haswell -mssse3 -ftree-vectorize -fPIC -fPIE "
             f"-fstack-protector-strong -O2 -pipe -isystem {conda_prefix}/include -DGBNCPUFEAT "
