### 2023-06-29 0763253
skip ssjit tests on macos for now
