### 2023-06-29 b7b25b7
Move `dtypes._core` to `core.dtypes` (#479)
* Move `dtypes._core` to `core.dtypes`

I think this follows established patterns that we typically use.
