### 2023-06-29 d7a9460
fix `ss.dtypes`
