### 2023-06-29 python-graphblas/python-graphblas#473 @eriknw
We discussed viz in our weekly meeting this week.

Since viz is not a primary purpose of `python-graphblas`, for now we would like to keep our viz functions to a minimum that we need for documentation. We're open to creating a new graphblas-viz package that could host functions such as `datashade`. In general, conversions to other common formats (for example, `igraph`) that have their own visualization should be what we primarily support.

I'm sad to learn that `gb.viz.datashade` isn't working properly for you, so thanks for raising the issue. Since it's not a high priority for us atm, maybe we can add some notes to the docstring about it being experimental.
