### 2023-06-29 python-graphblas/python-graphblas#474 @eriknw
Thanks for letting us know about viz in `igraph`!

After a quick scan of examples, it's not clear to me how viz from `igraph` would be able to handle showing the edge attributes of reciprocal edges any better: https://python.igraph.org/en/latest/tutorials/index.html
