### 2023-06-29 python-graphblas/python-graphblas#473 @szhorvat
> Since viz is not a primary purpose of `python-graphblas`, for now we would like to keep our viz functions to a minimum that we need for documentation.

This is reasonable.
