### 2023-06-29 python-graphblas/python-graphblas#474 @szhorvat
I'm sorry about that, it seems that contrary to my memory, multi-edge handling is not yet implemented with the matplotlib backend. It does work with the Cairo backend.
