### 2023-06-30 python-graphblas/python-graphblas#474 @szhorvat
@eriknw I need to correct myself again. It seems that multigraph plotting in igraph is broken in the development version, but it's working fine in the released version. Python is not my main language, so I was a bit confused.

To be clear: Yes, multigraph plotting works fine in igraph with either the Cairo or matplotlib backends.

Example:

```
import igraph as ig
import matplotlib.pyplot as plt

g = ig.Graph([(0,1), (1,0), (1,2), (2,0)], directed=True)

import matplotlib.pyplot as plt
fig, ax= plt.subplots()
ig.plot(g, edge_label=['a','b','c','d'], target=ax)
```
