+++ graphblas/core/ss/unary.py
@@ -4,8 +4,17 @@ from ...exceptions import check_status_carg
 from .. import NULL, ffi, lib
 from ..operator.base import TypedOpBase
 from ..operator.unary import TypedUserUnaryOp, UnaryOp
+from . import _IS_SSGB7
 
 ffi_new = ffi.new
+if _IS_SSGB7:
+    # JIT was introduced in SuiteSparse:GraphBLAS 8.0
+    import suitesparse_graphblas as ssgb
+
+    raise ImportError(
+        "JIT was added to SuiteSparse:GraphBLAS in version 8; "
+        f"current version is {ssgb.__version__}"
+    )
 
 
 class TypedJitUnaryOp(TypedOpBase):
