+++ graphblas/unary/ss.py
@@ -1,5 +1,8 @@
 from ..core import operator
-from ..core.ss.unary import register_new  # noqa: F401
+from ..core.ss import _IS_SSGB7
+
+if not _IS_SSGB7:
+    from ..core.ss.unary import register_new  # noqa: F401
 
 _delayed = {}
 
