+++ graphblas/dtypes/ss.py
@@ -1,1 +1,4 @@
-from ..core.ss.dtypes import register_new  # noqa: F401
+from ..core.ss import _IS_SSGB7
+
+if not _IS_SSGB7:
+    from ..core.ss.dtypes import register_new  # noqa: F401
