+++ graphblas/core/ss/__init__.py
@@ -1,0 +1,3 @@
+import suitesparse_graphblas as _ssgb
+
+_IS_SSGB7 = _ssgb.__version__.split(".", 1)[0] == "7"
