### 2023-07-02 9a9d24b
Enable SS 7 in CI; and a bit more coverage
