+++ graphblas/core/ss/indexunary.py
@@ -7,14 +7,6 @@ from ..operator.indexunary import IndexUnaryOp, TypedUserIndexUnaryOp
 from . import _IS_SSGB7
 
 ffi_new = ffi.new
-if _IS_SSGB7:
-    # JIT was introduced in SuiteSparse:GraphBLAS 8.0
-    import suitesparse_graphblas as ssgb
-
-    raise ImportError(
-        "JIT was added to SuiteSparse:GraphBLAS in version 8; "
-        f"current version is {ssgb.__version__}"
-    )
 
 
 class TypedJitIndexUnaryOp(TypedOpBase):
@@ -33,15 +25,22 @@ class TypedJitIndexUnaryOp(TypedOpBase):
 
 
 def register_new(name, jit_c_definition, input_type, thunk_type, ret_type):
-    if backend != "suitesparse":
+    if backend != "suitesparse":  # pragma: no cover (safety)
         raise RuntimeError(
             "`gb.indexunary.ss.register_new` invalid when not using 'suitesparse' backend"
+        )
+    if _IS_SSGB7:
+        # JIT was introduced in SuiteSparse:GraphBLAS 8.0
+        import suitesparse_graphblas as ssgb
+
+        raise RuntimeError(
+            "JIT was added to SuiteSparse:GraphBLAS in version 8; "
+            f"current version is {ssgb.__version__}"
         )
     input_type = lookup_dtype(input_type)
     thunk_type = lookup_dtype(thunk_type)
     ret_type = lookup_dtype(ret_type)
-    if not name.startswith("ss."):
-        name = f"ss.{name}"
+    name = name if name.startswith("ss.") else f"ss.{name}"
     module, funcname = IndexUnaryOp._remove_nesting(name)
 
     rv = IndexUnaryOp(name)
