+++ graphblas/core/ss/select.py
@@ -6,14 +6,6 @@ from ..operator.select import SelectOp, TypedUserSelectOp
 from . import _IS_SSGB7
 
 ffi_new = ffi.new
-if _IS_SSGB7:
-    # JIT was introduced in SuiteSparse:GraphBLAS 8.0
-    import suitesparse_graphblas as ssgb
-
-    raise ImportError(
-        "JIT was added to SuiteSparse:GraphBLAS in version 8; "
-        f"current version is {ssgb.__version__}"
-    )
 
 
 class TypedJitSelectOp(TypedOpBase):
@@ -32,14 +24,21 @@ class TypedJitSelectOp(TypedOpBase):
 
 
 def register_new(name, jit_c_definition, input_type, thunk_type):
-    if backend != "suitesparse":
+    if backend != "suitesparse":  # pragma: no cover (safety)
         raise RuntimeError(
             "`gb.select.ss.register_new` invalid when not using 'suitesparse' backend"
         )
+    if _IS_SSGB7:
+        # JIT was introduced in SuiteSparse:GraphBLAS 8.0
+        import suitesparse_graphblas as ssgb
+
+        raise RuntimeError(
+            "JIT was added to SuiteSparse:GraphBLAS in version 8; "
+            f"current version is {ssgb.__version__}"
+        )
     input_type = lookup_dtype(input_type)
     thunk_type = lookup_dtype(thunk_type)
-    if not name.startswith("ss."):
-        name = f"ss.{name}"
+    name = name if name.startswith("ss.") else f"ss.{name}"
     # Register to both `gb.indexunary.ss` and `gb.select.ss.`
     indexunary.ss.register_new(name, jit_c_definition, input_type, thunk_type, BOOL)
     module, funcname = SelectOp._remove_nesting(name, strict=False)
