### 2023-07-02 c9b3beb
Don't use `-isysroot` for SS JIT on macos
