+++ README.md
@@ -215,7 +215,7 @@ might find that other libraries in the Python ecosystem provide a more
 convenient high-level interface for data pre-processing and transformation
 (e.g. `pandas`, `scipy.sparse`), visualization (e.g. `networkx`, `igraph`),
 interactive exploration and analysis (e.g. `networkx`, `igraph`) or for
-algoritms that are not (yet) implemented in `graphblas-algorithms` (e.g.
+algorithms that are not (yet) implemented in `graphblas-algorithms` (e.g.
 `networkx`, `igraph`, `scipy.sparse.csgraph`). To facilitate communication with
 other libraries, `graphblas.io` contains multiple connectors, see the
 following section.
