### 2023-07-05 e3118da
Update README.md (#442)
* Update README.md

Follows up on #438:

>Regarding "comparing to similar packages in the ecosystem", I think it would actually be informative to compare to e.g. scipy.sparse, networkx, and igraph. I think such a comparison could go in the README and/or documentation.

* Update README.md

* Update README.md

Co-authored-by: Erik Welch <erik.n.welch@gmail.com>
