### 2023-07-05 79cf3fa
doc: update io.rst for awkward array (#457)
* Update io.rst

As raised in #436

* Update documentation for `awkward-array`-related io functions.

* Update the doc string and change the order of the functions to reflect their intended use case.

remove trailing blanks

* Use double backticks
