### 2023-07-05 546814b
Clean up; change dependency metadata to only use SS:GB 7
