+++ pyproject.toml
@@ -63,7 +63,7 @@ dependencies = [
     "pyyaml >=5.4",
     # These won't be installed by default after 2024.3.0
     # Use e.g. "python-graphblas[suitesparse]" or "python-graphblas[default]" instead
-    "suitesparse-graphblas >=7.4.0.0, <8.1",
+    "suitesparse-graphblas >=7.4.0.0, <7.5",
     "numba >=0.55; python_version<'3.12'",  # make optional where numba is not supported
 ]
 
@@ -75,7 +75,7 @@ changelog = "https://github.com/python-graphblas/python-graphblas/releases"
 
 [project.optional-dependencies]
 suitesparse = [
-    "suitesparse-graphblas >=7.4.0.0, <8.1",
+    "suitesparse-graphblas >=7.4.0.0, <7.5",
 ]
 networkx = [
     "networkx >=2.8",
@@ -313,6 +313,7 @@ ignore = [
     "TRY004",  # Prefer `TypeError` exception for invalid type (Note: good advice, but not worth the nuisance)
     "TRY200",  # Use `raise from` to specify exception cause (Note: sometimes okay to raise original exception)
     "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar` (Note: no annotations yet)
+    "PERF401",  # Use a list comprehension to create a transformed list (Note: poorly implemented atm)
 
     # Intentionally ignored
     "COM812",  # Trailing comma missing
@@ -324,6 +325,8 @@ ignore = [
     "N806",  # Variable ... in function should be lowercase
     "N807",  # Function name should not start and end with `__`
     "N818",  # Exception name ... should be named with an Error suffix (Note: good advice)
+    "PERF203",  # `try`-`except` within a loop incurs performance overhead (Note: too strict)
+    "PLC0205",  # Class `__slots__` should be a non-string iterable (Note: string is fine)
     "PLR0124", # Name compared with itself, consider replacing `x == x` (Note: too strict)
     "PLR0911",  # Too many return statements
     "PLR0912",  # Too many branches
