### 2023-07-05 fdd4cad
Test python-suitespare-graphblas 8 wheels
