### 2023-07-05 python-graphblas/python-graphblas#433 @jim22k
Refactoring the operator classes is definitely high up on the list. It will be needed to support multiple backends, but will also give us a much cleaner separation between code we want to work with and code which is only there to parse the CFFI pieces.
