### 2023-07-05 6270abf
Update ss config doc for experimental GPU support
