+++ graphblas/tests/test_ssjit.py
@@ -1,0 +1,269 @@
+import os
+import sys
+
+import numpy as np
+import pytest
+from numpy.testing import assert_array_equal
+
+import graphblas as gb
+from graphblas import backend, binary, dtypes, indexunary, select, unary
+from graphblas.core import _supports_udfs as supports_udfs
+from graphblas.core.ss import _IS_SSGB7
+
+from .conftest import autocompute, burble
+
+from graphblas import Vector  # isort:skip (for dask-graphblas)
+
+try:
+    import numba
+except ImportError:
+    numba = None
+
+if backend != "suitesparse":
+    pytest.skip("not suitesparse backend", allow_module_level=True)
+
+
+@pytest.fixture(scope="module", autouse=True)
+def _setup_jit():
+    # Configuration values below were obtained from the output of the JIT config
+    # in CI, but with paths changed to use `{conda_prefix}` where appropriate.
+    if "CONDA_PREFIX" not in os.environ or _IS_SSGB7:
+        return
+    conda_prefix = os.environ["CONDA_PREFIX"]
+    gb.ss.config["jit_c_control"] = "on"
+    if sys.platform == "linux":
+        gb.ss.config["jit_c_compiler_name"] = f"{conda_prefix}/bin/x86_64-conda-linux-gnu-cc"
+        gb.ss.config["jit_c_compiler_flags"] = (
+            "-march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong "
+            f"-fno-plt -O2 -ffunction-sections -pipe -isystem {conda_prefix}/include -Wundef "
+            "-std=c11 -lm -Wno-pragmas -fexcess-precision=fast -fcx-limited-range "
+            "-fno-math-errno -fwrapv -O3 -DNDEBUG -fopenmp -fPIC"
+        )
+        gb.ss.config["jit_c_linker_flags"] = (
+            "-Wl,-O2 -Wl,--sort-common -Wl,--as-needed -Wl,-z,relro -Wl,-z,now "
+            "-Wl,--disable-new-dtags -Wl,--gc-sections -Wl,--allow-shlib-undefined "
+            f"-Wl,-rpath,{conda_prefix}/lib -Wl,-rpath-link,{conda_prefix}/lib "
+            f"-L{conda_prefix}/lib -shared"
+        )
+        gb.ss.config["jit_c_libraries"] = (
+            f"-lm -ldl {conda_prefix}/lib/libgomp.so "
+            f"{conda_prefix}/x86_64-conda-linux-gnu/sysroot/usr/lib/libpthread.so"
+        )
+        gb.ss.config["jit_c_cmake_libs"] = (
+            f"m;dl;{conda_prefix}/lib/libgomp.so;"
+            f"{conda_prefix}/x86_64-conda-linux-gnu/sysroot/usr/lib/libpthread.so"
+        )
+    elif sys.platform == "darwin":
+        gb.ss.config["jit_c_compiler_name"] = f"{conda_prefix}/bin/clang"
+        gb.ss.config["jit_c_compiler_flags"] = (
+            "-march=core2 -mtune=haswell -mssse3 -ftree-vectorize -fPIC -fPIE "
+            f"-fstack-protector-strong -O2 -pipe -isystem {conda_prefix}/include -DGBNCPUFEAT "
+            "-Wno-pointer-sign -O3 -DNDEBUG -fopenmp=libomp -fPIC -arch x86_64"
+        )
+        gb.ss.config["jit_c_linker_flags"] = (
+            "-Wl,-pie -Wl,-headerpad_max_install_names -Wl,-dead_strip_dylibs "
[…]
