+++ graphblas/indexunary/ss.py
@@ -1,4 +1,5 @@
 from ..core import operator
+from ..core.ss.indexunary import register_new  # noqa: F401
 
 _delayed = {}
 
