+++ graphblas/binary/ss.py
@@ -1,4 +1,5 @@
 from ..core import operator
+from ..core.ss.binary import register_new  # noqa: F401
 
 _delayed = {}
 
