+++ graphblas/select/ss.py
@@ -1,4 +1,5 @@
 from ..core import operator
+from ..core.ss.select import register_new  # noqa: F401
 
 _delayed = {}
 
