### 2023-07-11 python-graphblas/python-graphblas#474 @eriknw
Nice, thanks!
