### 2023-07-11 python-graphblas/python-graphblas#475 @eriknw
We discussed this and #476 in a recent meeting. I don't recall whether we got consensus, but there is an argument to be made that the nodes should always be added or at least be added by default. Thanks for raising this issue @szhorvat!

Converting GraphBLAS objects _to_ and _from_ graph objects from other libraries is something we'd like to tackle this year, so I think this issue is related to #398. Defining node attributes with a vector could also make sense.
