### 2023-07-11 python-graphblas/python-graphblas#428 @eriknw
Huh. Good question. I have no idea what documentation that is referring to. @jim22k can you take a look?

In general, we have `.ss` attributes on classes (`Matrix`, `Vector`), objects (`A.ss`, `v.ss`), and modules (`gb.binary.ss`).

We the addition of the SuiteSparse:GraphBLAS JIT, the docs website probably does need a SuiteSparse-specific section. The `ss` property on e.g. `Matrix` ought to be documented in the API Reference too--maybe it's supposed to be but something isn't working anymore.
