### 2023-07-11 python-graphblas/python-graphblas#410 @eriknw
Awesome, thanks for taking the time to explore this! This looks great (up until it stopped working) and is a good start.

I like the idea of having a definitive set of names of default operators. `__dir__` gets us close, and the following should work for all the operator types:
```python
{x for x in dir(gb.unary) if not x.startswith('_')} - {'register_new', 'register_anonymous', 'ss', 'from_string', 'numpy'}
```
I think there probably should be a global such as `DEFAULT_NAMES` or whatever. (It's late, and I'm bad at coming up with names when it's late)

The discussion from this issue probably has a few action items. We should probably enumerate them and/or make new issues as appropriate.
