### 2023-07-11 python-graphblas/python-graphblas#409 @eriknw
I agree with this. For example, the API docs for Matrix, Vector, and Scalar should be on different pages instead of on one long page, right?
