### 2023-07-11 #81 @eriknw
I applaud the reviewers 👏 . I never expected such thorough and honest reviews. Thank you for positive feedback and the criticisms. I agree with all of it--I think you nailed both the strengths and the weaknesses. The reviews are valuable and I suspect will help shape our vision and effort for the next couple of years.

In particular, there are two main areas we need to give more attention:
- documentation
- maintainability

Heh, this is probably true for many projects. My main focus in the near and medium term will be maintainability.

Now if I may ramble on for a bit...

It's interesting that reviews occur at a specific moment in time. I know the history of `python-graphblas`, and, trust me, it has evolved significantly in every 6 month period of its 3.5 year lifetime. If we had been writing pristine user-facing documentation from the beginning, it would have needed to be rewritten and revised endlessly. Functionality would probably be 1.5-2 years behind where we are today. If you're curious, go back and look at versions around `1.3.8` to `1.3.14`. It's recognizable, but so different and missing so, so much!

Anyway. I want to highlight this comment:

> if one of the main GraphBLAS advantages is usability from high-level languages, then the target audience should be users of such languages

Absolutely! I agree 100%. We aspire to this. It will take time.

From a product perspective, I/we wanted to get the syntax and functionality stable enough to begin writing `graphblas-algorithms` in earnest. We are targeting networkx users and are adding dispatching to networkx.

Oh, and if you're curious why our test coverage is so high, it's' for multiple reasons:
- I was considering Cythonizing everything and moving it to `cygraphblas`, and thorough coverage would give me more confidence
  - I eventually dropped this idea, b/c it would be too great a barrier for maintainability (also, coverage now supports Cython)
- Similarly, we still plan to further develop a parallel, Dask-ified version of this library that will use the same tests.
- For greater confidence when adding a new implementation, or to help test a new implementation of GraphBLAS
- It's simply a good idea, especially for a package as complicated as ours!

Wrapping up... I think we have replied to all open issues from the review. They will keep us busy, that's for sure.

@jim22k @SultanOrazbayev want to say anything else? I don't think it's necessary for you to comment here.

Thanks again all, hope to see you around ❤️ !
