### 2023-07-11 python-graphblas/python-graphblas#428 @jim22k
I think our docstrings used to mention Suitespare-specific settings, so those could be found throughout the API section.

Refactoring likely broke this and requires an update to the API docs.
