### 2023-07-13 #81 @szhorvat
Yes. The suggestions I made are mostly for the long term.
