### 2023-07-14 #81 @tomalrussell
Thanks all, time and effort very much appreciated. The review process is done!

All that's left is to wrap up, publish the version of record and acknowledge all your contributions.

---

🎉 `python-graphblas` has been approved by pyOpenSci! Thank you @eriknw for submitting `python-graphblas` and many thanks to @szhorvat and @sneakers-the-rat for reviewing this package! 😸

## Author and Reviewer Wrap-Up Tasks

There are just a few things left to do to wrap up this submission, @eriknw, @jim22k, @SultanOrazbayev:
- [ ] Add the badge for pyOpenSci peer-review to the README.md of `python-graphblas`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/81)`.
- [ ] Add `python-graphblas` to the pyOpenSci website. Please open a PR to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors.

Both reviewers and maintainers (@sneakers-the-rat, @szhorvat too):
- [ ] Reviewers and maintainers, if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a PR so we can list you on our website as contributors!

Please complete the final steps to wrap up this review. @tomalrussell, please do the following:

---
