### 2023-07-14 #81 @tomalrussell
@eriknw and team, I'd also like to invite you to write a blog post on your package for us to promote your work. If you are interested - here are a few examples of other blog posts:
- [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
- [movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)
- [`2019-11-18-pandera-dataframe-validation.md`](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

This can be a really high-level motivation for the package, for a slightly-scientific-Python-user-audience, or could draw on your introductory tutorial material to get straight to what the package does..

This is totally optional and not a requirement, but if you have time, we'd love to spread the word about `python-graphblas` to pyOpenSci blog readers :relaxed:
