### 2023-07-14 aba6999
Allow `__index__` only for integral dtypes on Scalars
