+++ graphblas/dtypes/__init__.py
@@ -41,3 +41,6 @@ def __getattr__(key):
         globals()["ss"] = ss
         return ss
     raise AttributeError(f"module {__name__!r} has no attribute {key!r}")
+
+
+_index_dtypes = {BOOL, INT8, UINT8, INT16, UINT16, INT32, UINT32, INT64, UINT64, _INDEX}
