### 2023-07-16 39a4451
Add `matrix.power` to compute e.g. `A @ A @ A`
