+++ graphblas/core/matrix.py
@@ -28,6 +28,7 @@ from .utils import (
     class_property,
     get_order,
     ints_to_numpy_buffer,
+    maybe_integral,
     normalize_values,
     output_type,
     values_to_numpy_buffer,
@@ -89,6 +90,49 @@ def _m_union_v(updater, left, right, left_default, right_default, op):
 
 def _reposition(updater, indices, chunk):
     updater[indices] = chunk
+
+
+def _power(updater, A, n, op):
+    opts = updater.opts
+    if n == 1:
+        updater << A
+        return
+    # Use repeated squaring; compute A^2, A^4, A^8, etc., and combine terms
+    result = square = square_expr = None
+    n, bit = divmod(n, 2)
+    while True:
+        if bit != 0:
+            if square_expr is not None:
+                if n == 0 and result is None:
+                    updater << square_expr
+                    return
+                if square is None:
+                    square = square_expr.new(name="Squares", **opts)
+                else:
+                    square(**opts) << square_expr
+                square_expr = None
+            if result is None:
+                if square is None:
+                    result = A
+                else:
+                    result = square.dup(name="Power", **opts)
+            elif n == 0:
+                updater << op(result @ square)
+                return
+            elif result is A:
+                result = op(result @ square).new(name="Power", **opts)
+            else:
+                result(**opts) << op(result @ square)
+        n, bit = divmod(n, 2)
+        if square_expr is not None:
+            if square is None:
+                square = square_expr.new(name="Squares", **opts)
+            else:
+                square << square_expr
+        if square is None:
+            square_expr = op(A @ A)
+        else:
+            square_expr = op(square @ square)
 
 
 class Matrix(BaseType):
@@ -155,8 +199,6 @@ class Matrix(BaseType):
         This is SuiteSparse-specific and may change in the future.
         This does not copy the matrix.
         """
-        from .vector import Vector
-
         if self._ncols != 1:
             raise ValueError(
                 f"Matrix must have a single column (not {self._ncols}) to be cast to a Vector"
@@ -2690,6 +2732,26 @@ class Matrix(BaseType):
             dtype=self.dtype,
         )
 
+    def power(self, n, op=semiring.plus_times):
+        method_name = "power"
+        if self._nrows != self._ncols:
+            raise DimensionMismatch(f"power only works for square Matrix; shape is {self.shape}")
+        if (N := maybe_integral(n)) is None:
+            raise TypeError(f"n must be a positive integer; got bad type: {type(n)}")
+        if N <= 0:
+            raise ValueError(f"n must be a positive integer; got: {N}")
+        op = get_typed_op(op, self.dtype, kind="semiring")
+        self._expect_op(op, "Semiring", within=method_name, argname="op")
+        return MatrixExpression(
+            "power",
+            None,
+            [self, _power, (self, N, op)],  # [*expr_args, func, args]
+            expr_repr=f"{{0.name}}.power({N}, op={op})",
+            nrows=self._nrows,
+            ncols=self._ncols,
+            dtype=self.dtype,
+        )
+
     ##################################
     # Extract and Assign index methods
     ##################################
@@ -3358,6 +3420,7 @@ class MatrixExpression(BaseExpression):
     mxv = wrapdoc(Matrix.mxv)(property(automethods.mxv))
     name = wrapdoc(Matrix.name)(property(automethods.name)).setter(automethods._set_name)
[…]
