### 2023-07-18 cd609cd
Add Matrix.power docstring
