+++ graphblas/io/__init__.py
@@ -4,4 +4,3 @@ from ._networkx import from_networkx, to_networkx
 from ._numpy import from_numpy, to_numpy  # deprecated
 from ._scipy import from_scipy_sparse, to_scipy_sparse
 from ._sparse import from_pydata_sparse, to_pydata_sparse
-from ._viz import draw  # deprecated
