### 2023-07-26 3476731
Allow `__index__` only for integral dtypes on Scalars (#481)
