### 2023-07-26 941d95f
Code comments for `power` implementation
