### 2023-07-26 8e42ded
Add `matrix.power` to compute e.g. `A @ A @ A @ ...` (#483)
