+++ graphblas/core/matrix.py
@@ -28,6 +28,7 @@ from .utils import (
     class_property,
     get_order,
     ints_to_numpy_buffer,
+    maybe_integral,
     normalize_values,
     output_type,
     values_to_numpy_buffer,
@@ -89,6 +90,68 @@ def _m_union_v(updater, left, right, left_default, right_default, op):
 
 def _reposition(updater, indices, chunk):
     updater[indices] = chunk
+
+
+def _power(updater, A, n, op):
+    opts = updater.opts
+    if n == 1:
+        updater << A
+        return
+    # Use repeated squaring: compute A^2, A^4, A^8, etc., and combine terms as needed.
+    # See `numpy.linalg.matrix_power` for a simpler implementation to understand how this works.
+    # We reuse `result` and `square` outputs, and use `square_expr` so masks can be applied.
+    result = square = square_expr = None
+    n, bit = divmod(n, 2)
+    while True:
+        if bit != 0:
+            # Need to multiply `square_expr` or `A` into the result
+            if square_expr is not None:
+                # Need to evaluate `square_expr`; either into final result, or into `square`
+                if n == 0 and result is None:
+                    # Handle `updater << A @ A` without an intermediate value
+                    updater << square_expr
+                    return
+                if square is None:
+                    # Create `square = A @ A`
+                    square = square_expr.new(name="Squares", **opts)
+                else:
+                    # Compute `square << square @ square`
+                    square(**opts) << square_expr
+                square_expr = None
+            if result is None:
+                # First time needing the intermediate result!
+                if square is None:
+                    # Use `A` if possible to avoid unnecessary copying
+                    # We will detect and handle `result is A` below
+                    result = A
+                else:
+                    # Copy square as intermediate result
+                    result = square.dup(name="Power", **opts)
+            elif n == 0:
+                # All done! No more terms to compute
+                updater << op(result @ square)
+                return
+            elif result is A:
+                # Now we need to create a new matrix for the intermediate result
+                result = op(result @ square).new(name="Power", **opts)
+            else:
+                # Main branch: multiply `square` into `result`
+                result(**opts) << op(result @ square)
+        n, bit = divmod(n, 2)
+        if square_expr is not None:
+            # We need to perform another squaring, so evaluate current `square_expr` first
+            if square is None:
+                # Create `square`
+                square = square_expr.new(name="Squares", **opts)
+            else:
+                # Compute `square`
+                square << square_expr
+        if square is None:
+            # First iteration! Create expression for first square
+            square_expr = op(A @ A)
+        else:
+            # Expression for repeated squaring
+            square_expr = op(square @ square)
 
 
 class Matrix(BaseType):
@@ -155,8 +218,6 @@ class Matrix(BaseType):
         This is SuiteSparse-specific and may change in the future.
         This does not copy the matrix.
         """
-        from .vector import Vector
-
         if self._ncols != 1:
             raise ValueError(
                 f"Matrix must have a single column (not {self._ncols}) to be cast to a Vector"
@@ -2690,6 +2751,60 @@ class Matrix(BaseType):
             dtype=self.dtype,
         )
 
+    def power(self, n, op=semiring.plus_times):
+        """Raise a square Matrix to the (positive integer) power ``n``.
+
[…]
