### 2023-07-26 8c49f3e
Remove deprecated `{scan,selectk,compactify}_{row,column}wise`
