+++ graphblas/core/ss/matrix.py
@@ -1,5 +1,4 @@
 import itertools
-import warnings
 
 import numpy as np
 from suitesparse_graphblas.utils import claim_buffer, claim_buffer_2d, unclaim_buffer
@@ -3715,51 +3714,6 @@ class ss:
             parent = parent.T
         return prefix_scan(parent, op, name=name, within="scan", **opts)
 
-    def scan_columnwise(self, op=monoid.plus, *, name=None, **opts):
-        """Perform a prefix scan across columns with the given monoid.
-
-        .. deprecated:: 2022.11.1
-            ``Matrix.ss.scan_columnwise`` will be removed in a future release.
-            Use ``Matrix.ss.scan(order="columnwise")`` instead.
-            Will be removed in version 2023.7.0 or later
-
-        For example, use ``monoid.plus`` (the default) to perform a cumulative sum,
-        and ``monoid.times`` for cumulative product.  Works with any monoid.
-
-        Returns
-        -------
-        Matrix
-        """
-        warnings.warn(
-            "`Matrix.ss.scan_columnwise` is deprecated; "
-            'please use `Matrix.ss.scan(order="columnwise")` instead.',
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return prefix_scan(self._parent.T, op, name=name, within="scan_columnwise", **opts)
-
-    def scan_rowwise(self, op=monoid.plus, *, name=None, **opts):
-        """Perform a prefix scan across rows with the given monoid.
-
-        .. deprecated:: 2022.11.1
-            ``Matrix.ss.scan_rowwise`` will be removed in a future release.
-            Use ``Matrix.ss.scan`` instead.
-            Will be removed in version 2023.7.0 or later
-
-        For example, use ``monoid.plus`` (the default) to perform a cumulative sum,
-        and ``monoid.times`` for cumulative product.  Works with any monoid.
-
-        Returns
-        -------
-        Matrix
-        """
-        warnings.warn(
-            "`Matrix.ss.scan_rowwise` is deprecated; please use `Matrix.ss.scan` instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return prefix_scan(self._parent, op, name=name, within="scan_rowwise", **opts)
-
     def flatten(self, order="rowwise", *, name=None, **opts):
         """Return a copy of the Matrix collapsed into a Vector.
 
@@ -3901,99 +3855,6 @@ class ss:
             k, fmt, indices, sort_axis, choose_func, is_random, do_sort, name
         )
 
-    def selectk_rowwise(self, how, k, *, name=None):  # pragma: no cover (deprecated)
-        """Select (up to) k elements from each row.
-
-        .. deprecated:: 2022.11.1
-            ``Matrix.ss.selectk_rowwise`` will be removed in a future release.
-            Use ``Matrix.ss.selectk`` instead.
-            Will be removed in version 2023.7.0 or later
-
-        Parameters
-        ----------
-        how : str
-            "random": choose k elements with equal probability
-            "first": choose the first k elements
-            "last": choose the last k elements
-        k : int
-            The number of elements to choose from each row
-
-        **THIS API IS EXPERIMENTAL AND MAY CHANGE**
-        """
-        warnings.warn(
-            "`Matrix.ss.selectk_rowwise` is deprecated; please use `Matrix.ss.selectk` instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        how = how.lower()
-        fmt = "hypercsr"
-        indices = "col_indices"
-        sort_axis = "sorted_cols"
-        if how == "random":
-            choose_func = choose_random
-            is_random = True
-            do_sort = False
-        elif how == "first":
-            choose_func = choose_first
-            is_random = False
-            do_sort = True
[…]
